### 2024-09-26 #7294 @kyleniemeyer
Hello @tomjholland, we will use this pre-review issue to find an editor and reviewers, and sort out any issues prior to the review starting.

While your submission looks in scope for JOSS and meets most of our requirements, I can't find instructions for running a test suite—at minimum, software needs to be tested (e.g., unit and functional) and have some instructions for manually running tests. Ideally, this would be automated (e.g., with pytest), and at best hooked into continuous integration (e.g., GitHub Actions).
