### 2024-09-26 #7294 @tomjholland
> Hello @tomjholland, we will use this pre-review issue to find an editor and reviewers, and sort out any issues prior to the review starting.
>
> While your submission looks in scope for JOSS and meets most of our requirements, I can't find instructions for running a test suite—at minimum, software needs to be tested (e.g., unit and functional) and have some instructions for manually running tests. Ideally, this would be automated (e.g., with pytest), and at best hooked into continuous integration (e.g., GitHub Actions).

Hello @kyleniemeyer, my project has a test suite that you should be able to find in `tests/`, in addition:
- the test suite is fully automated with pytest
- instructions to run pytest are included in step 5 of the contributing guidelines (if more detail is required do let me know)
- the tests are run with GitHub Actions on every pull request as specified in `.github/workflows/ci.yml`

I hope this helps, and let me know if there is anything else you need me to do.
