### 2024-09-27 f7a0502
Strip back readme processor, output now a single dict
