### 2024-09-27 b0bb38e
Strip back readme processor, output now a single dict
