+++ pyprobe/readme_processor.py
@@ -1,316 +1,37 @@
 """Module for processing PyPrBE README files."""
-import warnings
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Tuple
 
-import pybamm
 import yaml
-from pydantic import BaseModel, Field, model_validator
 
 
-class ReadmeModel(BaseModel):
-    """A Pydantic BaseModel class for processing the README.yaml file."""
+class ReadmeModel:
+    """A class for processing the README.yaml file."""
 
-    readme_dict: Dict[str, Any]
-    """A dictionary containing the contents of the README.yaml file."""
-    readme_type: List[str] = Field(default_factory=list)
-    """A list of strings indicating the README format for each experiment."""
-    number_of_experiments: int = Field(default_factory=int)
-    """The number of experiments in the README file."""
-    titles: List[str] = Field(default_factory=list)
-    """A list of strings containing the titles of the experiments."""
-    step_numbers: List[List[int]] = Field(default_factory=list)
-    """A list of lists containing the step numbers for each experiment."""
-    step_indices: List[List[int]] = Field(default_factory=list)
-    """A list of lists containing the step indices for each experiment."""
-    step_descriptions: List[List[str]] = Field(default_factory=list)
-    """A list of lists containing the step descriptions for each experiment."""
-    cycle_details: List[List[Tuple[int, int, int]]] = Field(default_factory=list)
-    """A list of lists containing the cycle details for each experiment."""
-    pybamm_experiment_descriptions: List[Tuple[str, ...]] = Field(default_factory=list)
-    """A list of tuples containing the PyBaMM experiment descriptions for each
-    experiment."""
-    pybamm_experiment_list: List[pybamm.Experiment] = Field(default_factory=list)
-    """A list of PyBaMM experiment objects for each experiment."""
-    pybamm_experiment: Optional[pybamm.Experiment] = Field(default=None)
-    """A PyBaMM experiment object for all experiments."""
+    def __init__(self, readme_dict: Dict[str, Any]) -> None:
+        """Initialize the ReadmeModel class."""
+        experiment_names = readme_dict.keys()
 
-    class Config:
-        """Pydantic configuration settings."""
-
-        arbitrary_types_allowed = True
-
-    @model_validator(mode="before")
-    @classmethod
-    def check_readme_dict(cls, data: Any) -> "ReadmeModel":
-        """Validate the structure of the README.yaml file.
-
-        Args:
-            data (Any): The data to be validated.
-
-        Returns:
-            ReadmeModel: The validated data.
-
-        Raises:
-            TypeError: If the 'Steps' field is not a dictionary or a list.
-            TypeError:
-                If the 'Steps' field is a dictionary and the keys are not integers
-                or the values are not strings.
-            TypeError: If the 'Steps' field is a list and the values are not strings.
-            TypeError: If the 'cycle' fields are not dictionaries.
-            TypeError:
-                If the 'cycle' fields are dictionaries and the keys are not strings
-                or the values are not integers.
-        """
-        readme_dict = data["readme_dict"]
-        data["readme_type"] = []
-        cls.number_of_experiments = len(readme_dict)
-        for experiment in readme_dict:
-            if "Steps" in readme_dict[experiment]:
-                steps = readme_dict[experiment]["Steps"]
-                if isinstance(steps, dict):
-                    if not all(
-                        isinstance(k, int) and isinstance(v, str)
-                        for k, v in steps.items()
-                    ):
-                        raise TypeError(
-                            "The 'Steps' field must be a dictionary with keys of type"
-                            " int and values of type str"
-                        )
-                    cycle_keys = [
-                        key
-                        for key in readme_dict[experiment].keys()
-                        if "cycle" in key.lower()
-                    ]
[…]
