### 2024-09-27 4ddabed
Persistent step_descriptions dataframe on RawData objects to create pybamm experiment
