### 2024-09-27 5e00b94
Move cycle column calculation into filtering step to allow for nested cycles
