### 2024-09-27 aa49a53
Move cycle column calculation into filtering step to allow for nested cycles
