### 2024-09-27 ebec642
Fix missing Cycle column in cycling analysis module
