### 2024-09-27 9072777
Remove Cycle from expected columns in imported dataframes
