+++ tests/cyclers/test_biologic.py
@@ -72,7 +72,6 @@ def test_read_and_process(benchmark, biologic_cycler, biologic_MB_cycler):
         "Date",
         "Time [s]",
         "Step",
-        "Cycle",
         "Event",
         "Current [A]",
         "Voltage [V]",
@@ -100,7 +99,6 @@ def test_read_and_process(benchmark, biologic_cycler, biologic_MB_cycler):
         "Date",
         "Time [s]",
         "Step",
-        "Cycle",
         "Event",
         "Current [A]",
         "Voltage [V]",
