+++ tests/cyclers/test_basecycler.py
@@ -41,7 +41,6 @@ def sample_pyprobe_dataframe():
                 "2022-02-02 02:02:02.000000",
             ],
             "Time [s]": [1.0, 2.0, 3.0],
-            "Cycle": [0, 0, 0],
             "Step": [1, 2, 3],
             "Event": [0, 1, 2],
             "Current [A]": [7.0e-3, 8.0e-3, 9.0e-3],
