### 2024-09-27 2d29de9
Replace checking for Cycle column with verifying cycle_info list
