### 2024-09-28 4e5b544
Make utils module, move flatten function into utils
