### 2024-09-28 cb77d0e
Make utils module, move flatten function into utils
