### 2024-09-28 b06ec94
Add back implicit readme support
