### 2024-09-28 e44b3ee
Add back implicit readme support
