### 2024-09-28 fb77274
Restore functionality to process readme containing only total steps
