### 2024-09-28 80ed719
Restore functionality to process readme containing only total steps
