+++ pyprobe/readme_processor.py
@@ -27,7 +27,7 @@ class ReadmeModel:
                 else:
                     raise ValueError("Invalid format for steps in README file")
             elif "Total Steps" in self.readme_dict[experiment_name].keys():
-                pass
+                self._process_total_steps_experiment(experiment_name)
             else:
                 raise ValueError(
                     "Each experiment must have a 'Steps' or 'Total Steps' key."
@@ -60,12 +60,7 @@ class ReadmeModel:
         Args:
             experiment_name (str): The name of the experiment.
         """
-        all_steps = [
-            experiment["Steps"]
-            for experiment in self.experiment_dict.values()
-            if "Steps" in experiment
-        ]
-        max_step = max(utils.flatten_list(all_steps)) if all_steps else 0
+        max_step = self._get_max_step()
         step_descriptions = self.readme_dict[experiment_name]["Steps"]
         step_numbers = list(range(max_step + 1, max_step + len(step_descriptions) + 1))
 
@@ -74,6 +69,34 @@ class ReadmeModel:
         )  # cast to satisfy mypy
         self.experiment_dict[experiment_name]["Step Descriptions"] = step_descriptions
         self.experiment_dict[experiment_name]["Cycles"] = []
+
+    def _process_total_steps_experiment(self, experiment_name: str) -> None:
+        """Process an experiment with total steps.
+
+        Args:
+            experiment_name (str): The name of the experiment.
+        """
+        total_steps = self.readme_dict[experiment_name]["Total Steps"]
+        max_step = self._get_max_step()
+        step_numbers = list(range(max_step + 1, max_step + total_steps + 1))
+        self.experiment_dict[experiment_name]["Steps"] = cast(
+            List[Union[str, int, Tuple[int, int, int]]], step_numbers
+        )  # cast to satisfy mypy
+        self.experiment_dict[experiment_name]["Step Descriptions"] = []
+        self.experiment_dict[experiment_name]["Cycles"] = []
+
+    def _get_max_step(self) -> int:
+        """Get the maximum step number from the experiment dictionary.
+
+        Returns:
+            int: The maximum step number from previously processed experiments.
+        """
+        all_steps = [
+            experiment["Steps"]
+            for experiment in self.experiment_dict.values()
+            if "Steps" in experiment
+        ]
+        return max(utils.flatten_list(all_steps)) if all_steps else 0
 
 
 def process_readme(
