### 2024-09-28 44b513f
Test that nans are in place where no step descriptions are given
