### 2024-09-29 2f1ac1e
Remove cycle column in cell module tests
