### 2024-09-29 4688149
In Neware cycler, time is read from Total Time column if Date not present
