### 2024-09-29 0453f92
Neware- read from Capacity column if Charge or Discharge capacity not present
