+++ pyprobe/cyclers/neware.py
@@ -4,6 +4,7 @@
 import polars as pl
 
 from pyprobe.cyclers.basecycler import BaseCycler
+from pyprobe.units import Units
 
 
 class Neware(BaseCycler):
@@ -19,6 +20,7 @@ class Neware(BaseCycler):
         "DChg. Cap.(*)": "Discharge Capacity [*]",
         "T1(*)": "Temperature [*]",
         "Total Time": "Time [*]",
+        "Capacity(*)": "Capacity [*]",
     }
 
     @staticmethod
@@ -32,7 +34,6 @@ class Neware(BaseCycler):
             pl.DataFrame | pl.LazyFrame: The DataFrame.
         """
         dataframe = BaseCycler.read_file(filepath)
-        print(dataframe)
         if "Time" in dataframe.collect_schema().names():
             dataframe = dataframe.with_columns(
                 pl.col("Time")
@@ -76,3 +77,57 @@ class Neware(BaseCycler):
         else:
             return pl.col("Time [s]")
             # return Units("Time", self._column_map["Time"]["Unit"]).to_default_unit()
+
+    @property
+    def charge_capacity(self) -> pl.Expr:
+        """Identify and format the charge capacity column.
+
+        For the Maccor cycler, this is the capacity column when the current is positive.
+
+        Returns:
+            pl.Expr: A polars expression for the charge capacity column.
+        """
+        if "Charge Capacity" in self._column_map.keys():
+            return super().charge_capacity
+        else:
+            current_direction = self.current.sign()
+            charge_capacity = (
+                Units(
+                    "Capacity", self._column_map["Capacity"]["Unit"]
+                ).to_default_unit()
+                * current_direction.replace(-1, 0).abs()
+            )
+            return charge_capacity.alias("Charge Capacity [Ah]")
+
+    @property
+    def discharge_capacity(self) -> pl.Expr:
+        """Identify and format the discharge capacity column.
+
+        For the Maccor cycler, this is the capacity column when the current is negative.
+
+        Returns:
+            pl.Expr: A polars expression for the discharge capacity column.
+        """
+        if "Discharge Capacity" in self._column_map.keys():
+            return super().discharge_capacity
+        else:
+            current_direction = self.current.sign()
+            discharge_capacity = (
+                Units(
+                    "Capacity", self._column_map["Capacity"]["Unit"]
+                ).to_default_unit()
+                * current_direction.replace(1, 0).abs()
+            )
+            return discharge_capacity.alias("Discharge Capacity [Ah]")
+
+    @property
+    def capacity(self) -> pl.Expr:
+        """Identify and format the capacity column.
+
+        For the Maccor cycler remove the option to calculate the capacity from a single
+        capacity column.
+
+        Returns:
+            pl.Expr: A polars expression for the capacity column.
+        """
+        return self.capacity_from_ch_dch
