### 2024-09-30 2ad14ce
Correct docstrings in Neware module
