### 2024-10-01 0bf6624
Add method to basecycler for tabulating the created column map
