### 2024-10-02 2c99bc8
Add check for missing columns to BaseCycler validator
