### 2024-10-02 8025eb7
Fix dashboard errors looking for procedure.titles and "Cycle" column heading
