### 2024-10-02 8f5c47c
Add cycle method back into Procedure class
