### 2024-10-02 5a35914
Add basic pybamm solution reading
