### 2024-10-02 a65b6e6
Add basic pybamm solution reading
