+++ pyprobe/cell.py
@@ -6,7 +6,7 @@ from typing import Callable, Dict, List, Optional
 
 import distinctipy
 import polars as pl
-import pybamm.solvers.solution as pybamm_solution
+import pybamm.solvers.solution
 from pydantic import BaseModel, Field, field_validator, validate_call
 
 from pyprobe.cyclers import basecycler, biologic, neware
@@ -285,7 +285,8 @@ class Cell(BaseModel):
     def import_pybamm_solution(
         self,
         procedure_name: str,
-        pybamm_solution: pybamm_solution,
+        experiment_names: List[str] | str,
+        pybamm_solutions: List[pybamm.solvers.solution] | pybamm.solvers.solution,
         output_data_path: Optional[str] = None,
         optional_variables: Optional[List[str]] = None,
     ) -> None:
@@ -294,6 +295,9 @@ class Cell(BaseModel):
         Filtering a PyBaMM solution object by cycle and step reflects the behaviour of
         the :code:`cycles` and :code:`steps` dictionaries of the PyBaMM solution object.
 
+        Multiple experiments can be imported into the same procedure. This is achieved
+        by providing multiple solution objects and experiment names.
+
         This method optionally writes the data to a parquet file, if a data path is
         provided.
 
@@ -301,27 +305,103 @@ class Cell(BaseModel):
             procedure_name (str):
                 A name to give the procedure. This will be used when calling
                 :code:`cell.procedure[procedure_name]`.
-            pybamm_solution (pybamm.solvers.solution.Solution):
-                The PyBaMM solution object to import.
+            pybamm_solutions (list or pybamm_solution):
+                A list of PyBaMM solution objects or a single PyBaMM solution object.
+            experiment_names (list or str):
+                A list of experiment names or a single experiment name to assign to the
+                PyBaMM solution object.
             output_data_path (str, optional):
                 The path to write the parquet file. Defaults to None.
             optional_variables (list, optional):
                 A list of variables to import from the PyBaMM solution object in
                 addition to the PyProBE required variables. Defaults to None.
         """
+        # the minimum required variables to import from the PyBaMM solution object
         required_variables = [
             "Time [s]",
             "Current [A]",
             "Terminal voltage [V]",
             "Discharge capacity [A.h]",
         ]
+
+        # get the list of variables to import from the PyBaMM solution object
         if optional_variables is not None:
-            imported_variables = required_variables + optional_variables
+            import_variables = required_variables + optional_variables
         else:
-            imported_variables = required_variables
-        pybamm_data = pybamm_solution.get_data_dict(imported_variables)
-        base_dataframe = pl.LazyFrame(pybamm_data)
-        base_dataframe = base_dataframe.select(
+            import_variables = required_variables
+
+        # check if the experiment names and PyBaMM solutions are lists
+        if isinstance(experiment_names, list) and isinstance(pybamm_solutions, list):
+            if len(experiment_names) != len(pybamm_solutions):
+                raise ValueError(
+                    "The number of experiment names and PyBaMM solutions must be equal."
+                )
+        elif isinstance(experiment_names, list) != isinstance(pybamm_solutions, list):
+            if isinstance(experiment_names, list):
+                raise ValueError(
+                    "A list of experiment names must be provided with a list of PyBaMM"
+                    " solutions."
+                )
+            else:
+                raise ValueError(
+                    "A single experiment name must be provided with a single PyBaMM"
+                    " solution."
+                )
+        else:
+            experiment_names = [str(experiment_names)]
+            pybamm_solutions = [pybamm_solutions]
+
+        all_solution_data = pl.LazyFrame({})
[…]
