### 2024-10-03 bbcb98e
Add docstring for _map_columns
