### 2024-10-03 29d35ae
Fix raising valueerror whenever capacity is missing
