### 2024-10-03 6f70104
move checking for optional columns out of properties and into required columns list
