### 2024-10-03 39cdc23
Fix error with empty initialised lazyframe
