### 2024-10-03 c440117
Fix error with empty initialised lazyframe
