### 2024-10-03 1987f96
Add validation for input data path
