### 2024-10-03 c722c17
Add validator for column dict
