### 2024-10-04 cca8d06
Add bounds to the minimize optimizer used for ocv fitting
