### 2024-10-04 ddb5da2
Add version number property and test against pyproject
