+++ tests/test_package.py
@@ -1,0 +1,10 @@
+"""Test package-level functionality."""
+
+import toml
+
+import pyprobe
+
+
+def test_version():
+    """Test version."""
+    assert pyprobe.__version__ == toml.load("pyproject.toml")["project"]["version"]
