+++ pyprobe/__init__.py
@@ -3,3 +3,5 @@ from .cell import Cell, make_cell_list  # noqa: F401
 from .dashboard import launch_dashboard  # noqa: F401
 from .plot import Plot  # noqa: F401
 from .result import Result  # noqa: F401
+
+__version__ = "1.0.2"
