### 2024-10-05 ae5aede
Store step descriptions in dict rather than lazyframe
