### 2024-10-05 980969e
Store step descriptions in dict rather than lazyframe
