### 2024-10-05 2d96816
Save and load a cell object to a folder
