+++ pyprobe/__init__.py
@@ -1,5 +1,5 @@
 """The PyProBE package."""
-from .cell import Cell, make_cell_list  # noqa: F401
+from .cell import Cell, load_archive, make_cell_list  # noqa: F401
 from .dashboard import launch_dashboard  # noqa: F401
 from .plot import Plot  # noqa: F401
 from .result import Result  # noqa: F401
