### 2024-10-05 4916bc0
Save and load a cell object to a folder
