### 2024-10-05 af62df8
Allow saved archive to be packaged into a zip file
