### 2024-10-05 07f3150
Allow saved archive to be packaged into a zip file
