### 2024-10-05 be0225b
Add feature to remove an experiment from a procedure
