### 2024-10-05 fbb7ebf
Add feature to remove an experiment from a procedure
