### 2024-10-05 df5add3
Fix version number check in loading archive, get __version__ directly from project file
