### 2024-10-05 343eaac
Fix version number check in loading archive, get __version__ directly from project file
