### 2024-10-08 8f90944
Correct pybamm current column to pyprobe convention of negative for discharge
