### 2024-10-08 8813b5f
Correct pybamm current column to pyprobe convention of negative for discharge
