### 2024-10-09 81dbc79
Add options for different compression levels and include in performance example
