### 2024-10-09 78ad199
Add options for different compression levels and include in performance example
