+++ pyprobe/cell.py
@@ -2,7 +2,7 @@
 import os
 import time
 import warnings
-from typing import Callable, Dict, List, Optional
+from typing import Callable, Dict, List, Literal, Optional
 
 import distinctipy
 import polars as pl
@@ -63,6 +63,9 @@ class Cell(BaseModel):
         input_filename: str | Callable[[str], str],
         output_filename: str | Callable[[str], str],
         filename_inputs: Optional[List[str]] = None,
+        compression_priority: Literal[
+            "performance", "file size", "uncompressed"
+        ] = "performance",
     ) -> None:
         """Convert cycler file into PyProBE format.
 
@@ -83,6 +86,12 @@ class Cell(BaseModel):
             filename_inputs (list):
                 The list of inputs to input_filename and output_filename, if they are
                 functions. These must be keys of the cell info.
+            compression_priority (str):
+                The priority of the compression algorithm to use on the resulting
+                parquet file. Available options are:
+                - 'performance': Use the 'lz4' compression algorithm (default).
+                - 'file size': Use the 'zstd' compression algorithm.
+                - 'uncompressed': Do not use compression.
         """
         input_data_path = self._get_data_paths(
             folder_path, input_filename, filename_inputs
@@ -101,8 +110,17 @@ class Cell(BaseModel):
         }
         t1 = time.time()
         importer = cycler_dict[cycler](input_data_path=input_data_path)
-        self._write_parquet(importer, output_data_path)
-        print(f"\tparquet written in {time.time()-t1:.2f} seconds.")
+        compression_dict = {
+            "uncompressed": "uncompressed",
+            "performance": "lz4",
+            "file size": "zstd",
+        }
+        self._write_parquet(
+            importer,
+            output_data_path,
+            compression=compression_dict[compression_priority],
+        )
+        print(f"\tparquet written in {time.time()-t1: .2f} seconds.")
 
     @validate_call
     def process_generic_file(
@@ -112,6 +130,9 @@ class Cell(BaseModel):
         output_filename: str | Callable[[str], str],
         column_dict: Dict[str, str],
         filename_inputs: Optional[List[str]] = None,
+        compression_priority: Literal[
+            "performance", "file size", "uncompressed"
+        ] = "performance",
     ) -> None:
         """Convert generic file into PyProBE format.
 
@@ -133,6 +154,12 @@ class Cell(BaseModel):
             filename_inputs (list):
                 The list of inputs to input_filename and output_filename.
                 These must be keys of the cell info.
+            compression_priority (str):
+                The priority of the compression algorithm to use on the resulting
+                parquet file. Available options are:
+                - 'performance': Use the 'lz4' compression algorithm (default).
+                - 'file size': Use the 'zstd' compression algorithm.
+                - 'uncompressed': Do not use compression.
         """
         input_data_path = self._get_data_paths(
             folder_path, input_filename, filename_inputs
@@ -149,8 +176,17 @@ class Cell(BaseModel):
             input_data_path=input_data_path,
             column_dict=column_dict,
         )
-        self._write_parquet(importer, output_data_path)
-        print(f"\tparquet written in {time.time()-t1:.2f} seconds.")
+        compression_dict = {
+            "uncompressed": "uncompressed",
+            "performance": "lz4",
+            "file size": "zstd",
+        }
+        self._write_parquet(
+            importer,
+            output_data_path,
+            compression=compression_dict[compression_priority],
+        )
+        print(f"\tparquet written in {time.time()-t1: .2f} seconds.")
 
     @validate_call
     def add_procedure(
@@ -220,17 +256,19 @@ class Cell(BaseModel):
         self,
         importer: basecycler.BaseCycler,
         output_data_path: str,
+        compression: str,
[…]
