### 2024-10-15 #7294 @kyleniemeyer
@tomjholland ah, that's great - I missed those at quick glance. It might help to add badges to the README that indicate the CI status.

@phibeck this looks good to go, thanks!
