### 2024-10-18 dc0cb2b
Run checks on all PR types
