+++ .github/workflows/ci.yml
@@ -1,6 +1,7 @@
 name: "Build and test"
 on:
   pull_request:
+    types: [opened, synchronize, reopened, ready_for_review]
     branches:
       - main
   workflow_dispatch:
