### 2024-10-18 433a1e1
Run checks on all PR types
