### 2024-10-18 2f0320d
Fix time column definition when date is not present
