### 2024-10-18 001eef7
Fix test failures caused by edge case passing incorrect type to analysis method
