+++ _sources/api/pyprobe.cyclers.maccor.rst.txt
@@ -1,0 +1,10 @@
+pyprobe.cyclers.maccor module
+=============================
+
+.. automodule:: pyprobe.cyclers.maccor
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+.. footbibliography::
