### 2024-10-21 7afe620
Move versioning into cell module and hardcode
