### 2024-10-23 26ef07f
Strip empty characters before or after datetime string when importing data
