### 2024-10-23 ad43149
Fix arbin import and add tests for collected dataframe
