### 2024-10-23 66351d0
Collect imported maccor dataframe in test, add benchmark
