+++ tests/cyclers/test_maccor.py
@@ -8,12 +8,16 @@ from polars.testing import assert_frame_equal
 from pyprobe.cyclers.maccor import Maccor
 
 
-def test_read_and_process_maccor():
+def test_read_and_process_maccor(benchmark):
     """Test reading and processing a sample Maccor file."""
     maccor_cycler = Maccor(
         input_data_path="tests/sample_data/maccor/sample_data_maccor.csv"
     )
-    pyprobe_dataframe = maccor_cycler.pyprobe_dataframe
+
+    def read_and_process_maccor():
+        return maccor_cycler.pyprobe_dataframe.collect()
+
+    pyprobe_dataframe = benchmark(read_and_process_maccor)
     expected_columns = [
         "Date",
         "Time [s]",
@@ -25,7 +29,7 @@ def test_read_and_process_maccor():
         "Temperature [C]",
     ]
     assert set(pyprobe_dataframe.columns) == set(expected_columns)
-    last_row = pl.LazyFrame(
+    last_row = pl.DataFrame(
         {
             "Date": datetime(2023, 11, 23, 15, 56, 24, 60000),
             "Time [s]": [13.06],
