### 2024-10-23 cb382e6
Collect basytec dataframe in test and add benchmark
