+++ tests/cyclers/test_basytec.py
@@ -18,12 +18,16 @@ def test_read_file_basytec():
     assert dataframe["Date"][2] == "2023-06-19 17:56:54.002823"
 
 
-def test_read_and_process_basytec():
+def test_read_and_process_basytec(benchmark):
     """Test the full process of reading and processing a file."""
     basytec_cycler = Basytec(
         input_data_path="tests/sample_data/basytec/sample_data_basytec.txt"
     )
-    pyprobe_dataframe = basytec_cycler.pyprobe_dataframe
+
+    def read_and_process_maccor():
+        return basytec_cycler.pyprobe_dataframe.collect()
+
+    pyprobe_dataframe = benchmark(read_and_process_maccor)
     expected_columns = [
         "Date",
         "Time [s]",
@@ -35,7 +39,7 @@ def test_read_and_process_basytec():
         "Temperature [C]",
     ]
     assert set(pyprobe_dataframe.columns) == set(expected_columns)
-    last_row = pl.LazyFrame(
+    last_row = pl.DataFrame(
         {
             "Date": datetime(2023, 6, 19, 17, 58, 3, 235803),
             "Time [s]": [70.235804],
