### 2024-10-23 798df23
Make common read_and_process function across all cycler tests
