+++ tests/cyclers/test_arbin.py
@@ -3,9 +3,10 @@
 from datetime import datetime
 
 import polars as pl
-from polars.testing import assert_frame_equal
 
 from pyprobe.cyclers.arbin import Arbin
+
+from .test_basecycler import helper_read_and_process
 
 
 def test_read_and_process_arbin(benchmark):
@@ -13,35 +14,22 @@ def test_read_and_process_arbin(benchmark):
     arbin_cycler = Arbin(
         input_data_path="tests/sample_data/arbin/sample_data_arbin.csv"
     )
-
-    def read_and_process_arbin():
-        return arbin_cycler.pyprobe_dataframe.collect()
-
-    pyprobe_dataframe = benchmark(read_and_process_arbin)
-    expected_columns = [
-        "Date",
-        "Time [s]",
-        "Step",
-        "Event",
-        "Current [A]",
-        "Voltage [V]",
-        "Capacity [Ah]",
-        "Temperature [C]",
-    ]
-    assert set(pyprobe_dataframe.columns) == set(expected_columns)
-    assert set(pyprobe_dataframe.select("Event").unique().to_series().to_list()) == set(
-        [0, 1, 2]
-    )
     expected_df = pl.DataFrame(
         {
-            "Date": [datetime(2024, 9, 20, 8, 32, 34, 558000)],
-            "Time [s]": [30.0005],
-            "Step": [1],
-            "Event": [0],
-            "Current [A]": [0.0],
-            "Voltage [V]": [3.534595],
-            "Capacity [Ah]": [0.000400839],
-            "Temperature [C]": [24.66422],
+            "Date": [datetime(2024, 9, 20, 8, 37, 5, 772000)],
+            "Time [s]": [301.214],
+            "Step": [3],
+            "Event": [2],
+            "Current [A]": [2.650138],
+            "Voltage [V]": [3.599601],
+            "Capacity [Ah]": [0.0007812400999999999],
+            "Temperature [C]": [24.68785],
         }
     )
-    assert_frame_equal(expected_df, pyprobe_dataframe.head(1))
+    expected_events = set([0, 1, 2])
+    helper_read_and_process(
+        benchmark,
+        arbin_cycler,
+        expected_final_row=expected_df,
+        expected_events=expected_events,
+    )
