### 2024-10-23 7b63206
Remove duplicate pydantic fields in daughter classes
