### 2024-10-23 73d3ec7
Allow dict values in the info dictionary
