+++ pyprobe/result.py
@@ -35,7 +35,7 @@ class Result(BaseModel):
 
     base_dataframe: Union[pl.LazyFrame, pl.DataFrame]
     """The data as a polars DataFrame or LazyFrame."""
-    info: Dict[str, Optional[str | int | float]]
+    info: Dict[str, Optional[str | int | float | Dict[Any, Any]]]
     """Dictionary containing information about the cell."""
     column_definitions: Dict[str, str] = Field(default_factory=dict)
     """A dictionary containing the definitions of the columns in the data."""
@@ -364,7 +364,7 @@ class Result(BaseModel):
                 | Dict[str, NDArray[np.float64] | List[float]]
             ]
         ],
-        info: Dict[str, Optional[str | int | float]],
+        info: Dict[str, Optional[str | int | float | Dict[Any, Any]]],
     ) -> "Result":
         """Build a Result object from a list of dataframes.
 
