### 2024-10-28 fad9398
Make dataframe concatenation diagonal when loading data to fill empty columns with nulls
