### 2024-10-28 92c1b39
Create RPT class for running analysis over multiple RPT procedures
