### 2024-10-29 1636607
Add rpt_summary property
