### 2024-11-01 c281ab1
Add generic interpolator classes to analysis utils
