+++ tests/analysis/test_utils.py
@@ -3,6 +3,7 @@
 import numpy as np
 import polars as pl
 import pytest
+from scipy import interpolate
 
 import pyprobe.analysis.utils as utils
 from pyprobe.filters import Experiment
@@ -62,3 +63,113 @@ def test_base_analysis(input_data_fixture):
         input_data=input_data_fixture, required_columns=["Units [mAh]"]
     )
     np.testing.assert_array_equal(analysis.variables, np.array([7, 8, 9]) * 1000)
+
+
+def test_base_interpolator_model_valid():
+    """Test _BaseInterpolatorModel with valid x and y."""
+    x = np.array([1, 2, 3])
+    y = np.array([4, 5, 6])
+    model = utils._BaseInterpolatorModel(x=x, y=y)
+    np.testing.assert_array_equal(model.x, x)
+    np.testing.assert_array_equal(model.y, y)
+
+
+def test_base_interpolator_model_x_not_strictly_increasing_or_decreasing():
+    """Test _BaseInterpolatorModel with x not strictly increasing or decreasing."""
+    x = np.array([1, 3, 2])
+    y = np.array([4, 6, 5])
+    with pytest.raises(ValueError, match="x must be strictly increasing or decreasing"):
+        utils._BaseInterpolatorModel(x=x, y=y)
+
+
+def test_base_interpolator_model_x_and_y_different_lengths():
+    """Test _BaseInterpolatorModel with x and y of different lengths."""
+    x = np.array([1, 2, 3])
+    y = np.array([4, 5])
+    with pytest.raises(ValueError, match="x and y must have the same length"):
+        utils._BaseInterpolatorModel(x=x, y=y)
+
+
+def test_base_interpolator_model_x_strictly_increasing():
+    """Test _BaseInterpolatorModel with x strictly increasing."""
+    x = np.array([1, 2, 3])
+    y = np.array([4, 5, 6])
+    model = utils._BaseInterpolatorModel(x=x, y=y)
+    np.testing.assert_array_equal(model.x, x)
+    np.testing.assert_array_equal(model.y, y)
+
+
+def test_base_interpolator_model_x_strictly_decreasing():
+    """Test _BaseInterpolatorModel with x strictly decreasing."""
+    x = np.array([3, 2, 1])
+    y = np.array([6, 5, 4])
+    model = utils._BaseInterpolatorModel(x=x, y=y)
+    np.testing.assert_array_equal(model.x, x)
+    np.testing.assert_array_equal(model.y, y)
+
+
+def test_linear_interpolator_init():
+    """Test _LinearInterpolator initialization with valid x and y."""
+    x = np.array([1, 2, 3])
+    y = np.array([4, 5, 6])
+    interpolator = utils._LinearInterpolator(x=x, y=y)
+    np.testing.assert_array_equal(interpolator.x, x)
+    np.testing.assert_array_equal(interpolator.y, y)
+
+
+def test_create_interpolator_linear():
+    """Test _create_interpolator with linear interpolation."""
+    x = np.array([1, 2, 3])
+    y = np.array([4, 5, 6])
+    LinearInterpolator = utils._create_interpolator(utils._LinearInterpolator)
+    interpolator = LinearInterpolator(x=x, y=y)
+    assert type(interpolator._interpolator) == utils._LinearInterpolator
+    x_new = np.array([1.5, 2.5])
+    y_new = interpolator(x_new)
+    assert y_new.shape == x_new.shape
+
+
+def test_create_interpolator_cubic():
+    """Test _create_interpolator with cubic interpolation."""
+    x = np.array([1, 2, 3])
+    y = np.array([4, 5, 6])
[…]
