### 2024-11-01 20c56df
Add method to provide OCP data to DMA class
