### 2024-11-01 c8fa9ba
Full cell OCV estimation w composite electrode
