### 2024-11-02 b4c8e44
Replace pydantic validation for interpolator with utility func, adapt linear interpolator as PPoly
