### 2024-11-02 419d9cd
Add interpolation-free methods for calculating OCV and its gradient
