+++ tests/analysis/test_dma.py
@@ -7,6 +7,7 @@ import pytest
 from pydantic import ValidationError
 
 import pyprobe.analysis.base.degradation_mode_analysis_functions as dma_functions
+from pyprobe.analysis import utils
 from pyprobe.analysis.degradation_mode_analysis import DMA
 from pyprobe.result import Result
 
@@ -29,18 +30,18 @@ def test_set_ocp_from_data_pe(stoichiometry_data, ocp_data):
     dma.set_ocp_from_data(
         stoichiometry_data, ocp_data, electrode="pe", interpolation_method="cubic"
     )
-    assert dma.PE_ocp[0] is not None
-    assert callable(dma.PE_ocp[0])
-    assert np.isclose(dma.PE_ocp[0](0.4), np.sin(0.4))
+    assert dma.ocp_pe[0] is not None
+    assert callable(dma.ocp_pe[0])
+    assert np.isclose(dma.ocp_pe[0](0.4), np.sin(0.4))
 
 
 def test_set_ocp_from_data_ne(stoichiometry_data, ocp_data):
     """Test the set_ocp_from_data method."""
     dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
     dma.set_ocp_from_data(stoichiometry_data, ocp_data, electrode="ne")
-    assert dma.NE_ocp[0] is not None
-    assert callable(dma.NE_ocp[0])
-    assert np.isclose(dma.NE_ocp[0](0.1), np.sin(0.1))
+    assert dma.ocp_ne[0] is not None
+    assert callable(dma.ocp_ne[0])
+    assert np.isclose(dma.ocp_ne[0](0.1), np.sin(0.1))
 
 
 def test_set_ocp_from_data_linear_interpolation(stoichiometry_data, ocp_data):
@@ -49,9 +50,9 @@ def test_set_ocp_from_data_linear_interpolation(stoichiometry_data, ocp_data):
     dma.set_ocp_from_data(
         stoichiometry_data, ocp_data, electrode="pe", interpolation_method="linear"
     )
-    assert dma.PE_ocp[0] is not None
-    assert callable(dma.PE_ocp[0])
-    assert np.isclose(dma.PE_ocp[0](0.4), np.sin(0.4))
+    assert dma.ocp_pe[0] is not None
+    assert callable(dma.ocp_pe[0])
+    assert np.isclose(dma.ocp_pe[0](0.4), np.sin(0.4))
 
 
 def test_set_ocp_from_data_cubic_interpolation(stoichiometry_data, ocp_data):
@@ -60,9 +61,9 @@ def test_set_ocp_from_data_cubic_interpolation(stoichiometry_data, ocp_data):
     dma.set_ocp_from_data(
         stoichiometry_data, ocp_data, electrode="pe", interpolation_method="cubic"
     )
-    assert dma.PE_ocp[0] is not None
-    assert callable(dma.PE_ocp[0])
-    assert np.isclose(dma.PE_ocp[0](0.4), np.sin(0.4))
+    assert dma.ocp_pe[0] is not None
+    assert callable(dma.ocp_pe[0])
+    assert np.isclose(dma.ocp_pe[0](0.4), np.sin(0.4))
 
 
 def test_set_ocp_from_data_multiple_components(stoichiometry_data, ocp_data):
[…]
