+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -1,11 +1,13 @@
 """Module for degradation mode analysis methods."""
 
-from typing import Callable, List, Literal, Optional, Tuple
+from typing import Any, Callable, Dict, List, Literal, Optional, Tuple
 
 import numpy as np
 import polars as pl
 from numpy.typing import NDArray
 from pydantic import BaseModel
+from scipy import optimize
+from scipy.interpolate import PPoly
 
 import pyprobe.analysis.base.degradation_mode_analysis_functions as dma_functions
 from pyprobe.analysis import utils
@@ -30,10 +32,10 @@ class DMA(BaseModel):
     dma_result: Optional[Result] = None
     """The degradation mode analysis results."""
 
-    NE_ocp: List[None | Callable[[NDArray], NDArray]] = []
+    ocp_ne: List[None | Callable[[NDArray], NDArray]] = []
     """A list of functions for the negative electrode OCP."""
 
-    PE_ocp: List[None | Callable[[NDArray], NDArray]] = []
+    ocp_pe: List[None | Callable[[NDArray], NDArray]] = []
     """A list of functions for the positive electrode OCP."""
 
     def set_ocp_from_data(
@@ -45,16 +47,169 @@ class DMA(BaseModel):
         component_index: int = 0,
         total_electrode_components: int = 1,
     ) -> None:
-        """Provide the OCP data for a given electrode."""
+        """Provide the OCP data for a given electrode.
+
+        Args:
+            stoichiometry (NDArray[np.float64]): The stoichiometry data.
+            ocp (NDArray[np.float64]): The OCP data.
+            electrode (Literal["pe", "ne"]): The electrode to set the OCP data for.
+            interpolation_method
+                (Literal["linear", "cubic", "Pchip", "Akima"], optional):
+                The interpolation method to use. Defaults to "linear".
+            component_index (int, optional):
+                The index of the electrode component to set the OCP data for.
+                Defaults to 0.
+            total_electrode_components (int, optional):
+                The total number of electrode components. Defaults to 1.
+        """
         interpolator = utils.interpolators[interpolation_method](x=stoichiometry, y=ocp)
         if electrode == "pe":
-            if len(self.PE_ocp) == 0:
-                self.PE_ocp = [None] * total_electrode_components
-            self.PE_ocp[component_index] = interpolator
+            if len(self.ocp_pe) == 0:
+                self.ocp_pe = [None] * total_electrode_components
+            self.ocp_pe[component_index] = interpolator
         elif electrode == "ne":
-            if len(self.NE_ocp) == 0:
-                self.NE_ocp = [None] * total_electrode_components
-            self.NE_ocp[component_index] = interpolator
+            if len(self.ocp_ne) == 0:
+                self.ocp_ne = [None] * total_electrode_components
+            self.ocp_ne[component_index] = interpolator
+
+    def _f_OCV(
+        self,
+        SOC: NDArray[np.float64],
+        x_pe_lo: NDArray[np.float64],
+        x_pe_hi: NDArray[np.float64],
+        x_ne_lo: NDArray[np.float64],
+        x_ne_hi: NDArray[np.float64],
+    ) -> NDArray[np.float64]:
+        """Calculate the full cell OCV as a function of SOC.
+
+        Args:
+            SOC (NDArray[np.float64]): The full cell SOC.
+            x_pe_lo (float): The cathode stoichiometry at lowest cell SOC.
+            x_pe_hi (float): The cathode stoichiometry at highest cell SOC.
[…]
