### 2024-11-04 3b33624
Add support for sympy OCP expressions and add methods for differentiation
