+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -4,6 +4,7 @@ from typing import Any, Callable, Dict, List, Literal, Optional, Tuple
 
 import numpy as np
 import polars as pl
+import sympy as sp
 from numpy.typing import NDArray
 from pydantic import BaseModel
 from scipy import optimize
@@ -32,11 +33,124 @@ class DMA(BaseModel):
     dma_result: Optional[Result] = None
     """The degradation mode analysis results."""
 
-    ocp_ne: List[None | Callable[[NDArray], NDArray]] = []
-    """A list of functions for the negative electrode OCP."""
+    def _ocp_derivative(
+        self, ocp_list: List[None | Callable[[NDArray], NDArray]]
+    ) -> List[Callable[[NDArray], NDArray]]:
+        """Calculate the derivative of each OCP.
 
-    ocp_pe: List[None | Callable[[NDArray], NDArray]] = []
-    """A list of functions for the positive electrode OCP."""
+        Args:
+            ocp (Callable[[NDArray], NDArray]):
+                The OCP function. Must be a differentiable function. Currently supported
+                formats are scipy.interpolate.PPoly objects (from utils.interpolators)
+                or sympy expressions.
+
+        Returns:
+            Callable[[NDArray], NDArray]: The derivative of the OCP.
+        """
+
+        def _check_free_symbols(free_symbols: set[sp.Symbol]) -> sp.Symbol:
+            if len(free_symbols) == 1:
+                return free_symbols.pop()
+            else:
+                raise ValueError(
+                    "OCP must be a function of a single variable, " "the stoichiometry."
+                )
+
+        derivatives = []
+        for ocp in ocp_list:
+            if isinstance(ocp, PPoly):
+                derivatives.append(ocp.derivative())
+            elif isinstance(ocp, sp.Expr):
+                free_symbols = ocp.free_symbols
+                sto = _check_free_symbols(free_symbols)
+                gradient = sp.diff(ocp, sto)
+                derivatives.append(sp.lambdify(sto, gradient, "numpy"))
+            else:
+                raise ValueError(
+                    "OCP is not in a differentiable format. OCP must be a"
+                    " PPoly object or a sympy expression."
+                )
+        return derivatives
+
+    @property
+    def d_ocp_pe(self) -> List[Callable[[NDArray], NDArray]]:
+        """Return the derivative of the positive electrode OCP.
+
+        Returns:
+            List[Callable[[NDArray], NDArray]]: The derivative of the positive
+            electrode OCP.
+        """
+        return self._ocp_derivative(self._ocp_pe)
+
+    @property
+    def d_ocp_ne(self) -> List[Callable[[NDArray], NDArray]]:
+        """Return the derivative of the negative electrode OCP.
+
+        Returns:
+            List[Callable[[NDArray], NDArray]]: The derivative of the negative
+            electrode OCP.
+        """
+        return self._ocp_derivative(self._ocp_ne)
+
+    @property
+    def ocp_pe(
+        self,
+    ) -> List[Optional[Callable[[NDArray[np.float64]], NDArray[np.float64]]]]:
+        """Return a list of positive electrode OCPs as a function of stoichiometry.
+
+        Returns:
+            List[Optional[Callable[[NDArray[np.float64]], NDArray[np.float64]]]]:
+                A list of positive electrode OCPs as a function of stoichiometry.
+        """
+        if not hasattr(self, "_f_ocp_pe"):
+            self._f_ocp_pe = []
+            for ocp in self._ocp_pe:
+                if isinstance(ocp, sp.Expr):
[…]
