### 2024-11-04 0dae32d
Add method for running ocv fitting with new refactored methods
