+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -347,6 +347,137 @@ class DMA(BaseModel):
 
         return selected_optimizer(_ocv_curve_fit_objective, **optimizer_options).x
 
+    def run_ocv_curve_fit(
+        self,
+        fitting_target: Literal["OCV", "dQdV", "dVdQ"] = "OCV",
+        optimizer: Literal["minimize", "differential_evolution"] = "minimize",
+        optimizer_options: Dict[str, Any] = {
+            "x0": np.array([0.9, 0.1, 0.1, 0.9]),
+            "bounds": [(0, 1), (0, 1), (0, 1), (0, 1)],
+        },
+    ) -> Tuple[Result, Result]:
+        """Fit half cell open circuit potential curves to full cell OCV data.
+
+        Args:
+            fitting_target (Literal["OCV", "dQdV", "dVdQ"], optional):
+                The target for the curve fitting. Defaults to "OCV".
+            optimizer (Literal["minimize", "differential_evolution"], optional):
+                The optimization algorithm to use. Defaults to "minimize".
+            optimizer_options (Dict[str, Any], optional):
+                The options for the optimization algorithm. Defaults to
+                {"x0": np.array([0.9, 0.1, 0.1, 0.9]),
+                 "bounds": [(0, 1), (0, 1), (0, 1), (0, 1)]}.
+                Where x0 is the initial guess for the fit and bounds are the
+                limits for the fit. The fitting parameters are ordered [x_pe_lo,
+                x_pe_hi, x_ne_lo, x_ne_hi], where lo and hi indicate the stoichiometry
+                limits at low and high full-cell SOC respectively.
+
+        Returns:
+            Tuple[Result, Result]:
+                - Result: The stoichiometry limits and electrode capacities.
+                - Result: The fitted OCV data.
+        """
+        if "SOC" in self.input_data.column_list:
+            required_columns = ["Voltage [V]", "Capacity [Ah]", "SOC"]
+            validator = AnalysisValidator(
+                input_data=self.input_data, required_columns=required_columns
+            )
+            voltage, capacity, SOC = validator.variables
+            cell_capacity = np.abs(np.ptp(capacity)) / np.abs(np.ptp(SOC))
+        else:
+            required_columns = ["Voltage [V]", "Capacity [Ah]"]
+            validator = AnalysisValidator(
+                input_data=self.input_data, required_columns=required_columns
+            )
+            voltage, capacity = validator.variables
+            cell_capacity = np.abs(np.ptp(capacity))
+            SOC = (capacity - capacity.min()) / cell_capacity
+
+        dVdSOC = np.gradient(voltage, SOC)
+        dSOCdV = np.gradient(SOC, voltage)
+
+        fitting_target_data = {
+            "OCV": voltage,
+            "dQdV": dSOCdV,
+            "dVdQ": dVdSOC,
+        }[fitting_target]
+
+        x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi = self._curve_fit_ocv(
+            SOC,
+            fitting_target_data,
+            fitting_target,
+            optimizer,
+            optimizer_options,
+        )
+
+        (
+            pe_capacity,
+            ne_capacity,
+            li_inventory,
+        ) = dma_functions.calc_electrode_capacities(
+            x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi, cell_capacity
+        )
+
+        self.stoichiometry_limits = self.input_data.clean_copy(
+            pl.DataFrame(
+                {
+                    "x_pe low SOC": np.array([x_pe_lo]),
+                    "x_pe high SOC": np.array([x_pe_hi]),
[…]
