### 2024-11-04 2d9a020
Replace interpolators dict with separately defined functions
