### 2024-11-04 0d37c10
Move interpolation functions into smoothing module
