+++ tests/analysis/test_utils.py
@@ -3,7 +3,6 @@
 import numpy as np
 import polars as pl
 import pytest
-from scipy import interpolate
 
 import pyprobe.analysis.utils as utils
 from pyprobe.filters import Experiment
@@ -63,113 +62,3 @@ def test_base_analysis(input_data_fixture):
         input_data=input_data_fixture, required_columns=["Units [mAh]"]
     )
     np.testing.assert_array_equal(analysis.variables, np.array([7, 8, 9]) * 1000)
-
-
-def test_linear_interpolator():
-    """Test _LinearInterpolator initialization with valid x and y."""
-    x = np.array([1, 2, 3])
-    y = np.array([4, 5, 6])
-    interpolator = utils._LinearInterpolator(x, y)
-    x_new = np.array([1.5, 2.5])
-    y_new = interpolator(x_new)
-    np.testing.assert_array_equal(y_new, np.array([4.5, 5.5]))
-
-
-def test_create_interpolator_linear():
-    """Test _create_interpolator with linear interpolation."""
-    x = np.array([1, 2, 3])
-    y = np.array([4, 5, 6])
-    interpolator = utils._create_interpolator(utils._LinearInterpolator, x, y)
-    assert type(interpolator) == utils._LinearInterpolator
-    assert isinstance(interpolator, interpolate.PPoly)
-    x_new = np.array([1.5, 2.5])
-    y_new = interpolator(x_new)
-    assert y_new.shape == x_new.shape
-
-
-def test_create_interpolator_cubic():
-    """Test _create_interpolator with cubic interpolation."""
-    x = np.array([1, 2, 3])
-    y = np.array([4, 5, 6])
-    interpolator = utils._create_interpolator(interpolate.CubicSpline, x, y)
-    assert type(interpolator) == interpolate.CubicSpline
-    x_new = np.array([1.5, 2.5])
-    y_new = interpolator(x_new)
-    assert y_new.shape == x_new.shape
-
-
-def test_create_interpolator_pchip():
-    """Test _create_interpolator with Pchip interpolation."""
-    x = np.array([1, 2, 3])
-    y = np.array([4, 5, 6])
-    interpolator = utils._create_interpolator(interpolate.PchipInterpolator, x, y)
-    assert type(interpolator) == interpolate.PchipInterpolator
-    x_new = np.array([1.5, 2.5])
-    y_new = interpolator(x_new)
-    assert y_new.shape == x_new.shape
-
-
-def test_create_interpolator_akima():
-    """Test _create_interpolator with Akima interpolation."""
-    x = np.array([1, 2, 3])
-    y = np.array([4, 5, 6])
-    interpolator = utils._create_interpolator(interpolate.Akima1DInterpolator, x, y)
-    assert type(interpolator) == interpolate.Akima1DInterpolator
-    x_new = np.array([1.5, 2.5])
-    y_new = interpolator(x_new)
-    assert y_new.shape == x_new.shape
-
-
-def test_interpolators_dict():
-    """Test the interpolators dictionary."""
-    x = np.array([1, 2, 3])
-    y = np.array([4, 5, 6])
-    assert (
-        type(utils.interpolators["Akima"](x=x, y=y)) == interpolate.Akima1DInterpolator
-    )
-
-
-def test_validate_interp_input_vectors_valid():
-    """Test _validate_interp_input_vectors with valid input vectors."""
-    x = np.array([1, 2, 3])
-    y = np.array([4, 5, 6])
-    try:
-        x, y = utils._validate_interp_input_vectors(x, y)
-    except ValueError:
-        pytest.fail("Unexpected ValueError raised")
-
-
-def test_validate_interp_input_vectors_invalid_length():
[…]
