### 2024-11-05 52e2bad
Add downsampling tests for non-monotonic data
