### 2024-11-05 cab0222
Add downsampling utility function to DMA class
