### 2024-11-05 4155b8d
Add benchmarks to downsampling test
