### 2024-11-06 5baf077
Add methods for extending the data stored in result objects and joining new columns
