### 2024-11-07 326169e
Add class and methods for batch dma fitting
