+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -1,5 +1,7 @@
 """Module for degradation mode analysis methods."""
 
+import concurrent.futures
+import copy
 from typing import Any, Callable, Dict, List, Literal, Optional, Tuple
 
 import numpy as np
@@ -147,9 +149,13 @@ class DMA(BaseModel):
 
         Args:
             ocp (sp.Expr): _description_
-            electrode (Literal[&quot;pe&quot;, &quot;ne&quot;]): _description_
-            component_index (int, optional): _description_. Defaults to 0.
-            total_electrode_components (int, optional): _description_. Defaults to 1.
+            electrode (Literal["pe", "ne"]): Electrode to set the OCP data for.
+            component_index (int, optional):
+                Electrode component index in a composite electrode to set the OCP data
+                for. Defaults to 0.
+            total_electrode_components (int, optional):
+                Total number of electrode components in a composite electrode.
+                Defaults to 1.
         """
         if electrode == "pe":
             if not hasattr(self, "_ocp_pe"):
@@ -781,3 +787,276 @@ class DMA(BaseModel):
             )
         )
         return DMA(input_data=average_result)
+
+
+def _run_ocv_curve_fit_with_index(
+    index: int,
+    dma_object: DMA,
+    fitting_target: Literal["OCV", "dQdV", "dVdQ"],
+    optimizer: Literal["minimize", "differential_evolution"],
+    optimizer_options: Dict[str, Any],
+) -> Tuple[int, Tuple[Result, Result]]:
+    """Wrapper function for running the OCV curve fitting with an index."""
+    result = dma_object.run_ocv_curve_fit(
+        fitting_target=fitting_target,
+        optimizer=optimizer,
+        optimizer_options=optimizer_options,
+    )
+    return index, result
+
+
+def _quantify_degradation_modes_with_index(
+    index: int, dma_object: DMA, reference_stoichiometry_limits: Result
+) -> Tuple[int, Result]:
+    result = dma_object.quantify_degradation_modes(
+        reference_stoichiometry_limits=reference_stoichiometry_limits,
+    )
+    return index, result
+
+
+class BatchDMA(BaseModel):
+    """A class for running DMA analysis across multiple input data sets."""
+
+    input_data: List[PyProBEDataType]
+    """The a list of input data for the degradation mode analysis."""
+
+    fitted_OCV: Optional[List[Result]] = None
+    """A list of the fitted OCV data."""
+
+    batch_DMA_results: Optional[Result] = None
+    """The result of the DMA analysis across all provided input_data.
+
+    Contains the stoichiometry limits, SOH, LAM_pe, LAM_ne, and LLI for each of the
+    provided OCV fits.
+    """
+
+    @property
+    def dma_objects(self) -> List[DMA]:
+        """Return the DMA objects."""
+        if not hasattr(self, "_dma_objects"):
+            self._dma_objects = [DMA(input_data=data) for data in self.input_data]
+        return self._dma_objects
+
+    def downsample_ocv(
+        self,
+        sampling_interval: float,
+        occurrence: Literal["first", "last", "middle"] = "first",
+        time_column: str = "Time [s]",
+    ) -> List[Result]:
+        """Downsample the OCV data to a specified sampling interval.
+
+        This method updates the input_data attribute with the downsampled data, and
+        returns a Result object. The logic is based on
+        :func:`pyprobe.analysis.smoothing.downsample_data`, which also provides the
+        logic for the :func:`pyprobe.analysis.smoothing.Smoothing.downsample` method.
+
+        Args:
+            sampling_interval (float): The sampling interval for the downsampled data.
+            occurrence (Literal["first", "last", "middle"], optional):
+                The occurrence to use when downsampling. Defaults to "first".
[…]
