### 2024-11-08 #7294 @ritesh001
Hi, I would be willing to review this submission.

From: Sophie Beck ***@***.***>
Date: Friday, November 8, 2024 at 12:20 PM
Cc: Ritesh Kumar ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: PyProBE: Python Processing for Battery Experiments (Issue #7294)

👋 @DEARLIBS<https://github.com/DEARLIBS>, @mefuller<https://github.com/mefuller> & @ritesh001<https://github.com/ritesh001>, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
