### 2024-11-08 d6eb7a8
provide ocp functions to method for calculating full cell OCV
