### 2024-11-09 2cba16f
Switch setting ocp data into adding ocp data to list
