+++ tests/analysis/test_dma.py
@@ -25,10 +25,10 @@ def ocp_data():
     return np.sin(np.linspace(0, np.pi, 1000))
 
 
-def test_set_ocp_from_data_pe(stoichiometry_data, ocp_data):
-    """Test the set_ocp_from_data method."""
+def test_add_ocp_from_data_pe(stoichiometry_data, ocp_data):
+    """Test the add_ocp_from_data method."""
     dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
-    dma.set_ocp_from_data(
+    dma.add_ocp_from_data(
         stoichiometry_data, ocp_data, electrode="pe", interpolation_method="cubic"
     )
     assert dma.ocp_pe[0] is not None
@@ -36,19 +36,19 @@ def test_set_ocp_from_data_pe(stoichiometry_data, ocp_data):
     assert np.isclose(dma.ocp_pe[0](0.4), np.sin(0.4))
 
 
-def test_set_ocp_from_data_ne(stoichiometry_data, ocp_data):
-    """Test the set_ocp_from_data method."""
+def test_add_ocp_from_data_ne(stoichiometry_data, ocp_data):
+    """Test the add_ocp_from_data method."""
     dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
-    dma.set_ocp_from_data(stoichiometry_data, ocp_data, electrode="ne")
+    dma.add_ocp_from_data(stoichiometry_data, ocp_data, electrode="ne")
     assert dma.ocp_ne[0] is not None
     assert callable(dma.ocp_ne[0])
     assert np.isclose(dma.ocp_ne[0](0.1), np.sin(0.1))
 
 
-def test_set_ocp_from_data_linear_interpolation(stoichiometry_data, ocp_data):
-    """Test the set_ocp_from_data method with linear interpolation."""
+def test_add_ocp_from_data_linear_interpolation(stoichiometry_data, ocp_data):
+    """Test the add_ocp_from_data method with linear interpolation."""
     dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
-    dma.set_ocp_from_data(
+    dma.add_ocp_from_data(
         stoichiometry_data, ocp_data, electrode="pe", interpolation_method="linear"
     )
     assert dma.ocp_pe[0] is not None
@@ -56,10 +56,10 @@ def test_set_ocp_from_data_linear_interpolation(stoichiometry_data, ocp_data):
     assert np.isclose(dma.ocp_pe[0](0.4), np.sin(0.4))
 
 
-def test_set_ocp_from_data_cubic_interpolation(stoichiometry_data, ocp_data):
-    """Test the set_ocp_from_data method with cubic interpolation."""
+def test_add_ocp_from_data_cubic_interpolation(stoichiometry_data, ocp_data):
+    """Test the add_ocp_from_data method with cubic interpolation."""
     dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
-    dma.set_ocp_from_data(
+    dma.add_ocp_from_data(
         stoichiometry_data, ocp_data, electrode="pe", interpolation_method="cubic"
     )
     assert dma.ocp_pe[0] is not None
@@ -67,40 +67,31 @@ def test_set_ocp_from_data_cubic_interpolation(stoichiometry_data, ocp_data):
     assert np.isclose(dma.ocp_pe[0](0.4), np.sin(0.4))
 
 
[…]
