### 2024-11-09 89b4dcf
Add numerical calculation of derivative on pure python functions
