### 2024-11-09 0cec7c1
Create factory method for objective function to allow for flexibility to composite electrode
