+++ tests/analysis/test_dma.py
@@ -5,7 +5,6 @@ import numpy as np
 import polars as pl
 import pytest
 import sympy as sp
-from numpy.typing import NDArray
 from pydantic import ValidationError
 
 import pyprobe.analysis.base.degradation_mode_analysis_functions as dma_functions
@@ -102,7 +101,7 @@ def test_ocp_derivative_ppoly():
     y = np.array([0, 1, 0, -1])
     ppoly_ocp = smoothing.linear_interpolator(x, y)
     dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
-    derivative = dma._ocp_derivative([ppoly_ocp])[0]
+    derivative = dma._ocp_derivative(ppoly_ocp)
     assert callable(derivative)
     x = np.array([0, 1, 2, 3])
     np.testing.assert_allclose(derivative(x), np.array([1, -1, -1, -1]))
@@ -113,7 +112,7 @@ def test_ocp_derivative_sympy():
     dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
     x = sp.symbols("x")
     sympy_ocp = 2 * x**2 + 3 * x + 1
-    derivative = dma._ocp_derivative([sympy_ocp])[0]
+    derivative = dma._ocp_derivative(sympy_ocp)
     assert callable(derivative)
     x = np.array([0, 1, 2, 3])
     np.testing.assert_allclose(derivative(x), np.array([3, 7, 11, 15]))
@@ -123,11 +122,11 @@ def test_ocp_derivative_function():
     """Test _ocp_derivative with a python function."""
     dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
 
-    def ocp(x: NDArray[np.float64]) -> NDArray[np.float64]:
+    def ocp(x):
         """Sample OCP function."""
         return 2 * x**2 + 3 * x + 1
 
-    derivative = dma._ocp_derivative([ocp])[0]
+    derivative = dma._ocp_derivative(ocp)
     assert callable(derivative)
     x = np.linspace(0, 100, 100)
     np.testing.assert_allclose(derivative(x)[1:-1], (4 * x + 3)[1:-1])
@@ -200,13 +199,11 @@ def test_f_grad_OCV():
     ocp_pe = smoothing.cubic_interpolator(x=x_pts, y=ocp_pe_pts)
     ocp_ne_pts = 3 * x_pts**3
     ocp_ne = smoothing.cubic_interpolator(x=x_pts, y=ocp_ne_pts)
-    dma.ocp_ne = [ocp_ne]
-    dma.ocp_pe = [ocp_pe]
     x_pe_lo = 0
     x_pe_hi = 1
     x_ne_lo = 0
     x_ne_hi = 1
-    d_ocv = dma._f_grad_OCV(x_pts, x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi)
+    d_ocv = dma._f_grad_OCV(x_pts, ocp_pe, ocp_ne, x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi)
     x_pe = np.linspace(x_pe_lo, x_pe_hi, 100)
     x_ne = np.linspace(x_ne_lo, x_ne_hi, 100)
     d_ocv_expected = 4 * x_pe - 9 * x_ne**2
@@ -216,119 +213,10 @@ def test_f_grad_OCV():
     x_pe_hi = 0.1
     x_ne_lo = 0.1
     x_ne_hi = 0.7
-    d_ocv = dma._f_grad_OCV(x_pts, x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi)
[…]
