+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -2,7 +2,7 @@
 
 import concurrent.futures
 import copy
-from typing import Any, Callable, Dict, List, Literal, Optional, Tuple
+from typing import Any, Callable, Dict, List, Literal, Optional, Tuple, Union
 
 import numpy as np
 import polars as pl
@@ -35,334 +35,21 @@ class DMA(BaseModel):
     dma_result: Optional[Result] = None
     """The degradation mode analysis results."""
 
-    @property
-    def ocp_pe(
-        self,
-    ) -> List[Optional[Callable[[NDArray[np.float64]], NDArray[np.float64]]]]:
-        """Return a list of positive electrode OCPs as a function of stoichiometry.
+    _ocp_pe: List[
+        Union[
+            Callable[[NDArray[np.float64]], NDArray[np.float64]],
+            PPoly,
+            sp.Expr,
+        ]
+    ] = []
 
-        Returns:
-            List[Optional[Callable[[NDArray[np.float64]], NDArray[np.float64]]]]:
-                A list of positive electrode OCPs as a function of stoichiometry.
-        """
-        if not hasattr(self, "_f_ocp_pe"):
-            self._f_ocp_pe = []
-            for ocp in self._ocp_pe:
-                if isinstance(ocp, sp.Expr):
-                    self._f_ocp_pe.append(
-                        sp.lambdify(ocp.free_symbols.pop(), ocp, "numpy")
-                    )
-                else:
-                    self._f_ocp_pe.append(ocp)
-        return self._f_ocp_pe
-
-    @ocp_pe.setter
-    def ocp_pe(
-        self,
-        value: List[Optional[Callable[[NDArray[np.float64]], NDArray[np.float64]]]],
-    ) -> None:
-        """Set the OCP data for the positive electrode."""
-        self._f_ocp_pe = value
-        self._ocp_pe = value
-
-    @property
-    def ocp_ne(
-        self,
-    ) -> List[Optional[Callable[[NDArray[np.float64]], NDArray[np.float64]]]]:
-        """Return a list of negative electrode OCPs as a function of stoichiometry.
-
-        Returns:
-            List[Optional[Callable[[NDArray[np.float64]], NDArray[np.float64]]]]:
-                A list of negative electrode OCPs as a function of stoichiometry.
-        """
-        if not hasattr(self, "_f_ocp_ne"):
-            self._f_ocp_ne = []
-            for ocp in self._ocp_ne:
-                if isinstance(ocp, sp.Expr):
-                    self._f_ocp_ne.append(
-                        sp.lambdify(ocp.free_symbols.pop(), ocp, "numpy")
-                    )
-                else:
-                    self._f_ocp_ne.append(ocp)
-        return self._f_ocp_ne
-
-    @ocp_ne.setter
-    def ocp_ne(
-        self,
-        value: List[Optional[Callable[[NDArray[np.float64]], NDArray[np.float64]]]],
-    ) -> None:
-        """Set the OCP data for the negative electrode."""
-        self._f_ocp_ne = value
-        self._ocp_ne = value
-
-    def add_ocp_from_data(
-        self,
-        stoichiometry: NDArray[np.float64],
-        ocp: NDArray[np.float64],
-        electrode: Literal["pe", "ne"],
-        interpolation_method: Literal["linear", "cubic", "Pchip", "Akima"] = "linear",
-    ) -> None:
-        """Provide a OCP data for a given electrode.
-
-        Appends to the ocp list for the given electrode. Composite electrodes require
-        multiple calls to this method to provide the OCP data for each component.
-
-        Args:
-            stoichiometry (NDArray[np.float64]): The stoichiometry data.
-            ocp (NDArray[np.float64]): The OCP data.
-            electrode (Literal["pe", "ne"]): The electrode to set the OCP data for.
[…]
