### 2024-11-11 a33a3ba
Add an OCP class for dealing with half cell data
