### 2024-11-11 23a9942
Integrate new OCP classes into DMA methods
