+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -141,6 +141,12 @@ class DMA(BaseModel):
     input_data: PyProBEDataType
     """The input data for the degradation mode analysis."""
 
+    ocp_pe: OCP
+    """An object containing the positive electrode OCP functions."""
+
+    ocp_ne: OCP
+    """An object containing the negative electrode OCP functions."""
+
     # 3. Define the attributes that will be populated by the methods.
     stoichiometry_limits: Optional[Result] = None
     """The stoichiometry limits and electrode capacities."""
@@ -149,21 +155,10 @@ class DMA(BaseModel):
     dma_result: Optional[Result] = None
     """The degradation mode analysis results."""
 
-    _ocp_pe: List[
-        Union[
-            Callable[[NDArray[np.float64]], NDArray[np.float64]],
-            PPoly,
-            sp.Expr,
-        ]
-    ] = []
+    class Config:
+        """Pydantic configuration settings."""
 
-    _ocp_ne: List[
-        Union[
-            Callable[[NDArray[np.float64]], NDArray[np.float64]],
-            PPoly,
-            sp.Expr,
-        ]
-    ] = []
+        arbitrary_types_allowed = True
 
     def downsample_ocv(
         self,
@@ -199,176 +194,9 @@ class DMA(BaseModel):
         )
         return self.input_data
 
-    def _get_ocp_functions(
-        self,
-        ocp_list: List[Callable[[NDArray[np.float64]], NDArray[np.float64]]],
-    ) -> List[Callable[[NDArray[np.float64]], NDArray[np.float64]]]:
-        """Helper method to initialize and return OCP functions."""
-        ocp_functions = []
-        for ocp in ocp_list:
-            if isinstance(ocp, sp.Expr):
-                ocp_functions.append(sp.lambdify(ocp.free_symbols.pop(), ocp, "numpy"))
-            else:
-                ocp_functions.append(ocp)
-        return ocp_functions
-
-    @property
-    def ocp_pe(
-        self,
-    ) -> List[Callable[[NDArray[np.float64]], NDArray[np.float64]]]:
-        """Return a list of positive electrode OCPs as a function of stoichiometry.
-
-        Returns:
-            List[Callable[[NDArray[np.float64]], NDArray[np.float64]]]:
-                A list of positive electrode OCPs as a function of stoichiometry.
-        """
-        if not hasattr(self, "_f_ocp_pe"):
-            self._f_ocp_pe = self._get_ocp_functions(self._ocp_pe)
-        return self._f_ocp_pe
-
-    @ocp_pe.setter
-    def ocp_pe(
-        self,
-        value: List[Callable[[NDArray[np.float64]], NDArray[np.float64]]],
-    ) -> None:
-        """Set the OCP data for the positive electrode.
-
-        Args:
-            value (List[Callable[[NDArray[np.float64]], NDArray[np.float64]]]):
-                The OCP data for the positive electrode. Provide a list of functions
-                that take stoichiometry as input and return OCP as output.
-        """
-        self._f_ocp_pe = value
-        self._ocp_pe = value
-
-    @property
-    def ocp_ne(
-        self,
-    ) -> List[Callable[[NDArray[np.float64]], NDArray[np.float64]]]:
-        """Return a list of negative electrode OCPs as a function of stoichiometry.
-
-        Returns:
-            List[Callable[[NDArray[np.float64]], NDArray[np.float64]]]:
-                A list of negative electrode OCPs as a function of stoichiometry.
-        """
-        if not hasattr(self, "_f_ocp_ne"):
[…]
