+++ tests/analysis/test_dma.py
@@ -116,120 +116,6 @@ def test_get_gradient_invalid_input():
         OCP._get_gradient(42)
 
 
-def test_add_ocp_from_data_pe(stoichiometry_data, ocp_data):
-    """Test the add_ocp_from_data method."""
-    dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
-    dma.add_ocp_from_data(
-        stoichiometry_data, ocp_data, electrode="pe", interpolation_method="cubic"
-    )
-    assert dma.ocp_pe[0] is not None
-    assert callable(dma.ocp_pe[0])
-    assert np.isclose(dma.ocp_pe[0](0.4), np.sin(0.4))
-
-
-def test_add_ocp_from_data_ne(stoichiometry_data, ocp_data):
-    """Test the add_ocp_from_data method."""
-    dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
-    dma.add_ocp_from_data(stoichiometry_data, ocp_data, electrode="ne")
-    assert dma.ocp_ne[0] is not None
-    assert callable(dma.ocp_ne[0])
-    assert np.isclose(dma.ocp_ne[0](0.1), np.sin(0.1))
-
-
-def test_add_ocp_from_data_linear_interpolation(stoichiometry_data, ocp_data):
-    """Test the add_ocp_from_data method with linear interpolation."""
-    dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
-    dma.add_ocp_from_data(
-        stoichiometry_data, ocp_data, electrode="pe", interpolation_method="linear"
-    )
-    assert dma.ocp_pe[0] is not None
-    assert callable(dma.ocp_pe[0])
-    assert np.isclose(dma.ocp_pe[0](0.4), np.sin(0.4))
-
-
-def test_add_ocp_from_data_cubic_interpolation(stoichiometry_data, ocp_data):
-    """Test the add_ocp_from_data method with cubic interpolation."""
-    dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
-    dma.add_ocp_from_data(
-        stoichiometry_data, ocp_data, electrode="pe", interpolation_method="cubic"
-    )
-    assert dma.ocp_pe[0] is not None
-    assert callable(dma.ocp_pe[0])
-    assert np.isclose(dma.ocp_pe[0](0.4), np.sin(0.4))
-
-
-def test_add_ocp_from_data_multiple_components(stoichiometry_data, ocp_data):
-    """Test the add_ocp_from_data method with multiple components."""
-    dma = DMA(input_data=Result(base_dataframe=pl.DataFrame({}), info={}))
-    dma.add_ocp_from_data(
-        stoichiometry_data,
-        ocp_data,
-        electrode="pe",
-    )
-    dma.add_ocp_from_data(
-        stoichiometry_data,
-        ocp_data,
-        electrode="pe",
-    )
-    assert len(dma.ocp_pe) == 2
-    assert np.isclose(dma.ocp_pe[0](0.4), np.sin(0.4))
-    assert np.isclose(dma.ocp_pe[1](0.8), np.sin(0.8))
-
-
-def test_add_ocp_from_expression():
-    """Test the add_ocp_from_expression method."""
[…]
