### 2024-11-11 #7294 @yaomz16
> 👋 @TomTranter, @EricaEgg & @yaomz16, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi Sophie, sorry for the slow reply but I won't be able to help review the paper
