### 2024-11-11 ed8dd47
Add a CompositeOCP class
