+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -19,16 +19,73 @@ from pyprobe.result import Result
 from pyprobe.typing import FilterToCycleType, PyProBEDataType
 
 
-class OCP:
-    """A class for single-component electrode open circuit potential data."""
+def _get_gradient(
+    any_function: Union[
+        Callable[[NDArray[np.float64]], NDArray[np.float64]],
+        PPoly,
+        sp.Expr,
+    ]
+) -> Callable[[NDArray[np.float64]], NDArray[np.float64]]:
+    """Retrieve the gradient of the OCP function.
 
-    def set_from_data(
+    Args:
+        any_function: The OCP function.
+
+    Returns:
+        Callable[[NDArray[np.float64]], NDArray[np.float64]]:
+            The gradient of the OCP.
+    """
+    if isinstance(any_function, PPoly):
+        return any_function.derivative()
+    elif isinstance(any_function, sp.Expr):
+        free_symbols = any_function.free_symbols
+        sto = free_symbols.pop()
+        gradient = sp.diff(any_function, sto)
+        return sp.lambdify(sto, gradient, "numpy")
+    elif callable(any_function):
+
+        def function_derivative(
+            sto: NDArray[np.float64],
+        ) -> NDArray[np.float64]:
+            """Numerically calculate the derivative."""
+            return np.gradient(any_function(sto), sto)
+
+        return function_derivative
+    else:
+        raise ValueError(
+            "OCP is not in a differentiable format. OCP must be a"
+            " PPoly object, a sympy expression or a callable function with a "
+            "single NDArray input and single NDArray output."
+        )
+
+
+class OCP:
+    """A class for single-component electrode open circuit potential data.
+
+    Args:
+        ocp_function (Union[Callable[[NDArray[np.float64]],
+            NDArray[np.float64]], PPoly, sp.Expr]):
+            The OCP function for the electrode.
+    """
+
+    def __init__(
         self,
+        ocp_function: Union[
+            Callable[[NDArray[np.float64]], NDArray[np.float64]],
+            PPoly,
+            sp.Expr,
+        ],
+    ) -> None:
+        """Initialize the OCP object."""
+        self.ocp_function = ocp_function
+
+    @staticmethod
+    def from_data(
         stoichiometry: NDArray[np.float64],
         ocp: NDArray[np.float64],
         interpolation_method: Literal["linear", "cubic", "Pchip", "Akima"] = "linear",
-    ) -> None:
-        """Provide a OCP data for a given electrode.
+    ) -> "OCP":
+        """Create an OCP object from stoichiometry and OCP data.
 
         Appends to the ocp list for the given electrode. Composite electrodes require
         multiple calls to this method to provide the OCP data for each component.
@@ -39,6 +96,9 @@ class OCP:
             interpolation_method
                 (Literal["linear", "cubic", "Pchip", "Akima"], optional):
                 The interpolation method to use. Defaults to "linear".
+
+        Returns:
+            OCP: The OCP object.
         """
         interpolator = {
             "linear": smoothing.linear_interpolator,
@@ -46,91 +106,136 @@ class OCP:
             "Pchip": smoothing.pchip_interpolator,
             "Akima": smoothing.akima_interpolator,
         }[interpolation_method]
-        self.base_form = interpolator(stoichiometry, ocp)
+        return OCP(interpolator(stoichiometry, ocp))
 
-    def set_from_expression(self, sympy_expression: sp.Expr) -> None:
-        """Provide an OCP in analytical function form.
+    @staticmethod
+    def from_expression(sympy_expression: sp.Expr) -> "OCP":
+        """Create an OCP object from a sympy expression.
 
         Args:
[…]
