### 2024-11-11 f7225b7
Reinstate logic for composite ocp fitting
