### 2024-11-11 68ecc9e
Fix tests for calculate_dma_parameters and downsample_ocv methods
