### 2024-11-11 f94d79c
BREAKING: remove fit_ocv method
