+++ tests/analysis/test_dma.py
@@ -5,7 +5,6 @@ import numpy as np
 import polars as pl
 import pytest
 import sympy as sp
-from pydantic import ValidationError
 from scipy.interpolate import PPoly
 
 import pyprobe.analysis.base.degradation_mode_analysis_functions as dma_functions
@@ -507,135 +506,6 @@ def test_run_batch_dma():
 
 n_points = 1000
 z = np.linspace(0, 1, 1000)
-
-
-def test_fit_ocv():
-    """Test the fit_ocv method."""
-    capacity = np.linspace(0, 1, n_points)
-    x_real = [0.8, 0.1, 0.1, 0.7]
-    x_pe_real = np.linspace(x_real[0], x_real[1], n_points)
-    x_ne_real = np.linspace(x_real[2], x_real[3], n_points)
-
-    # test charge
-    voltage = nmc_LGM50_ocp_Chen2020(x_pe_real) - graphite_LGM50_ocp_Chen2020(x_ne_real)
-
-    result = Result(
-        base_dataframe=pl.DataFrame(
-            {"Voltage [V]": voltage, "Capacity [Ah]": capacity}
-        ),
-        info={},
-    )
-    dma = DMA(input_data=result)
-    x_guess = [0.8, 0.4, 0.2, 0.6]
-    params, fit = dma.fit_ocv(
-        x_ne=z,
-        x_pe=z,
-        ocp_ne=graphite_LGM50_ocp_Chen2020(z),
-        ocp_pe=nmc_LGM50_ocp_Chen2020(z),
-        x_guess=x_guess,
-    )
-    assert isinstance(params, Result)
-    assert params.data.columns == [
-        "x_pe low SOC",
-        "x_pe high SOC",
-        "x_ne low SOC",
-        "x_ne high SOC",
-        "Cell Capacity [Ah]",
-        "Cathode Capacity [Ah]",
-        "Anode Capacity [Ah]",
-        "Li Inventory [Ah]",
-    ]
-
-    assert isinstance(fit, Result)
-    assert fit.data.columns == [
-        "Capacity [Ah]",
-        "SOC",
-        "Input Voltage [V]",
-        "Fitted Voltage [V]",
-        "Input dSOCdV [1/V]",
-        "Fitted dSOCdV [1/V]",
-        "Input dVdSOC [V]",
-        "Fitted dVdSOC [V]",
-    ]
-
-    param_values = list(params.data.row(0))
-    np.testing.assert_allclose(
-        np.array(param_values)[:4],
-        np.array(x_real),
-        rtol=1e-4,
-    )
-
-    # test discharge
-    voltage = np.flip(voltage)
-    capacity = -1 * capacity
-    result = Result(
-        base_dataframe=pl.DataFrame(
-            {"Voltage [V]": voltage, "Capacity [Ah]": capacity}
-        ),
-        info={},
-    )
-    dma = DMA(input_data=result)
-    params, _ = dma.fit_ocv(
-        x_ne=z,
-        x_pe=z,
-        ocp_ne=graphite_LGM50_ocp_Chen2020(z),
-        ocp_pe=nmc_LGM50_ocp_Chen2020(z),
-        x_guess=x_guess,
-    )
-
-    param_values = list(params.data.row(0))
-    np.testing.assert_allclose(
-        np.array(param_values)[:4],
-        np.array(x_real),
-        rtol=1e-4,
-    )
-
-    result = Result(
-        base_dataframe=pl.DataFrame({"Voltage [V]": voltage, "Time [s]": capacity}),
-        info={},
-    )
-    dma = DMA(input_data=result)
[…]
