+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -667,148 +667,6 @@ class DMA(BaseModel):
 
         return self.stoichiometry_limits, self.fitted_OCV
 
-    def fit_ocv(
-        self,
-        x_ne: NDArray[np.float64],
-        x_pe: NDArray[np.float64],
-        ocp_ne: NDArray[np.float64],
-        ocp_pe: NDArray[np.float64],
-        x_guess: NDArray[np.float64],
-        fitting_target: str = "OCV",
-        optimizer: Optional[str] = None,
-    ) -> Tuple[Result, Result]:
-        """Fit half cell open circuit potential curves to full cell OCV data.
-
-        Based on the code from :footcite:t:`Kirkaldy_batteryDAT_2023`. The fitting
-        algorithm is based on the scipy.optimize.minimize function.
-
-        Args:
-            x_ne (NDArray[np.float64]): The anode stoichiometry data.
-            x_pe (NDArray[np.float64]): The cathode stoichiometry data.
-            ocp_ne (NDArray[np.float64]): The anode OCP data.
-            ocp_pe (NDArray[np.float64]): The cathode OCP data.
-            x_guess (NDArray[np.float64]): The initial guess for the fit.
-            fitting_target (str, optional):
-                The target for the curve fitting. Defaults to "OCV".
-            optimizer (str, optional):
-                The optimization algorithm to use. Defaults to None. If None, the
-                optimizer will be set to scipy.optimize.minimize for
-                fitting_target="OCV" and scipy.optimize.differential_evolution for
-                fitting_target="dQdV" or "dVdQ".
-
-        Returns:
-            Tuple[Result, Result]:
-                - Result: The stoichiometry limits and electrode capacities.
-                - Result: The fitted OCV data.
-        """
-        if "SOC" in self.input_data.column_list:
-            required_columns = ["Voltage [V]", "Capacity [Ah]", "SOC"]
-            validator = AnalysisValidator(
-                input_data=self.input_data, required_columns=required_columns
-            )
-            voltage, capacity, SOC = validator.variables
-            cell_capacity = np.abs(np.ptp(capacity)) / np.abs(np.ptp(SOC))
-        else:
-            required_columns = ["Voltage [V]", "Capacity [Ah]"]
-            validator = AnalysisValidator(
-                input_data=self.input_data, required_columns=required_columns
-            )
-            voltage, capacity = validator.variables
-            cell_capacity = np.abs(np.ptp(capacity))
-            SOC = (capacity - capacity.min()) / cell_capacity
-
-        dVdSOC = np.gradient(voltage, SOC)
-        dSOCdV = np.gradient(SOC, voltage)
-
-        if optimizer is None:
-            if fitting_target == "OCV":
-                optimizer = "minimize"
-            else:
-                optimizer = "differential_evolution"
-
-        fitting_target_data = {
-            "OCV": voltage,
-            "dQdV": dSOCdV,
-            "dVdQ": dVdSOC,
-        }[fitting_target]
-
-        x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi = dma_functions.ocv_curve_fit(
-            SOC,
-            fitting_target_data,
-            x_pe,
-            ocp_pe,
-            x_ne,
-            ocp_ne,
-            fitting_target,
-            optimizer,
-            x_guess,
-        )
-
-        (
-            pe_capacity,
-            ne_capacity,
-            li_inventory,
-        ) = dma_functions.calc_electrode_capacities(
-            x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi, cell_capacity
-        )
-
[…]
