### 2024-11-11 d11b7e6
Fix tests for batch DMA
