### 2024-11-11 df4d496
Fix tests for averaging input ocvs
