+++ tests/analysis/test_dma.py
@@ -653,7 +653,12 @@ def test_average_ocvs(BreakinCycles_fixture):
     """Test the average_ocvs method."""
     break_in = BreakinCycles_fixture.cycle(0)
     break_in.set_SOC()
-    dma = DMA.average_ocvs(input_data=break_in, charge_filter="constant_current(1)")
+    dma = DMA.average_ocvs(
+        input_data=break_in,
+        charge_filter="constant_current(1)",
+        ocp_pe=OCP(nmc_LGM50_ocp_Chen2020),
+        ocp_ne=OCP(graphite_LGM50_ocp_Chen2020),
+    )
     assert math.isclose(dma.input_data.get_only("Voltage [V]")[0], 3.14476284763849)
     assert math.isclose(dma.input_data.get_only("Voltage [V]")[-1], 4.170649780122139)
     np.testing.assert_allclose(
@@ -661,7 +666,11 @@ def test_average_ocvs(BreakinCycles_fixture):
     )
     # test invalid input
     with pytest.raises(ValueError):
-        DMA.average_ocvs(input_data=break_in.charge(0))
+        DMA.average_ocvs(
+            input_data=break_in.charge(0),
+            ocp_pe=OCP(nmc_LGM50_ocp_Chen2020),
+            ocp_ne=OCP(graphite_LGM50_ocp_Chen2020),
+        )
 
 
 def test_calc_full_cell_ocv_composite():
