### 2024-11-12 ec00a6c
Create ocp abstract class
