### 2024-11-12 15e1c7d
Remove type descriptions from docstrings as auto generated in sphinx
