+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -33,8 +33,7 @@ def _get_gradient(
         any_function: The OCP function.
 
     Returns:
-        Callable[[NDArray[np.float64]], NDArray[np.float64]]:
-            The gradient of the OCP.
+        The gradient of the OCP.
     """
     if isinstance(any_function, PPoly):
         return any_function.derivative()
@@ -80,9 +79,7 @@ class OCP(_AbstractOCP):
     """A class for single-component electrode open circuit potential data.
 
     Args:
-        ocp_function (Union[Callable[[NDArray[np.float64]],
-            NDArray[np.float64]], PPoly, sp.Expr]):
-            The OCP function for the electrode.
+        ocp_function: The OCP function for the electrode.
     """
 
     def __init__(
@@ -108,14 +105,12 @@ class OCP(_AbstractOCP):
         multiple calls to this method to provide the OCP data for each component.
 
         Args:
-            stoichiometry (NDArray[np.float64]): The stoichiometry data.
-            ocp (NDArray[np.float64]): The OCP data.
-            interpolation_method
-                (Literal["linear", "cubic", "Pchip", "Akima"], optional):
-                The interpolation method to use. Defaults to "linear".
+            stoichiometry: The stoichiometry data.
+            ocp: The OCP data.
+            interpolation_method: The interpolation method to use. Defaults to "linear".
 
         Returns:
-            OCP: The OCP object.
+            The OCP object.
         """
         interpolator = {
             "linear": smoothing.linear_interpolator,
@@ -133,7 +128,7 @@ class OCP(_AbstractOCP):
             sympy_expression: A sympy expression for the OCP.
 
         Returns:
-            OCP: The OCP object.
+            The OCP object.
         """
         return OCP(sympy_expression)
 
@@ -161,11 +156,8 @@ class CompositeOCP(_AbstractOCP):
     """A class for composite electrode open circuit potential data.
 
     Args:
-        ocp_list (List[Union[Callable[[NDArray[np.float64]],
-        NDArray[np.float64]], PPoly, sp.Expr]]):
-            A list of OCP functions for the composite electrode.
-
-        ocp_vector (NDArray[np.float64]): The OCP vector for the composite electrode.
+        ocp_list: A list of OCP functions for the composite electrode.
+        ocp_vector: The OCP vector for the composite electrode.
     """
 
     def __init__(
@@ -195,15 +187,11 @@ class CompositeOCP(_AbstractOCP):
         """Create a CompositeOCP object from stoichiometry and OCP data.
 
         Args:
-            stoichiometry_comp1 (NDArray[np.float64]):
-                The stoichiometry data for the first component.
-            ocp_comp1 (NDArray[np.float64]): The OCP data for the first component.
-            stoichiometry_comp2 (NDArray[np.float64]):
-                The stoichiometry data for the second component.
-            ocp_comp2 (NDArray[np.float64]): The OCP data for the second component.
-            interpolation_method
-                (Literal["linear", "cubic", "Pchip", "Akima"], optional):
-                The interpolation method to use. Defaults to "linear".
+            stoichiometry_comp1: The stoichiometry data for the first component.
+            ocp_comp1: The OCP data for the first component.
+            stoichiometry_comp2: The stoichiometry data for the second component.
+            ocp_comp2: The OCP data for the second component.
+            interpolation_method: The interpolation method to use. Defaults to "linear".
         """
         # Determine common voltage range for anode components
         composite_voltage_limits = (
@@ -296,14 +284,12 @@ class DMA(BaseModel):
         logic for the :func:`pyprobe.analysis.smoothing.Smoothing.downsample` method.
 
         Args:
-            sampling_interval (float): The sampling interval for the downsampled data.
[…]
