### 2024-11-12 15a12a9
Create new resistance correction class and add average dma to it
