### 2024-11-12 03601de
Remove DMA class, module just holds methods
