+++ tests/analysis/test_dma.py
@@ -8,15 +8,9 @@ import sympy as sp
 from scipy.interpolate import PPoly
 
 import pyprobe.analysis.base.degradation_mode_analysis_functions as dma_functions
+import pyprobe.analysis.degradation_mode_analysis as dma
 from pyprobe.analysis import smoothing
-from pyprobe.analysis.degradation_mode_analysis import (
-    DMA,
-    OCP,
-    BatchDMA,
-    CompositeOCP,
-    _get_gradient,
-    average_ocvs,
-)
+from pyprobe.analysis.degradation_mode_analysis import OCP, CompositeOCP, _get_gradient
 from pyprobe.result import Result
 
 """Tests for the OCP class in the degradation mode analysis module."""
@@ -195,17 +189,14 @@ def pe_ocp_fixture():
 
 def test_f_OCV(ne_ocp_fixture, pe_ocp_fixture):
     """Test the f_OCV method."""
-    dma = DMA(
-        input_data=Result(base_dataframe=pl.DataFrame({}), info={}),
-        ocp_ne=ne_ocp_fixture,
-        ocp_pe=pe_ocp_fixture,
-    )
     x_pe_lo = 0.8
     x_pe_hi = 0.1
     x_ne_lo = 0.1
     x_ne_hi = 0.7
     soc = np.linspace(0, 1, 100)
     ocv = dma._f_OCV(
+        pe_ocp_fixture,
+        ne_ocp_fixture,
         soc,
         x_pe_lo,
         x_pe_hi,
@@ -221,18 +212,17 @@ def test_f_OCV(ne_ocp_fixture, pe_ocp_fixture):
 
 def test_f_grad_OCV(ne_ocp_fixture, pe_ocp_fixture):
     """Test the f_grad_OCV method."""
-    dma = DMA(
-        input_data=Result(base_dataframe=pl.DataFrame({}), info={}),
-        ocp_ne=ne_ocp_fixture,
-        ocp_pe=pe_ocp_fixture,
-    )
     soc = np.linspace(0, 1, 1000)
     x_pe_lo = 0.8
     x_pe_hi = 0.1
     x_ne_lo = 0.1
     x_ne_hi = 0.7
-    d_ocv = dma._f_grad_OCV(soc, x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi)
-    ocv_pts = dma._f_OCV(soc, x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi)
+    d_ocv = dma._f_grad_OCV(
+        pe_ocp_fixture, ne_ocp_fixture, soc, x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi
+    )
+    ocv_pts = dma._f_OCV(
+        pe_ocp_fixture, ne_ocp_fixture, soc, x_pe_lo, x_pe_hi, x_ne_lo, x_ne_hi
+    )
     numerical_d_ocv = np.gradient(ocv_pts, soc)
     np.testing.assert_allclose(d_ocv, numerical_d_ocv)
 
@@ -249,19 +239,17 @@ def test_run_ocv_curve_fit(ne_ocp_fixture, pe_ocp_fixture):
     ocv_ne = graphite_LGM50_ocp_Chen2020(x_ne)
     soc = np.linspace(0, 1, 10000)
     ocv_target = ocv_pe - ocv_ne
-    dma = DMA(
-        input_data=Result(
-            base_dataframe=pl.DataFrame(
-                {"Voltage [V]": ocv_target, "Capacity [Ah]": soc}
-            ),
-            info={},
-        ),
-        ocp_ne=ne_ocp_fixture,
-        ocp_pe=pe_ocp_fixture,
+
+    input_data = Result(
[…]
