+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -9,7 +9,7 @@ import numpy as np
 import polars as pl
 import sympy as sp
 from numpy.typing import NDArray
-from pydantic import BaseModel, validate_call
+from pydantic import validate_call
 from scipy import optimize
 from scipy.interpolate import PPoly
 
@@ -243,558 +243,516 @@ class CompositeOCP(_AbstractOCP):
         return _get_gradient(self.eval)
 
 
-# 1. Define the class as a Pydantic BaseModel.
-class DMA(BaseModel):
-    """A class for degradation mode analysis methods."""
+def _f_OCV(
+    ocp_pe: Union[OCP, CompositeOCP],
+    ocp_ne: Union[OCP, CompositeOCP],
+    SOC: NDArray[np.float64],
+    x_pe_lo: NDArray[np.float64],
+    x_pe_hi: NDArray[np.float64],
+    x_ne_lo: NDArray[np.float64],
+    x_ne_hi: NDArray[np.float64],
+) -> NDArray[np.float64]:
+    """Calculate the full cell OCV as a function of SOC.
 
-    # 2. Define the input_data attribute, giving it a type
-    input_data: PyProBEDataType
-    """The input data for the degradation mode analysis."""
+    Args:
+        SOC: The full cell SOC.
+        ocp_pe: The positive electrode OCP as a function of stoichiometry.
+        ocp_ne: The negative electrode OCP as a function of stoichiometry.
+        x_pe_lo: The cathode stoichiometry at lowest cell SOC.
+        x_pe_hi: The cathode stoichiometry at highest cell SOC.
+        x_ne_lo: The cathode stoichiometry at lowest cell SOC.
+        x_ne_hi: The anode stoichiometry at highest cell SOC.
 
-    ocp_pe: OCP | CompositeOCP
-    """An object containing the positive electrode OCP functions."""
+    Returns:
+        A function to calculate the full cell OCV as a function of SOC.
+    """
+    # Calculate the stoichiometry at the given SOC for each electrode
+    z_pe = x_pe_lo + (x_pe_hi - x_pe_lo) * SOC
+    z_ne = x_ne_lo + (x_ne_hi - x_ne_lo) * SOC
 
-    ocp_ne: OCP | CompositeOCP
-    """An object containing the negative electrode OCP functions."""
+    # Return the full cell OCV
+    return ocp_pe.eval(z_pe) - ocp_ne.eval(z_ne)
 
-    # 3. Define the attributes that will be populated by the methods.
-    stoichiometry_limits: Optional[Result] = None
-    """The stoichiometry limits and electrode capacities."""
-    fitted_OCV: Optional[Result] = None
-    """The fitted OCV data."""
-    dma_result: Optional[Result] = None
-    """The degradation mode analysis results."""
 
-    class Config:
-        """Pydantic configuration settings."""
+def _f_grad_OCV(
+    ocp_pe: Union[OCP, CompositeOCP],
+    ocp_ne: Union[OCP, CompositeOCP],
+    SOC: NDArray[np.float64],
+    x_pe_lo: NDArray[np.float64],
+    x_pe_hi: NDArray[np.float64],
+    x_ne_lo: NDArray[np.float64],
+    x_ne_hi: NDArray[np.float64],
+) -> NDArray[np.float64]:
+    """Calculate the full cell OCV gradient as a function of SOC.
 
-        arbitrary_types_allowed = True
+    Derivative is calculated using the chain rule:
+    d(OCV)/d(SOC) = d(OCV)/d(z_pe) * d(z_pe)/d(SOC)
+                    - d(OCV)/d(z_ne) * d(z_ne)/d(SOC)
+    Args:
+        SOC: The full cell SOC.
[…]
