### 2024-11-12 3469647
Add validate_call decorators to public functions
