### 2024-11-12 0edc07e
Remove type checking from functionality of AnalysisValidator
