+++ tests/analysis/test_utils.py
@@ -5,7 +5,6 @@ import polars as pl
 import pytest
 
 import pyprobe.analysis.utils as utils
-from pyprobe.filters import Experiment
 from pyprobe.result import Result
 
 
@@ -36,7 +35,6 @@ def test_base_analysis(input_data_fixture):
     analysis = utils.AnalysisValidator(
         input_data=input_data_fixture, required_columns=["x"]
     )
-    assert analysis.validate_input_data_type() == analysis
     assert analysis.validate_required_columns() == analysis
 
     with pytest.raises(ValueError):
@@ -50,13 +48,6 @@ def test_base_analysis(input_data_fixture):
     x, y = analysis.variables
     np.testing.assert_array_equal(x, np.array([1, 2, 3]))
     np.testing.assert_array_equal(y, np.array([4, 5, 6]))
-
-    with pytest.raises(ValueError):
-        analysis = utils.AnalysisValidator(
-            input_data=input_data_fixture,
-            required_columns=["z"],
-            required_type=Experiment,
-        )
 
     analysis = utils.AnalysisValidator(
         input_data=input_data_fixture, required_columns=["Units [mAh]"]
