### 2024-11-12 e1619ed
Add method for extracting only ocv points from pulsing
