+++ tests/analysis/__init__.py
@@ -1,0 +1,1 @@
+"""Module for testing the analysis module."""
