+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -1,12 +1,11 @@
 """Module for degradation mode analysis methods."""
 
-import concurrent.futures
-import copy
 from abc import ABC, abstractmethod
 from typing import Any, Callable, Dict, List, Literal, Optional, Tuple, Union, cast
 
 import numpy as np
 import polars as pl
+import ray
 import sympy as sp
 from numpy.typing import NDArray
 from pydantic import ConfigDict, validate_call
@@ -675,15 +674,13 @@ def run_ocv_curve_fit(
 
 @validate_call
 def quantify_degradation_modes(
-    input_stoichiometry_limits: Result, reference_stoichiometry_limits: Result
+    stoichiometry_limits_list: List[Result],
 ) -> Result:
     """Quantify the change in degradation modes between at least two OCV fits.
 
     Args:
-        input_stoichiometry_limits:
-            A result object containing the stoichiometry limits for the OCV fits.
-        reference_stoichiometry_limits:
-            A result object containing the beginning of life stoichiometry limits.
+        stoichiometry_limits_list: A list of Result objects containing the
+        stoichiometry limits for the OCV fits.
 
     Returns:
         A result object containing the SOH, LAM_pe, LAM_ne, and LLI for each of
@@ -695,36 +692,27 @@ def quantify_degradation_modes(
         "Anode Capacity [Ah]",
         "Li Inventory [Ah]",
     ]
-    AnalysisValidator(
-        input_data=reference_stoichiometry_limits, required_columns=required_columns
-    )
-    if input_stoichiometry_limits is None:
-        raise ValueError("No stoichiometry limits have been calculated.")
-    AnalysisValidator(
-        input_data=input_stoichiometry_limits, required_columns=required_columns
-    )
+    for stoichiometry_limits in stoichiometry_limits_list:
+        AnalysisValidator(
+            input_data=stoichiometry_limits, required_columns=required_columns
+        )
 
-    electrode_capacity_results = [
-        reference_stoichiometry_limits,
-        input_stoichiometry_limits,
-    ]
     cell_capacity = utils.assemble_array(
-        electrode_capacity_results, "Cell Capacity [Ah]"
+        stoichiometry_limits_list, "Cell Capacity [Ah]"
     )
     pe_capacity = utils.assemble_array(
-        electrode_capacity_results, "Cathode Capacity [Ah]"
+        stoichiometry_limits_list, "Cathode Capacity [Ah]"
     )
-    ne_capacity = utils.assemble_array(
-        electrode_capacity_results, "Anode Capacity [Ah]"
-    )
-    li_inventory = utils.assemble_array(electrode_capacity_results, "Li Inventory [Ah]")
+    ne_capacity = utils.assemble_array(stoichiometry_limits_list, "Anode Capacity [Ah]")
+    li_inventory = utils.assemble_array(stoichiometry_limits_list, "Li Inventory [Ah]")
     SOH, LAM_pe, LAM_ne, LLI = dma_functions.calculate_dma_parameters(
         cell_capacity, pe_capacity, ne_capacity, li_inventory
     )
 
-    dma_result = electrode_capacity_results[0].clean_copy(
+    dma_result = stoichiometry_limits_list[0].clean_copy(
         pl.DataFrame(
             {
+                "Index": np.arange(len(stoichiometry_limits_list)),
                 "SOH": SOH[:, 0],
                 "LAM_pe": LAM_pe[:, 0],
                 "LAM_ne": LAM_ne[:, 0],
@@ -732,7 +720,11 @@ def quantify_degradation_modes(
             }
         )
     )
+    dma_result.base_dataframe = dma_result.base_dataframe.with_columns(
+        pl.col("Index").cast(pl.Int64)
+    )
[…]
