+++ requirements.txt
@@ -4,6 +4,8 @@
 #
 #    pip-compile
 #
+aiosignal==1.3.1
+    # via ray
 altair==5.4.0
     # via streamlit
 annotated-types==0.7.0
@@ -27,7 +29,9 @@ certifi==2024.7.4
 charset-normalizer==3.3.2
     # via requests
 click==8.1.7
-    # via streamlit
+    # via
+    #   ray
+    #   streamlit
 contourpy==1.3.0
     # via matplotlib
 cycler==0.12.1
@@ -40,8 +44,14 @@ executing==2.0.1
     # via stack-data
 fastexcel==0.11.5
     # via PyProBE (pyproject.toml)
+filelock==3.16.1
+    # via ray
 fonttools==4.53.1
     # via matplotlib
+frozenlist==1.5.0
+    # via
+    #   aiosignal
+    #   ray
 gitdb==4.0.11
     # via gitpython
 gitpython==3.1.43
@@ -59,7 +69,9 @@ jinja2==3.1.4
 joblib==1.4.2
     # via scikit-learn
 jsonschema==4.23.0
-    # via altair
+    # via
+    #   altair
+    #   ray
 jsonschema-specifications==2023.12.1
     # via jsonschema
 kaleido==0.2.1
@@ -78,6 +90,8 @@ mdurl==0.1.2
     # via markdown-it-py
 mpmath==1.3.0
     # via sympy
+msgpack==1.1.0
+    # via ray
 narwhals==1.8.0
     # via altair
 numpy==1.26.4
@@ -103,6 +117,7 @@ packaging==24.1
     #   matplotlib
     #   plotly
     #   pooch
+    #   ray
     #   streamlit
     #   xarray
 pandas==2.2.2
@@ -129,7 +144,9 @@ pooch==1.8.2
 prompt-toolkit==3.0.47
     # via ipython
 protobuf==5.27.3
-    # via streamlit
+    # via
+    #   ray
+    #   streamlit
 ptyprocess==0.7.0
     # via pexpect
 pure-eval==0.2.3
@@ -159,6 +176,10 @@ python-dateutil==2.9.0.post0
 pytz==2024.1
     # via pandas
 pyyaml==6.0.2
+    # via
+    #   PyProBE (pyproject.toml)
+    #   ray
+ray==2.39.0
     # via PyProBE (pyproject.toml)
 referencing==0.35.1
     # via
@@ -167,6 +188,7 @@ referencing==0.35.1
 requests==2.32.3
     # via
     #   pooch
+    #   ray
     #   streamlit
 rich==13.7.1
     # via streamlit
