### 2024-11-13 36e7a24
Report ray parallel performance and ensure shutdown
