### 2024-11-13 f76e6af
Include stoichiometry limits in result of quantify degradation modes
