### 2024-11-13 9a2ab5c
Allow kwargs to pass to interpolators
