### 2024-11-13 bee0f9e
Add tests for clean copy
