### 2024-11-13 #7474 @editorialbot
[REVIEW]: PyProBE: Python Processing for Battery Experiments
**Submitting author:** @tomjholland (<a href="http://orcid.org/0009-0009-0670-1901">Thomas Holland</a>)
**Repository:** https://github.com/ImperialCollegeLondon/PyProBE
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
