### 2024-11-15 d0179c7
Add kwargs to linear interpolator
