### 2024-11-15 4e71545
Add dma method for sequentially performing the ocv fitting
