### 2024-11-15 e280234
Add deprecated type stubs
