### 2024-11-15 89b06ae
Add standalone method for getting resistances from each pulse
