+++ tests/analysis/test_pulsing.py
@@ -14,24 +14,26 @@ def Pulsing_fixture(procedure_fixture):
     procedure_fixture.set_SOC(
         reference_charge=procedure_fixture.experiment("Break-in Cycles").charge(-1)
     )
-    return Pulsing(input_data=procedure_fixture.experiment("Discharge Pulses"))
+    return procedure_fixture.experiment("Discharge Pulses")
 
 
 def test_pulse(Pulsing_fixture):
     """Test the pulse method."""
-    pulse = Pulsing_fixture.pulse(0)
+    pulse_obj = Pulsing(input_data=Pulsing_fixture)
+    pulse = pulse_obj.pulse(0)
     assert pulse.data["Time [s]"][0] == 483572.397
     assert (pulse.data["Step"] == 10).all()
 
-    pulse = Pulsing_fixture.pulse(6)
+    pulse = pulse_obj.pulse(6)
     assert pulse.data["Time [s]"][0] == 531149.401
     assert (pulse.data["Step"] == 10).all()
 
 
 def test_pulse_summary(Pulsing_fixture):
     """Test the pulse_summary method."""
-    pulse_summary = Pulsing_fixture.pulse_summary([10])
-    assert isinstance(Pulsing_fixture.pulse_summary(), Result)
+    pulse_obj = Pulsing(input_data=Pulsing_fixture)
+    pulse_summary = pulse_obj.pulse_summary([10])
+    assert isinstance(pulse_obj.pulse_summary(), Result)
     assert isinstance(pulse_summary, Result)
     assert pulse_summary.get("OCV [V]")[0] == 4.1919
     assert np.isclose(pulse_summary.get("R0 [Ohms]")[0], (4.1558 - 4.1919) / -0.0199936)
@@ -40,13 +42,17 @@ def test_pulse_summary(Pulsing_fixture):
     )
 
 
-def test_get_ocv_curve(procedure_fixture):
+def test_get_resistances(Pulsing_fixture):
+    """Test the get_resistances method."""
+    resistances = pulsing.get_resistances(Pulsing_fixture, [10])
+    assert isinstance(resistances, Result)
+    assert resistances.get("R0 [Ohms]")[0] == (4.1558 - 4.1919) / -0.0199936
+    assert resistances.get("R_10s [Ohms]")[0] == (4.1337 - 4.1919) / -0.0199936
+
+
+def test_get_ocv_curve(Pulsing_fixture):
     """Test the get_ocv_curve method."""
-    procedure_fixture.set_SOC(
-        reference_charge=procedure_fixture.experiment("Break-in Cycles").charge(-1)
-    )
-    input_data = procedure_fixture.experiment("Discharge Pulses")
-    result = pulsing.get_ocv_curve(input_data)
+    result = pulsing.get_ocv_curve(Pulsing_fixture)
     expected_ocv_points = [
         4.1919,
         4.0949,
@@ -61,5 +67,5 @@ def test_get_ocv_curve(procedure_fixture):
         3.4513,
     ]
     assert isinstance(result, Result)
-    assert result.column_list == input_data.column_list
+    assert result.column_list == Pulsing_fixture.column_list
     assert np.allclose(result.get("Voltage [V]"), expected_ocv_points)
