+++ pyprobe/analysis/pulsing.py
@@ -3,6 +3,7 @@
 from typing import List
 
 import polars as pl
+from deprecated import deprecated
 from pydantic import BaseModel
 
 from pyprobe.analysis.utils import AnalysisValidator
@@ -39,6 +40,7 @@ class Pulsing(BaseModel):
         """
         return self.input_data.cycle(rest_number).rest(0)
 
+    @deprecated(reason="Use the get_resistances function instead.", version="1.1.0")
     def pulse_summary(self, r_times: List[float] = []) -> Result:
         """Returns a result object summarising the pulsing experiment.
 
@@ -258,3 +260,160 @@ def get_ocv_curve(input_data: PyProBEDataType) -> Result:
     return input_data.clean_copy(
         ocv_df, column_definitions=input_data.column_definitions
     )
+
+
+def get_resistances(input_data: PyProBEDataType, r_times: List[float] = []) -> Result:
+    """Returns a result object summarising the pulsing experiment.
+
+    Args:
+        r_times:
+            A list of times (in seconds) after each pulse at which to evaluate
+            the cell resistance.
+
+    Returns:
+        Result:
+            A result object containing key summary statistics for a pulsing
+            experiment. Includes:
+            - Experiment Capacity [Ah]
+            - SOC
+            - OCV [V]
+            - R0 [Ohms], calculated from the OCV and the first data point in the
+            pulse where the current is within 1% of the median pulse current
+            - Resistance calculated at each time provided in seconds in the r_times
+            argument
+    """
+    AnalysisValidator(
+        input_data=input_data,
+        required_columns=["Current [A]", "Voltage [V]", "Time [s]", "Event", "SOC"],
+    )
+    all_data_df = input_data.base_dataframe
+
+    # get the pulse number for each row
+    all_data_df = _get_pulse_number(all_data_df)
+    # get the last OCV point and timestamp before each pulse
+    ocv = (
+        all_data_df.filter(pl.col("Current [A]") == 0)
+        .group_by("Pulse Number")
+        .agg(
+            pl.col("Voltage [V]").last().alias("OCV [V]"),
+            pl.col("Time [s]").last().alias("Start Time [s]"),
+        )
+        .with_columns(pl.col("Pulse Number") + 1)
+    )
+    # get the median current for each pulse
+    pulse_current = (
+        all_data_df.filter(pl.col("Current [A]") != 0)
+        .group_by("Pulse Number")
+        .agg(pl.col("Current [A]").median().alias("Pulse Current"))
+    )
+    # recombine the dataframes
+    all_data_df = (
+        all_data_df.join(ocv, on="Pulse Number", how="left")
+        .join(pulse_current, on="Pulse Number", how="left")
+        .sort("Time [s]")
+    )
+    # get the first point in each pulse where the current is within 1% of the pulse
+    # current
+    pulse_df = (
+        all_data_df.filter(
+            (pl.col("Current [A]").abs() > 0.99 * pl.col("Pulse Current").abs())
+            & (pl.col("Current [A]").abs() < 1.01 * pl.col("Pulse Current").abs())
+        )
+        .group_by("Pulse Number")
+        .first()
+        .sort("Pulse Number")
+    )
+
+    # calculate the resistance at the start of the pulse
+    R0 = ((pl.col("Voltage [V]") - pl.col("OCV [V]")) / pl.col("Current [A]")).alias(
+        "R0 [Ohms]"
+    )
+    pulse_df = pulse_df.with_columns(R0)
+
+    t_col_names = [f"t_{time}s [s]" for time in r_times]
+    r_t_col_names = [f"R_{time}s [Ohms]" for time in r_times]
+    if t_col_names != []:
+        # add columns for the timestamps requested after each pulse
+        pulse_df = pulse_df.with_columns(
+            [
[…]
