### 2024-11-15 7c8a646
Move smoothing methods into separate functions and fix downsample logic
