+++ tests/analysis/test_smoothing.py
@@ -1,12 +1,10 @@
 """Tests for the smoothing analysis module."""
 import numpy as np
 import polars as pl
-import polars.testing as pl_testing
 import pytest
 from scipy import interpolate
 
 from pyprobe.analysis import smoothing
-from pyprobe.analysis.smoothing import Smoothing
 from pyprobe.result import Result
 
 
@@ -15,7 +13,7 @@ def noisy_data():
     """Generate noisy data."""
     np.random.seed(42)
     x = np.arange(1, 6, 0.01)
-    y = x**2 + np.random.normal(0, 0.1, size=x.size)  # y = x^2 with noise
+    y = x**2 + np.random.normal(0, 0.01, size=x.size)  # y = x^2 with noise
 
     return Result(
         base_dataframe=pl.LazyFrame({"x": x, "y": y}),
@@ -41,14 +39,15 @@ def noisy_data_reversed():
 
 def test_spline_smoothing(noisy_data, noisy_data_reversed, benchmark):
     """Test the spline smoothing method with noisy data."""
-    smoothing = Smoothing(input_data=noisy_data)
 
     def smooth():
-        return smoothing.spline_smoothing(x="x", target_column="y").get_only("y")
+        return smoothing.spline_smoothing(
+            input_data=noisy_data, x="x", target_column="y"
+        ).get_only("y")
 
     benchmark(smooth)
 
-    result = smoothing.spline_smoothing(x="x", target_column="y")
+    result = smoothing.spline_smoothing(input_data=noisy_data, x="x", target_column="y")
     x = np.arange(1, 6, 0.01)
     expected_y = x**2
 
@@ -63,54 +62,26 @@ def test_spline_smoothing(noisy_data, noisy_data_reversed, benchmark):
     np.testing.assert_allclose(result.get_only("d(y)/d(x)"), expected_dydx, rtol=0.2)
 
     # reverse the data
-    smoothing = Smoothing(input_data=noisy_data_reversed)
     flipped_x = np.flip(x)
-    result = smoothing.spline_smoothing(x="x", target_column="y")
+    result = smoothing.spline_smoothing(
+        input_data=noisy_data_reversed, x="x", target_column="y"
+    )
     flipped_expected_y = flipped_x**2
     np.testing.assert_allclose(result.get_only("y"), flipped_expected_y, rtol=0.2)
 
 
-def test_level_smoothing(noisy_data, noisy_data_reversed, benchmark):
-    """Test the level smoothing method with noisy data."""
-    smoothing = Smoothing(input_data=noisy_data)
-
-    result = smoothing.level_smoothing(target_column="y", interval=1, monotonic=True)
-
-    assert result.data.select(pl.col("y").diff().min().alias("min_diff")).item(0, 0) > 1
-    all_data = result.data.join(noisy_data.data, on="y")
-    assert len(all_data) == len(result.data)
-    pl_testing.assert_frame_equal(
-        all_data.select("x"), all_data.select(pl.col("x_right").alias("x"))
-    )
-    assert set(result.column_list) == set(noisy_data.column_list)
-
-    # reverse the data
-    smoothing = Smoothing(input_data=noisy_data_reversed)
-    result = smoothing.level_smoothing(target_column="y", interval=1)
-    assert (
-        result.data.select(pl.col("y").diff().abs().min().alias("min_diff")).item(0, 0)
-        > 1
-    )
-    all_data = result.data.join(noisy_data_reversed.data, on="y")
-    assert len(all_data) == len(result.data)
-    pl_testing.assert_frame_equal(
-        all_data.select("x"), all_data.select(pl.col("x_right").alias("x"))
-    )
-
-
 def test_savgol_smoothing(noisy_data, noisy_data_reversed, benchmark):
[…]
