+++ pyprobe/analysis/smoothing.py
@@ -13,6 +13,9 @@ from scipy.signal import savgol_filter
 
 from pyprobe.analysis.utils import AnalysisValidator
 from pyprobe.result import Result
+from pyprobe.typing import PyProBEDataType
+
+pl.Config.set_tbl_rows(50)
 
 
 class _LinearInterpolator(interpolate.PPoly):
@@ -130,7 +133,7 @@ def akima_interpolator(
     return _create_interpolator(interpolate.Akima1DInterpolator, x, y, **kwargs)
 
 
-def downsample_data(
+def _downsample_data(
     df: pl.DataFrame | pl.LazyFrame,
     target: str,
     sampling_interval: float,
@@ -138,6 +141,9 @@ def downsample_data(
     time_column: str = "Time [s]",
 ) -> pl.DataFrame | pl.LazyFrame:
     """Resample a DataFrame to a specified interval.
+
+    This method bins the data into intervals of the specified size and then selects
+    the desired occurrence within each bin. The target column must be monotonic.
 
     Args:
         df (pl.DataFrame | pl.LazyFrame):
@@ -153,47 +159,177 @@ def downsample_data(
         pl.DataFrame | pl.LazyFrame:
             The downsampled DataFrame.
     """
-    # Calculate the absolute difference between consecutive values
     df = df.with_columns(
         [
-            (pl.col(target) - pl.col(target).shift(1))
-            .abs()
-            .fill_null(0)
-            .alias("target_diff")
+            ((pl.col(target) / sampling_interval).floor() * sampling_interval).alias(
+                "bin"
+            )
         ]
     )
-
-    # Calculate cumulative sum of value differences
-    df = df.with_columns([pl.col("target_diff").cum_sum().alias("target_cumsum")])
-
-    # Create groups based on min_distance
-    df = df.with_columns(
-        [
-            (pl.col("target_cumsum") / sampling_interval)
-            .floor()
-            .cast(pl.Int64)
-            .alias("group")
-        ]
-    )
-
     # Group by 'group' and select the desired occurrence
     if occurrence == "first":
-        resampled_times = df.group_by("group").agg(
+        resampled_times = df.group_by("bin").agg(
             [pl.col(time_column).first().alias(time_column)]
         )
     elif occurrence == "last":
-        resampled_times = df.group_by("group").agg(
+        resampled_times = df.group_by("bin").agg(
             [
                 pl.col(time_column).last().alias(time_column),
             ]
         )
     elif occurrence == "middle":
         resampled_times = (
-            df.group_by("group")
+            df.group_by("bin")
             .agg(pl.col(time_column).quantile(0.5, "nearest").alias(time_column))
             .sort(time_column)
         )
     return resampled_times.join(df, on=time_column)
+
+
+def spline_smoothing(
+    input_data: PyProBEDataType,
+    target_column: str,
+    smoothing_lambda: Optional[float] = None,
+    x: str = "Time [s]",
+) -> Result:
+    """A method for smoothing noisy data using a spline.
+
+    Args:
+        input_data (PyProBEDataType):
+            The input data to smooth.
+        target_column (str):
+            The name of the target variable to smooth.
+        smoothing_lambda (float, optional):
+            The smoothing parameter. Default is None.
+        x (str, optional):
+            The name of the x variable for the spline curve fit.
+            Default is "Time [s]".
+
+    Returns:
+        Result:
+            A result object containing the data from input data with the target
+            column smoothed using a spline, and the gradient of the smoothed data
+            with respect to the x variable.
+    """
+    # validate and identify variables
+    validator = AnalysisValidator(
+        input_data=input_data, required_columns=[x, target_column]
+    )
+    x_data, y_data = validator.variables
+
[…]
