### 2024-11-15 cc76d69
Remove time column argument from downsampling
