### 2024-11-17 9471496
Move cycling methods out of class
