+++ tests/analysis/test_cycling.py
@@ -4,7 +4,7 @@ import math
 
 import pytest
 
-from pyprobe.analysis.cycling import Cycling
+from pyprobe.analysis import cycling
 from pyprobe.filters import Experiment
 from pyprobe.result import Result
 
@@ -18,26 +18,18 @@ def Cycling_fixture(lazyframe_fixture, info_fixture, step_descriptions_fixture):
         step_descriptions=step_descriptions_fixture,
         cycle_info=[],
     )
-    return Cycling(input_data=input_data)
-
-
-def test_init():
-    """Test the __init__ method."""
-    with pytest.raises(ValueError):
-        Cycling(input_data=Result(base_dataframe=None, info=None))
+    return input_data
 
 
 def test_set_capacity_throughput(Cycling_fixture):
     """Test the set_capacity_throughput method."""
-    Cycling_fixture._create_capacity_throughput()
-    assert "Capacity Throughput [Ah]" in Cycling_fixture.input_data.data.columns
-    assert Cycling_fixture.input_data.data["Capacity Throughput [Ah]"].head(1)[0] == 0
-    assert (
-        Cycling_fixture.input_data.step(0).data["Capacity Throughput [Ah]"].max()
-        == Cycling_fixture.input_data.step(0).capacity
-    )
+    result = cycling._create_capacity_throughput(
+        Cycling_fixture.base_dataframe
+    ).collect()
+    assert "Capacity Throughput [Ah]" in result.columns
+    assert result["Capacity Throughput [Ah]"].head(1)[0] == 0
     assert math.isclose(
-        Cycling_fixture.input_data.data["Capacity Throughput [Ah]"].tail(1)[0],
+        result["Capacity Throughput [Ah]"].tail(1)[0],
         0.472115,
         rel_tol=1e-5,
     )
@@ -45,9 +37,9 @@ def test_set_capacity_throughput(Cycling_fixture):
 
 def test_summary(BreakinCycles_fixture):
     """Test the summary property."""
-    cycling_instance = Cycling(input_data=BreakinCycles_fixture)
-    assert isinstance(cycling_instance.summary(), Result)
-    columns = cycling_instance.summary().data.columns
+    summary = cycling.summary(BreakinCycles_fixture)
+    assert isinstance(summary, Result)
+    columns = summary.data.columns
     required_columns = [
         "Capacity Throughput [Ah]",
         "Time [s]",
@@ -58,22 +50,22 @@ def test_summary(BreakinCycles_fixture):
         "Coulombic Efficiency",
     ]
     assert all(item in columns for item in required_columns)
-    assert cycling_instance.summary().data.shape[0] == 5
-    assert cycling_instance.summary().data["SOH Charge [%]"].head(1)[0] == 100
-    assert cycling_instance.summary().data["SOH Discharge [%]"].head(1)[0] == 100
+    assert summary.data.shape[0] == 5
+    assert summary.data["SOH Charge [%]"].head(1)[0] == 100
+    assert summary.data["SOH Discharge [%]"].head(1)[0] == 100
     assert math.isclose(
-        cycling_instance.summary().data["Charge Capacity [Ah]"].tail(1)[0],
+        summary.data["Charge Capacity [Ah]"].tail(1)[0],
         0.04139,
         rel_tol=1e-5,
     )
     assert math.isclose(
-        cycling_instance.summary().data["Discharge Capacity [Ah]"].tail(1)[0],
+        summary.data["Discharge Capacity [Ah]"].tail(1)[0],
         0.0413295,
         rel_tol=1e-5,
     )
 
     assert math.isclose(
-        cycling_instance.summary().data["Coulombic Efficiency"].tail(1)[0],
+        summary.data["Coulombic Efficiency"].tail(1)[0],
         0.999212,
         rel_tol=1e-7,
     )
