### 2024-11-17 db76030
Move differentiation methods out of class
