+++ pyprobe/analysis/differentiation.py
@@ -4,12 +4,150 @@ from typing import List
 
 import numpy as np
 import polars as pl
+from deprecated import deprecated
 from pydantic import BaseModel
 
 import pyprobe.analysis.base.differentiation_functions as diff_functions
 from pyprobe.analysis.utils import AnalysisValidator
 from pyprobe.result import Result
 from pyprobe.typing import PyProBEDataType
+
+
+def gradient(  # 1. Define the method
+    input_data: PyProBEDataType,
+    x: str,
+    y: str,
+) -> Result:
+    """Differentiate smooth data with a finite difference method.
+
+    A wrapper of the numpy.gradient function. This method calculates the gradient
+    of the data in the y column with respect to the data in the x column.
+
+    Args:
+        x (str): The name of the x variable.
+        y (str): The name of the y variable.
+
+    Returns:
+        Result:
+            A result object containing the columns, `x`, `y` and the
+            calculated gradient.
+    """
+    # 2. Validate the inputs to the method
+    validator = AnalysisValidator(
+        input_data=input_data,
+        required_columns=[x, y],
+        # required_type not neccessary here as type specified when declaring
+        # input_data attribute is strict enough
+    )
+    # 3. Retrieve the validated columns as numpy arrays
+    x_data, y_data = validator.variables
+
+    # 4. Perform the computation
+    gradient_title = f"d({y})/d({x})"
+    gradient_data = np.gradient(y_data, x_data)
+
+    # 5. Create a Result object to store the results
+    gradient_result = input_data.clean_copy(
+        pl.DataFrame({x: x_data, y: y_data, gradient_title: gradient_data})
+    )
+    # 6. Define the column definitions for the Result object
+    gradient_result.column_definitions = {
+        x: input_data.column_definitions[x],
+        y: input_data.column_definitions[y],
+        gradient_title: "The calculated gradient.",
+    }
+    # 7. Return the Result object
+    return gradient_result
+
+
+def differentiate_LEAN(
+    input_data: PyProBEDataType,
+    x: str,
+    y: str,
+    k: int = 1,
+    gradient: str = "dydx",
+    smoothing_filter: List[float] = [0.0668, 0.2417, 0.3830, 0.2417, 0.0668],
+    section: str = "longest",
+) -> Result:
+    r"""A method for differentiating noisy data.
+
+    Uses 'Level Evaluation ANalysis' (LEAN) method described in the paper of
+    :footcite:t:`Feng2020`.
+
+    This method assumes :math:`x` datapoints to be evenly spaced, it can return
+    either :math:`\frac{dy}{dx}` or :math:`\frac{dx}{dy}` depending on the argument
+    provided to the `gradient` parameter.
+
+    Args:
+        x (str):
+            The name of the x variable.
+        y (str):
+            The name of the y variable.
+        k (int, optional):
+            The integer multiple to apply to the sampling interval for the bin size
+            (:math:`\delta R` in paper). Default is 1.
+        gradient (str, optional):
+            The gradient to calculate, either 'dydx' or 'dxdy'. Default is 'dydx'.
+        smoothing_filter (List[float], optional):
+            The coefficients of the smoothing matrix.
+
+            Examples provided by :footcite:t:`Feng2020` include:
+
+                - [0.25, 0.5, 0.25] for a 3-point smoothing filter.
+                - [0.0668, 0.2417, 0.3830, 0.2417, 0.0668] (default) for a 5-point
+                    smoothing filter.
+                - [0.1059, 0.121, 0.1745, 0.1972, 0.1745, 0.121, 0.1059] for a
+                    7-point smoothing filter.
+
+        section (str, optional):
[…]
