+++ tests/analysis/test_differentiation.py
@@ -4,7 +4,7 @@ import numpy as np
 import polars as pl
 import pytest
 
-from pyprobe.analysis.differentiation import Differentiation
+from pyprobe.analysis import differentiation
 from pyprobe.result import Result
 
 x_data = np.array([1, 2, 3, 4, 5])
@@ -18,33 +18,21 @@ def differentiation_fixture():
         base_dataframe=pl.DataFrame({"x": x_data, "y": y_data}), info={}
     )
     input_data.column_definitions = {"x": "The x data", "y": "The y data"}
-    return Differentiation(input_data=input_data)
+    return input_data
 
 
-def test_differentiate_FD(differentiation_fixture):
+def test_gradient(differentiation_fixture):
     """Test the finite difference differentiation method."""
-    # Test case 1: dydx gradient
-    result = differentiation_fixture.differentiate_FD("x", "y", gradient="dydx")
+    result = differentiation.gradient(differentiation_fixture, "x", "y")
     expected_gradient = np.array([2, 2, 2, 2, 2])
     assert isinstance(result, Result)
     assert np.allclose(result.get_only("d(y)/d(x)"), expected_gradient)
-
-    # Test case 2: dxdy gradient
-    result = differentiation_fixture.differentiate_FD("x", "y", gradient="dxdy")
-    expected_gradient = np.array([0.5, 0.5, 0.5, 0.5, 0.5])
-    assert isinstance(result, Result)
-    assert np.allclose(result.get_only("d(x)/d(y)"), expected_gradient)
-
-    # Test case 3: Invalid gradient option
-    try:
-        result = differentiation_fixture.differentiate_FD("x", "y", gradient="invalid")
-        assert False  # Should raise a ValueError
-    except ValueError:
-        assert True
 
 
 def test_differentiate_LEAN(differentiation_fixture):
     """Test the LEAN differentiation method."""
     # Test case 1: dydx gradient
-    result = differentiation_fixture.differentiate_LEAN("x", "y", gradient="dydx")
+    result = differentiation.differentiate_LEAN(
+        differentiation_fixture, "x", "y", gradient="dydx"
+    )
     assert isinstance(result, Result)
