### 2024-11-17 1287c50
Add method for non-monotonic smoothing to downsample
