### 2024-11-17 b4453d1
Add method validation to smoothing module and deprecate old methods
