### 2024-11-18 37f1af2
Add deprecated package to requirements
