### 2024-11-18 566f911
Add validate_call decorators to analysis functions
