+++ tests/analysis/test_pulsing.py
@@ -46,8 +46,9 @@ def test_get_resistances(Pulsing_fixture):
     """Test the get_resistances method."""
     resistances = pulsing.get_resistances(Pulsing_fixture, [10])
     assert isinstance(resistances, Result)
+    print(resistances.data)
     assert resistances.get("R0 [Ohms]")[0] == (4.1558 - 4.1919) / -0.0199936
-    assert resistances.get("R_10s [Ohms]")[0] == (4.1337 - 4.1919) / -0.0199936
+    assert resistances.get("R_10.0s [Ohms]")[0] == (4.1337 - 4.1919) / -0.0199936
 
 
 def test_get_ocv_curve(Pulsing_fixture):
