### 2024-11-18 a0a27d0
Minor reorganisation and docstring updates
