+++ pyprobe/analysis/smoothing.py
@@ -17,6 +17,256 @@ from pyprobe.result import Result
 from pyprobe.typing import PyProBEDataType
 
 
+@validate_call
+def spline_smoothing(
+    input_data: PyProBEDataType,
+    target_column: str,
+    smoothing_lambda: Optional[float] = None,
+    x: str = "Time [s]",
+) -> Result:
+    """A method for smoothing noisy data using a spline.
+
+    Args:
+        input_data:
+            The input data to smooth.
+        target_column:
+            The name of the target variable to smooth.
+        smoothing_lambda:
+            The smoothing parameter. Default is None.
+        x:
+            The name of the x variable for the spline curve fit.
+            Default is "Time [s]".
+
+    Returns:
+        Result:
+            A result object containing the data from input data with the target
+            column smoothed using a spline, and the gradient of the smoothed data
+            with respect to the x variable.
+    """
+    # validate and identify variables
+    validator = AnalysisValidator(
+        input_data=input_data, required_columns=[x, target_column]
+    )
+    x_data, y_data = validator.variables
+
+    data_flipped = False
+    if x_data[0] > x_data[-1]:  # flip the data if it is not in ascending order
+        x_data = np.flip(x_data)
+        y_data = np.flip(y_data)
+        data_flipped = True
+
+    y_spline = make_smoothing_spline(x_data, y_data, lam=smoothing_lambda)
+
+    smoothed_y = y_spline(x_data)
+    if data_flipped:
+        x_data = np.flip(x_data)
+        smoothed_y = np.flip(smoothed_y)
+
+    derivative = y_spline.derivative()
+    smoothed_dydx = derivative(x_data)
+
+    result = copy.deepcopy(input_data)
+    smoothed_data_column = pl.Series(target_column, smoothed_y)
+    result.base_dataframe = result.base_dataframe.with_columns(
+        smoothed_data_column.alias(target_column)
+    )
+
+    gradient_column_name = f"d({target_column})/d({x})"
+    dydx_column = pl.Series(gradient_column_name, smoothed_dydx)
+    result.base_dataframe = result.base_dataframe.with_columns(
+        dydx_column.alias(gradient_column_name)
+    )
+    result.define_column(
+        f"d({target_column})/d({x})",
+        "The gradient of the smoothed data.",
+    )
+    return result
+
+
+def _downsample_monotonic_data(
+    df: pl.DataFrame | pl.LazyFrame,
+    target: str,
+    sampling_interval: float,
+    occurrence: Literal["first", "last", "middle"] = "first",
+) -> pl.DataFrame | pl.LazyFrame:
+    """Resample a DataFrame to a specified interval.
+
+    This method bins the data into intervals of the specified size and then selects
+    the desired occurrence within each bin. The target column must be monotonic.
+
+    Args:
+        df (pl.DataFrame | pl.LazyFrame):
+            The DataFrame to downsample.
+        target (str):
+            The target column to downsample.
+        sampling_interval (float):
+            The desired minimum interval between points.
+        occurrence (Literal['first', 'last', 'middle'], optional):
+            The occurrence to take when downsampling. Default is 'first'.
+
+    Returns:
+        pl.DataFrame | pl.LazyFrame:
+            The downsampled DataFrame.
+    """
+    df = df.with_columns(
+        [
+            ((pl.col(target) / sampling_interval).floor() * sampling_interval).alias(
+                "bin"
+            )
+        ]
+    )
+    # Group by 'group' and select the desired occurrence
+    if occurrence == "first":
+        return df.group_by("bin", maintain_order=True).first().drop("bin")
+    elif occurrence == "last":
[…]
