+++ pyprobe/analysis/pulsing.py
@@ -12,197 +12,6 @@ from pyprobe.result import Result
 from pyprobe.typing import PyProBEDataType
 
 
-class Pulsing(BaseModel):
-    """A pulsing experiment in a battery procedure."""
-
-    input_data: Experiment
-    """The input data for the pulsing experiment."""
-
-    def pulse(self, pulse_number: int) -> Step:
-        """Return a step object for a pulse in the pulsing experiment.
-
-        Args:
-            pulse_number (int): The Pulse Number to return.
-
-        Returns:
-            Step: A step object for a pulse in the pulsing experiment.
-        """
-        return self.input_data.cycle(pulse_number).chargeordischarge(0)
-
-    def pulse_rest(self, rest_number: int) -> Step:
-        """Return a step object for a rest in the pulsing experiment.
-
-        Args:
-            rest_number (int): The rest number to return.
-
-        Returns:
-            Step: A step object for a rest in the pulsing experiment.
-        """
-        return self.input_data.cycle(rest_number).rest(0)
-
-    @deprecated(reason="Use the get_resistances function instead.", version="1.1.0")
-    def pulse_summary(self, r_times: List[float] = []) -> Result:
-        """Returns a result object summarising the pulsing experiment.
-
-        Args:
-            r_times:
-                A list of times (in seconds) after each pulse at which to evaluate
-                the cell resistance.
-
-        Returns:
-            Result:
-                A result object containing key summary statistics for a pulsing
-                experiment. Includes:
-                - Experiment Capacity [Ah]
-                - SOC
-                - OCV [V]
-                - R0 [Ohms], calculated from the OCV and the first data point in the
-                pulse where the current is within 1% of the median pulse current
-                - Resistance calculated at each time provided in seconds in the r_times
-                argument
-        """
-        AnalysisValidator(
-            input_data=self.input_data,
-            required_columns=["Current [A]", "Voltage [V]", "Time [s]", "Event", "SOC"],
-        )
-        all_data_df = self.input_data.base_dataframe
-
-        # get the pulse number for each row
-        all_data_df = all_data_df.with_columns(
-            ((pl.col("Current [A]").shift() == 0) & (pl.col("Current [A]") != 0))
-            .cum_sum()
-            .alias("Pulse Number")
-        )
-        # get the last OCV point and timestamp before each pulse
-        ocv = (
-            all_data_df.filter(pl.col("Current [A]") == 0)
-            .group_by("Pulse Number")
-            .agg(
-                pl.col("Voltage [V]").last().alias("OCV [V]"),
-                pl.col("Time [s]").last().alias("Start Time [s]"),
-            )
-            .with_columns(pl.col("Pulse Number") + 1)
-        )
-        # get the median current for each pulse
-        pulse_current = (
-            all_data_df.filter(pl.col("Current [A]") != 0)
-            .group_by("Pulse Number")
-            .agg(pl.col("Current [A]").median().alias("Pulse Current"))
-        )
-        # recombine the dataframes
-        all_data_df = (
-            all_data_df.join(ocv, on="Pulse Number", how="left")
-            .join(pulse_current, on="Pulse Number", how="left")
-            .sort("Time [s]")
-        )
-        # get the first point in each pulse where the current is within 1% of the pulse
-        # current
-        pulse_df = (
-            all_data_df.filter(
-                (pl.col("Current [A]").abs() > 0.99 * pl.col("Pulse Current").abs())
-                & (pl.col("Current [A]").abs() < 1.01 * pl.col("Pulse Current").abs())
-            )
-            .group_by("Pulse Number")
-            .first()
-            .sort("Pulse Number")
-        )
-
-        # calculate the resistance at the start of the pulse
-        R0 = (
[…]
