### 2024-11-18 6e73b84
Update docs and examples for refactored analysis module
