### 2024-11-18 6d9c501
Fix error with r_time provided as int to get_resistances method
