+++ _sources/examples/differentiating-voltage-data.ipynb.txt
@@ -74,10 +74,9 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from pyprobe.analysis.differentiation import Differentiation\n",
+    "from pyprobe.analysis import differentiation\n",
     "\n",
-    "discharge_differentiation = Differentiation(input_data = final_cycle.discharge(0))\n",
-    "raw_data_dVdQ = discharge_differentiation.differentiate_FD(\"Capacity [Ah]\", \"Voltage [V]\", gradient='dydx')\n",
+    "raw_data_dVdQ = differentiation.gradient(final_cycle.discharge(0),\"Capacity [Ah]\", \"Voltage [V]\")\n",
     "print(raw_data_dVdQ.column_list)\n",
     "fig = pyprobe.Plot()\n",
     "fig.add_line(raw_data_dVdQ, 'Voltage [V]', 'd(Voltage [V])/d(Capacity [Ah])')\n",
@@ -99,17 +98,16 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from pyprobe.analysis.smoothing import Smoothing\n",
+    "from pyprobe.analysis import smoothing\n",
     "\n",
-    "smoother = Smoothing(input_data = final_cycle.discharge(0))\n",
     "\n",
-    "level_smooothed_data = smoother.level_smoothing(target_column='Voltage [V]',\n",
-    "                                                  interval=0.002,\n",
-    "                                                  monotonic=True,\n",
+    "downsampled_data = smoothing.downsample(input_data = final_cycle.discharge(0),\n",
+    "                                            target_column='Voltage [V]',\n",
+    "                                                  sampling_interval=0.002,\n",
     "                                                )\n",
     "fig = pyprobe.Plot()\n",
-    "fig.add_line(final_cycle.discharge(0), 'Capacity [Ah]', 'Voltage [V]', label='Raw data', color='red',)\n",
-    "fig.add_line(level_smooothed_data, 'Capacity [Ah]', 'Voltage [V]', label='Level smoothed data', color='blue', dash='dash')\n",
+    "# fig.add_line(final_cycle.discharge(0), 'Capacity [Ah]', 'Voltage [V]', label='Raw data', color='red',)\n",
+    "fig.add_line(downsampled_data, 'Capacity [Ah]', 'Voltage [V]', label='Level smoothed data', color='blue', dash='dash')\n",
     "fig.show_image()"
    ]
   },
@@ -126,11 +124,10 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "differentiation_object = Differentiation(input_data = level_smooothed_data)\n",
-    "level_smoothed_data_dVdQ = differentiation_object.differentiate_FD(\"Capacity [Ah]\", \"Voltage [V]\", gradient='dxdy')\n",
+    "downsampled_data_dVdQ = differentiation.gradient(downsampled_data, \"Voltage [V]\", \"Capacity [Ah]\")\n",
     "\n",
     "fig = pyprobe.Plot()\n",
-    "fig.add_line(level_smoothed_data_dVdQ, 'Voltage [V]', 'd(Capacity [Ah])/d(Voltage [V])')\n",
+    "fig.add_line(downsampled_data_dVdQ, 'Voltage [V]', 'd(Capacity [Ah])/d(Voltage [V])')\n",
     "fig.show_image()"
    ]
   },
@@ -147,19 +144,17 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "smoother = Smoothing(input_data = final_cycle.discharge(0))\n",
-    "\n",
-    "spline_smoothed_data = smoother.spline_smoothing(x='Capacity [Ah]',\n",
[…]
