### 2024-11-18 42933e0
Add test to check that missing columns are filled with null values
