### 2024-11-18 7b65e17
Add functionality to specify header rows in experiment log
