### 2024-11-18 644f187
Add functionality to specify header rows in experiment log
