### 2024-11-18 ce17fc5
Allow datetime objects in info dict
