### 2024-11-18 0490860
Allow datetime objects in info dict
