### 2024-11-18 1cd820d
Add method to get SOH values across RPTs
