+++ pyprobe/analysis/RPT.py
@@ -67,3 +67,33 @@ class RPT(BaseModel):
             name: "The cell capacity.",
         }
         return capacities_result
+
+    def process_soh(self, filter: str, name: str = "SOH") -> Result:
+        """Calculate the SOH for a particular experiment step across the RPTs.
+
+        Args:
+            filter (str): The filter to apply to the data.
+            name (str): The name of the column to store the SOH.
+
+        Returns:
+            Result: A result object for the cell SOH in each RPT.
+        """
+        all_soh = np.zeros(len(self.input_data))
+        for rpt_number, experiment in enumerate(self.input_data):
+            AnalysisValidator(input_data=experiment, required_columns=["Capacity [Ah]"])
+            reference_step = eval(f"experiment.{filter}")
+            all_soh[rpt_number] = reference_step.capacity
+        all_soh = all_soh / all_soh[0]
+
+        soh_df = pl.DataFrame(
+            {
+                name: all_soh,
+            }
+        )
+        soh_result = self.input_data[0].clean_copy(self._base_summary_df.hstack(soh_df))
+        self._summary_df = self._summary_df.hstack(soh_df)
+        soh_result.column_definitions = {
+            "RPT Number": "The RPT number.",
+            name: "The cell SOH.",
+        }
+        return soh_result
