### 2024-11-25 a4c7c22
Add flexible method for converting neware hh:mm:ss time format to floats
