### 2024-11-29 a4f0a99
Bump altair from 5.4.0 to 5.5.0
Bumps [altair](https://github.com/vega/altair) from 5.4.0 to 5.5.0.
- [Release notes](https://github.com/vega/altair/releases)
- [Commits](https://github.com/vega/altair/compare/v5.4.0...v5.5.0)

---
updated-dependencies:
- dependency-name: altair
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
