+++ requirements.txt
@@ -6,7 +6,7 @@
 #
 aiosignal==1.3.1
     # via ray
-altair==5.4.0
+altair==5.5.0
     # via streamlit
 annotated-types==0.7.0
     # via pydantic
