### 2024-11-29 3b781e5
Skip the parquet conversion if a parquet file already exists
