### 2024-11-29 c38e2a3
Create common function for converting data to parquet
