+++ pyprobe/cell.py
@@ -4,9 +4,8 @@ import os
 import shutil
 import time
 import warnings
-from typing import Callable, Dict, List, Literal, Optional
 import zipfile
-from typing import Any, Callable, Dict, List, Optional
+from typing import Any, Callable, Dict, List, Literal, Optional
 
 import distinctipy
 import polars as pl
@@ -62,10 +61,11 @@ class Cell(BaseModel):
         values = info
         return values
 
-    @validate_call
-    def process_cycler_file(
+    def _convert_to_parquet(
         self,
-        cycler: str,
+        cycler: Literal[
+            "neware", "biologic", "biologic_MB", "arbin", "basytec", "maccor", "generic"
+        ],
         folder_path: str,
         input_filename: str | Callable[[str], str],
         output_filename: str | Callable[[str], str],
@@ -74,36 +74,40 @@ class Cell(BaseModel):
             "performance", "file size", "uncompressed"
         ] = "performance",
         overwrite_existing: bool = False,
+        column_dict: Optional[Dict[str, str]] = None,
     ) -> None:
-        """Convert cycler file into PyProBE format.
+        """Convert a file into PyProBE format.
 
         Args:
-            cycler (str):
-                The cycler used to produce the data. Available cyclers are:
-                - 'neware'
-                - 'biologic'
-                - 'biologic_MB' (for Modulo Bat Biologic data)
-            folder_path (str):
+            cycler:
+                The cycler used to produce the data.
+            folder_path:
                 The path to the folder containing the data file.
-            input_filename (str | function):
+            input_filename:
                 A filename string or a function to generate the file name for cycler
                 data.
-            output_filename (str | function):
+            output_filename:
                 A filename string or a function to generate the file name for PyProBE
                 data.
-            filename_inputs (list):
+            filename_inputs:
                 The list of inputs to input_filename and output_filename, if they are
                 functions. These must be keys of the cell info.
-            compression_priority (str):
+            compression_priority:
                 The priority of the compression algorithm to use on the resulting
                 parquet file. Available options are:
                 - 'performance': Use the 'lz4' compression algorithm (default).
                 - 'file size': Use the 'zstd' compression algorithm.
                 - 'uncompressed': Do not use compression.
-            overwrite_existing (bool):
+            overwrite_existing:
                 If True, any existing parquet file with the output_filename will be
                 overwritten. If False, the function will skip the conversion if the
                 parquet file already exists.
+            column_dict:
+                A dictionary mapping the column names in the cycler file to the PyProBE
+                column names. The keys of the dictionary are the cycler column names and
+                the values are the PyProBE column names. You must use asterisks to
+                indicate the units of the columns. Only used for the 'generic' cycler.
+                E.g. :code:`{"V (*)": "Voltage [*]"}`.
         """
         input_data_path = self._get_data_paths(
             folder_path, input_filename, filename_inputs
@@ -120,9 +124,22 @@ class Cell(BaseModel):
                 "neware": neware.Neware,
                 "biologic": biologic.Biologic,
                 "biologic_MB": biologic.BiologicMB,
+                "arbin": arbin.Arbin,
+                "basytec": basytec.Basytec,
+                "maccor": maccor.Maccor,
             }
             t1 = time.time()
-            importer = cycler_dict[cycler](input_data_path=input_data_path)
+            if cycler == "generic":
+                if column_dict is None:
+                    raise ValueError(
+                        "column_dict must be provided for the generic cycler."
+                    )
+                importer = basecycler.BaseCycler(
+                    input_data_path=input_data_path,
+                    column_dict=column_dict,
+                )
+            else:
+                importer = cycler_dict[cycler](input_data_path=input_data_path)
             compression_dict = {
[…]
