### 2024-11-29 0bd25e2
Update version number to 1.1.0
