### 2024-12-03 37b4981
Fix search for experiment capacity on pulsing module methods
