### 2024-12-07 4dcd17a
Add tests for the RPT module
