+++ pyprobe/analysis/RPT.py
@@ -1,7 +1,5 @@
-"""A module for processing separate RPTs."""
+"""A module to streamline processing of identical reference performance tests."""
 
-
-import copy
 from typing import Any, List
 
 import numpy as np
@@ -26,7 +24,10 @@ class RPT(BaseModel):
                 "RPT Number": list(range(len(self.input_data))),
             }
         )
-        self._summary_df = copy.deepcopy(self._base_summary_df)
+        self._rpt_summary = self.input_data[0].clean_copy(self._base_summary_df)
+        self._rpt_summary.column_definitions = {
+            "RPT Number": "The RPT number.",
+        }
 
     @property
     def rpt_summary(self) -> Result:
@@ -35,17 +36,16 @@ class RPT(BaseModel):
         Returns:
             Result: A result object for the RPT summary.
         """
-        return self.input_data[0].clean_copy(self._summary_df)
+        return self._rpt_summary
 
-    def process_cell_capacity(self, filter: str, name: str = "Capacity [Ah]") -> Result:
+    def process_cell_capacity(self, filter: str, name: str = "Capacity [Ah]") -> None:
         """Calculate the capacity for a particular experiment step across the RPTs.
+
+        Results are stored in the :property:`rpt_summary` attribute.
 
         Args:
             filter (str): The filter to apply to the data.
             name (str): The name of the column to store the capacity.
-
-        Returns:
-            Result: A result object for the cell capacity in each RPT.
         """
         all_capacities = np.zeros(len(self.input_data))
         for rpt_number, experiment in enumerate(self.input_data):
@@ -58,25 +58,20 @@ class RPT(BaseModel):
                 name: all_capacities,
             }
         )
-        capacities_result = self.input_data[0].clean_copy(
-            self._base_summary_df.hstack(capacity_df)
+        self._rpt_summary.base_dataframe = self._rpt_summary.base_dataframe.hstack(
+            capacity_df
         )
-        self._summary_df = self._summary_df.hstack(capacity_df)
-        capacities_result.column_definitions = {
-            "RPT Number": "The RPT number.",
-            name: "The cell capacity.",
-        }
-        return capacities_result
+        column_definition = {name: "The cell capacity."}
+        self._rpt_summary.column_definitions.update(column_definition)
 
-    def process_soh(self, filter: str, name: str = "SOH") -> Result:
+    def process_soh(self, filter: str, name: str = "SOH") -> None:
         """Calculate the SOH for a particular experiment step across the RPTs.
+
+        Results are stored in the :property:`rpt_summary` attribute.
 
         Args:
             filter (str): The filter to apply to the data.
             name (str): The name of the column to store the SOH.
-
-        Returns:
-            Result: A result object for the cell SOH in each RPT.
         """
         all_soh = np.zeros(len(self.input_data))
         for rpt_number, experiment in enumerate(self.input_data):
@@ -90,10 +85,8 @@ class RPT(BaseModel):
                 name: all_soh,
             }
         )
-        soh_result = self.input_data[0].clean_copy(self._base_summary_df.hstack(soh_df))
-        self._summary_df = self._summary_df.hstack(soh_df)
-        soh_result.column_definitions = {
-            "RPT Number": "The RPT number.",
-            name: "The cell SOH.",
-        }
-        return soh_result
+        self._rpt_summary.base_dataframe = self._rpt_summary.base_dataframe.hstack(
+            soh_df
+        )
+        column_definition = {name: "The cell SOH."}
+        self._rpt_summary.column_definitions.update(column_definition)
