### 2024-12-07 7c9b4e0
Remove SOC column requirement for pulsing get_resitances method
