### 2024-12-07 4b43c6a
Add method for extracting pulse resistances from RPTs
