### 2024-12-07 ce14d5c
Add column for the start date of each RPT
