### 2024-12-07 98d6f8c
Rename typing module to avoid clash with built-in python module
