### 2024-12-07 5b9eb18
feat: create seaborn wrapper
