### 2024-12-07 651a723
fix:Turn off estimator for seaborn lineplot by default for performance and create example
