### 2024-12-07 1b22b02
feat:Add function to retrieve relevant columns for plotting only
