### 2024-12-07 af523ef
feat:Add built-in unit conversion to column retrieval for plotting
