### 2024-12-07 b842f71
Add built-in unit conversion to column retrieval for plotting
