### 2024-12-09 ImperialCollegeLondon/PyProBE#176 @jakobhaervig
Installation fails
I'm following instructions found here https://imperialcollegelondon.github.io/PyProBE/user_guide/installation.html
All seems to install just fine but when I try to load the module with

`import pyprobe`

I get:
`[WARNING] citations.read_citations(76): Citations could not be read because the 'pybtex' library is not installed. Install 'pybamm[cite]' to enable citation reading.`

Also, my I suggest that all dependencies are kept in pyproject.toml only. The `requirements.txt` seems redundant to me?
