### 2024-12-09 ImperialCollegeLondon/PyProBE#177 @jakobhaervig
Can't run first example
Trying to run the example from https://imperialcollegelondon.github.io/PyProBE/examples/getting-started.html
There is a reference to some data `data_directory = '../../../tests/sample_data/neware'`, which obviously fails because I don't have access to the data.
