### 2024-12-09 #7474 @jakobhaervig
Added two initial issues. Will you check them out @tomjholland?
https://github.com/ImperialCollegeLondon/PyProBE/issues/176
https://github.com/ImperialCollegeLondon/PyProBE/issues/177
