### 2024-12-09 ImperialCollegeLondon/PyProBE#176 @tomjholland
Thanks @jakobhaervig, that user warning is from PyBaMM (a dependency). It causes no runtime issues but I will try to supress it.
