### 2024-12-09 #7474 @tomjholland
Hi @jakobhaervig, @ritesh001,

Regarding https://github.com/ImperialCollegeLondon/PyProBE/issues/177: the sample data is stored in the repo: https://github.com/tomjholland/PyProBE/tree/main/tests/sample_data.

The relative import in the example means that if you run the examples from the cloned repository the sample data files should load correctly - this is how the CI tests run and how the example notebooks are embedded in the documentation with github actions.
