### 2024-12-09 ImperialCollegeLondon/PyProBE#176 @tomjholland
Regarding `requirements.txt`, this is auto-generated from the `pyproject.toml` using pip-tools which ensures that the dependencies are compatible. This way they are not independent of each other. Having it in the repo removes the  `pip-compile` step in the installation.
