### 2024-12-09 1d7ae55
refactor: Add cache for collected columns of the base dataframe
