### 2024-12-10 77e3b36
refactor:Deprecate the get_only() function
