+++ pyprobe/analysis/utils.py
@@ -22,7 +22,7 @@ def assemble_array(input_data: List[Result], name: str) -> NDArray[Any]:
     Returns:
         NDArray: The assembled array.
     """
-    return np.vstack([input.get_only(name) for input in input_data])
+    return np.vstack([input.get(name) for input in input_data])
 
 
 class AnalysisValidator(BaseModel):
