+++ tests/test_rawdata.py
@@ -1,6 +1,7 @@
 """Tests for the RawData class."""
 
 import copy
+import random
 
 import numpy as np
 import polars as pl
@@ -31,6 +32,21 @@ def test_init(RawData_fixture, step_descriptions_fixture):
     with pytest.raises(ValueError):
         RawData(base_dataframe=data, info={"test": 1})
 
+
+def test_data(RawData_fixture):
+    """Test the data property."""
+    columns = copy.deepcopy(RawData_fixture.data.collect_schema().names())
+    random.shuffle(columns)
+    RawData_fixture.live_dataframe = RawData_fixture.live_dataframe.select(columns)
+    assert RawData_fixture.data.columns == [
+        "Time [s]",
+        "Step",
+        "Event",
+        "Current [A]",
+        "Voltage [V]",
+        "Capacity [Ah]",
+        "Date",
+    ]
 
 def test_capacity(BreakinCycles_fixture):
     """Test the capacity property."""
