### 2024-12-10 f0c1f10
feat: Add utility methods for plot and hvplot to result objects
