+++ pyprobe/result.py
@@ -1,13 +1,17 @@
 """A module for the Result class."""
+from functools import wraps
 from pprint import pprint
 from typing import Any, Dict, List, Literal, Optional, Set, Tuple, Union
 
+import hvplot.polars
 import numpy as np
+import pandas as pd
 import polars as pl
 from deprecated import deprecated
 from numpy.typing import NDArray
 from pydantic import BaseModel, Field, model_validator
 
+from pyprobe.plot import _retrieve_relevant_columns
 from pyprobe.units import split_quantity_unit, unit_from_regexp
 
 logger = logging.getLogger(__name__)
@@ -209,6 +213,31 @@ class Result(BaseModel):
         if complete_dataframe.is_empty():
             raise ValueError("No data exists for this filter.")
         return complete_dataframe
+
+    @wraps(pd.DataFrame.plot)
+    def plot(self, *args: Any, **kwargs: Any) -> None:
+        """Wrapper for plotting using the pandas library."""
+        data_to_plot = _retrieve_relevant_columns(self, args, kwargs)
+        return data_to_plot.to_pandas().plot(*args, **kwargs)
+
+    plot.__doc__ = (
+        "This is a wrapper around the pandas plot method. It will perform"
+        "exactly as you would expect the pandas plot method to perform"
+        "when called on a DataFrame.\n\n" + (plot.__doc__ or "")
+    )
+
+    @wraps(hvplot.hvPlot)
+    def hvplot(self, *args: Any, **kwargs: Any) -> None:
+        """Wrapper for plotting using the hvplot library."""
+        data_to_plot = _retrieve_relevant_columns(self, args, kwargs)
+        return data_to_plot.hvplot(*args, **kwargs)
+
+    hvplot.__doc__ = (
+        "HvPlot is a library for creating fast and interactive plots.\n\n"
+        "The default backend is bokeh, which can be changed by setting the backend "
+        "with :code:`hvplot.extension('matplotlib')` or "
+        ":code:`hvplot.extension('plotly')`.\n\n" + (hvplot.__doc__ or "")
+    )
 
     def _get_data_subset(self, *column_names: str) -> pl.DataFrame:
         """Return a subset of the data with the specified columns.
