### 2024-12-11 2866a91
feat: Add user control of the header row index and date format when importing a generic file
