+++ pyprobe/cyclers/basytec.py
@@ -23,11 +23,14 @@ class Basytec(BaseCycler):
     }
 
     @staticmethod
-    def read_file(filepath: str) -> pl.DataFrame | pl.LazyFrame:
+    def read_file(
+        filepath: str, header_row_index: int = 0
+    ) -> pl.DataFrame | pl.LazyFrame:
         """Read a battery cycler file into a DataFrame.
 
         Args:
             filepath: The path to the file.
+            header_row_index: The index of the header row.
 
         Returns:
             pl.DataFrame | pl.LazyFrame: The DataFrame.
