+++ pyprobe/cell.py
@@ -77,6 +77,8 @@ class Cell(BaseModel):
         ] = "performance",
         overwrite_existing: bool = False,
         column_dict: Optional[Dict[str, str]] = None,
+        header_row_index: int = 0,
+        date_column_format: Optional[str] = None,
     ) -> None:
         """Convert a file into PyProBE format.
 
@@ -110,6 +112,12 @@ class Cell(BaseModel):
                 the values are the PyProBE column names. You must use asterisks to
                 indicate the units of the columns. Only used for the 'generic' cycler.
                 E.g. :code:`{"V (*)": "Voltage [*]"}`.
+            header_row_index:
+                The index of the header row in the file. Only used for the 'generic'
+                cycler.
+            date_column_format:
+                The format of the date column in the generic file. Only used for the
+                'generic' cycler.
         """
         input_data_path = self._get_data_paths(
             folder_path, input_filename, filename_inputs
@@ -141,6 +149,8 @@ class Cell(BaseModel):
                 importer = basecycler.BaseCycler(
                     input_data_path=input_data_path,
                     column_dict=column_dict,
+                    header_row_index=header_row_index,
+                    datetime_format=date_column_format,
                 )
             else:
                 importer = cycler_dict[cycler](input_data_path=input_data_path)
@@ -217,6 +227,8 @@ class Cell(BaseModel):
         input_filename: str | Callable[[str], str],
         output_filename: str | Callable[[str], str],
         column_dict: Dict[str, str],
+        header_row_index: int = 0,
+        date_column_format: Optional[str] = None,
         filename_inputs: Optional[List[str]] = None,
         compression_priority: Literal[
             "performance", "file size", "uncompressed"
@@ -239,6 +251,10 @@ class Cell(BaseModel):
                 the values are the PyProBE column names. You must use asterisks to
                 indicate the units of the columns.
                 E.g. :code:`{"V (*)": "Voltage [*]"}`.
+            header_row_index (int, optional):
+                The index of the header row in the file. Defaults to 0.
+            date_column_format (str, optional):
+                The format of the date column in the generic file. Defaults to None.
             filename_inputs (list):
                 The list of inputs to input_filename and output_filename.
                 These must be keys of the cell info.
@@ -257,6 +273,8 @@ class Cell(BaseModel):
             filename_inputs,
             compression_priority,
             column_dict=column_dict,
+            header_row_index=header_row_index,
+            date_column_format=date_column_format,
         )
 
     @validate_call
