### 2024-12-11 70743b7
feat: Add full-word Seconds unit
