### 2024-12-11 4b9e80d
feat: Add a method for adding a procedure without a README file
