### 2024-12-12 2b1c62d
feat: Add buffer to charge and discharge filters to exclude noise around zero current
