### 2024-12-17 ImperialCollegeLondon/PyProBE#176 @jakobhaervig
Step 3 and the requirements.txt is redundant. You may have to specify versions in `pyproject.toml` directly. You can basically leave it out (step 3) and simply do `pip install .`, which will read dependencies from `pyproject.toml`. Only having dependencies listed one place is cleaner in my opinion.
