### 2024-12-20 0a3f82d
feat: Add capability to read biologic mpt files that have no header
