### 2024-12-20 98d69b3
Minor update to figure 5 description of analysis methods
