### 2024-12-20 #7474 @tomjholland
@TomTranter, @jakobhaervig, @ritesh001 thanks a lot for your time reviewing my work and your nice comments.

@phibeck I caught a typo on Figure 5 in the manuscript, so took the opportunity to correct a very minor UI change that I have made in the package since submission in the same figure. The changes are made in this PR: https://github.com/ImperialCollegeLondon/PyProBE/pull/179, if there are no issues I will merge it with your permission. I can then create a final repo release, which I assume the version on the JOSS submission can be updated to reflect?

Thanks!
