### 2024-12-22 744f496
feat: Add method to combine multiple result objects
This will integrate their info dicts into the dataframe
