### 2024-12-23 cb09608
Add a logger module and implement in cell
