+++ pyprobe/__init__.py
@@ -1,5 +1,8 @@
 """The PyProBE package."""
 from .cell import Cell, __version__, load_archive, make_cell_list  # noqa: F401
 from .dashboard import launch_dashboard  # noqa: F401
+from .logger import configure_logging  # noqa: F401
 from .plot import Plot  # noqa: F401
 from .result import Result  # noqa: F401
+
+configure_logging()
