### 2024-12-23 ef95577
Add logger to filters module
