### 2024-12-23 8b4a12e
Add logger to rawdata module
