### 2024-12-23 ce37807
Add logger to readme_processor
