### 2024-12-23 575bae3
Add logging to units module
