### 2024-12-23 6cbf898
Add logging to neware module
