### 2024-12-23 8dc6b0f
Add logger to dma module
