### 2024-12-23 4bcdc1c
Overwrite logging handlers instead of appending
