### 2024-12-23 4ae0136
Replace print statements with logger
