### 2024-12-30 5bdb071
Make pybamm import only during function call requiring it
