### 2024-12-31 fc8102e
Pybamm_experiment property of rawdata returns input to Experiment class
Removes pybamm import
