### 2024-12-31 de8157e
Replace pybamm import in cell with a protocol class
