### 2024-12-31 5c2dbdb
Update ci workflow for uv
