+++ .github/workflows/ci.yml
@@ -10,9 +10,9 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - uses: actions/setup-python@v4
+      - uses: actions/setup-python@v5
         with:
-          python-version: "3.11"
+          python-version-file: "pyproject.toml"
       - uses: pre-commit/action@v3.0.0
   test:
     needs: qa
@@ -20,10 +20,20 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v4
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
         with:
-          python-version: "3.11"
-      - run: pip install -r requirements-dev.txt
-      - run: pytest tests --benchmark-json output.json
+          python-version-file: "pyproject.toml"
+      
+      - name: Install the project
+        run: uv sync --all-extras
+
+      - name: Run tests
+        run: uv run pytest tests --benchmark-json output.json
+
       - name: Store benchmark result
         uses: benchmark-action/github-action-benchmark@v1
         with:
@@ -37,6 +47,5 @@ jobs:
           comment-on-alert: true
           # Enable Job Summary for PRs
           summary-always: true
-      - run: pip install nbmake
-      - run: pip install .
-      - run: pytest --nbmake docs/source/examples/*ipynb 
+      - run: uv pip install nbmake
+      - run: uv run pytest --nbmake docs/source/examples/*ipynb 
