### 2024-12-31 a378c00
Update documentation workflow for uv
