+++ .github/workflows/ci.yml
@@ -29,7 +29,7 @@ jobs:
           python-version-file: "pyproject.toml"
       
       - name: Install the project
-        run: uv sync --all-extras
+        run: uv sync --all-extras --group format --group test --group docs --group external_integration
 
       - name: Run tests
         run: uv run pytest tests --benchmark-json output.json
