+++ .github/workflows/ci.yml
@@ -29,7 +29,7 @@ jobs:
           python-version-file: "pyproject.toml"
       
       - name: Install the project
-        run: uv sync --all-extras --group format --group test --group docs --group external_integration
+        run: uv sync --all-extras --group format --group test --group docs --group external-integration
 
       - name: Run tests
         run: uv run pytest tests --benchmark-json output.json
@@ -47,5 +47,7 @@ jobs:
           comment-on-alert: true
           # Enable Job Summary for PRs
           summary-always: true
-      - run: uv pip install nbmake
-      - run: uv run pytest --nbmake docs/source/examples/*ipynb 
+      - name: Install nbmake
+        run: uv pip install nbmake
+      - name: Test notebooks
+        run: uv run pytest --nbmake docs/source/examples/*ipynb
