+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "PyProBE"
+name = "PyProBE-Data"
 version = "1.1.0"
 requires-python = ">=3.11,<3.13"
 dependencies = [
