### 2024-12-31 0e66cb8
Update author, licence and build directories in pyproject
