+++ pyproject.toml
@@ -1,7 +1,11 @@
 [project]
 name = "PyProBE-Data"
+authors = [
+  {name = "Thomas Holland"},
+]
 version = "1.1.0"
 requires-python = ">=3.11,<3.13"
+license = { file = "LICENSE" }
 dependencies = [
     "deprecated>=1.2.15",
     "distinctipy>=1.3.4",
@@ -60,7 +64,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
 exclude = [
-    "htmlcov",
+    "htmlcov", "docs*", "tests*", "paper*"
 ] # Exclude the coverage report file from setuptools package finder
 
 [tool.mypy]
