### 2024-12-31 402d2d5
Add workflow for publishing to pypi
