### 2024-12-31 82840de
Correct pypi urls
