### 2024-12-31 c922d32
Update user installation docs for pypi install
