+++ docs/source/user_guide/installation.rst
@@ -30,56 +30,9 @@ virtual environment to install PyProBE, for example venv or conda.
 
    .. image:: images/VSCode_select_kernel.png
 
-
-You should have two separate directories: 
-
-* one for the PyProBE repository 
-* one for your processing script
-
 The steps to install PyProBE are as follows:
 
-1. Enter a directory in which you wish to install PyProBE:
-   
-   .. code-block:: bash
-
-      cd /path/to/installation/directory
-
-2. Clone the repository to your local machine. This creates a directory called PyProBE.
-
-   .. code-block:: bash
-
-      git clone https://github.com/ImperialCollegeLondon/PyProBE.git
-      cd PyProBE
-
-   It is recommended to used the tagged releases for a stable version of the code. 
-   You can list the available tags with the following command:
-
-   .. code-block:: bash
-
-      git tag
-
-   You can then checkout the desired tag with the following command:
-   
-   .. code-block:: bash
-
-      git checkout <tag name>
-
-   For example:
-
-   .. code-block:: bash
-
-      git checkout v0.1.0
-
-
-   To update your installation you can run:
-   
-   .. code-block:: bash
-
-      git fetch --tags
-      git tag
-      git checkout <tag name>
-
-3. Create and activate a virtual environment.
+1. Create and activate a virtual environment.
   
   .. tabs::
       .. tab:: venv
@@ -100,27 +53,20 @@ The steps to install PyProBE are as follows:
             conda create -n pyprobe python=3.12
             conda activate pyprobe
 
-3. Install PyProBE's dependencies:
+2. Install PyProBE with pip:
    
    .. code-block:: bash
 
-      cd /path/to/your/directory/PyProBE
-      pip install -r requirements.txt
+      pip install PyProBE-Data
 
-4. Install PyProBE as a package into your virtual environment:
-   
-   .. code-block:: bash
-
-      pip install .
-
-5. In a second directory you can create a new python script or jupyter notebook to 
+3. You can create a new python script or jupyter notebook to 
    process your data. You can import PyProBE into your script as follows:
 
    .. code-block:: python
 
       import pyprobe
 
-6. Before being able to launch the dashboard you will need to initialise streamlit.
+4. Before being able to launch the dashboard you will need to initialise streamlit.
    Do this by running the streamlit Hello app from your command line:
 
    .. code-block:: bash
