+++ requirements.txt
@@ -64,7 +64,7 @@ ipython==8.26.0
     # via PyProBE (pyproject.toml)
 jedi==0.19.1
     # via ipython
-jinja2==3.1.4
+jinja2==3.1.5
     # via
     #   altair
     #   pydeck
