### 2025-01-01 b621de0
chore: use _version.py for version numbering
