### 2025-01-01 0523185
chore: add setup for semantic release
