+++ pyproject.toml
@@ -74,6 +74,28 @@ exclude = [
 [tool.hatch.version]
 path = "pyprobe/_version.py"
 
+[tool.semantic_release]
+version_source = "tag"
+version_variables = ["pyprobe/_version.py:__version__"]
+changelog_file = false
+branch = "main" 
+commit_version_number = false  # disable auto commits
+tag_commit = false            # disable auto tagging
+prerelease = true
+upload_to_pypi = false        # disable PyPI upload
+upload_to_release = false     # disable GitHub release creation
+hooks = [
+    "bash -c 'echo \"Running uv lock\" && uv lock --upgrade-package pyprobe-data'"
+]
+
+[tool.semantic_release.branches.main]
+match = "main"
+prerelease = false
+
+[tool.semantic_release.branches.development]
+match = ".*"  # Match any branch except main
+prerelease = true
+
 [tool.mypy]
 disallow_any_explicit = false
 disallow_any_generics = true
