### 2025-01-01 507e396
chore: test semantic release in github action
