### 2025-01-01 d917238
chore: Add python-semantic-release as dependency
