### 2025-01-01 c4dd660
move update version into ci workflow
