### 2025-01-01 83d8d90
Revert "move update version into ci workflow"
This reverts commit c4dd660c66acf44fa23a16bb99735866a0c27153.
