+++ .github/workflows/ci.yml
@@ -1,19 +1,25 @@
-name: "Pull Request Checks"
-
+name: "Build and test"
 on:
   pull_request:
     types: [opened, synchronize, reopened, ready_for_review]
-    branches: [main]
-
+    branches:
+      - main
+  workflow_dispatch:
 jobs:
-  test-and-version:
+  qa:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
         with:
-          fetch-depth: 0
-          ref: ${{ github.head_ref }}
-
+          python-version-file: "pyproject.toml"
+      - uses: pre-commit/action@v3.0.0
+  test:
+    needs: qa
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v4
       - name: Install uv
         uses: astral-sh/setup-uv@v5
 
@@ -21,32 +27,27 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version-file: "pyproject.toml"
-
+      
       - name: Install the project
         run: uv sync --all-extras --group format --group test --group docs --group external-integration
-  
-      - name: Get branch name
-        id: branch-name
-        run: echo "branch=${GITHUB_HEAD_REF:-${GITHUB_REF#refs/heads/}}" >> $GITHUB_OUTPUT
 
       - name: Run tests
-        run: uv run pytest
+        run: uv run pytest tests --benchmark-json output.json
 
+      - name: Store benchmark result
+        uses: benchmark-action/github-action-benchmark@v1
+        with:
+          name: Python Benchmark with pytest-benchmark
+          tool: 'pytest'
+          output-file-path: output.json
+          # Use personal access token instead of GITHUB_TOKEN due to https://github.community/t/github-action-not-triggering-gh-pages-upon-push/16096
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+          # Show alert with commit comment on detecting possible performance regression
+          alert-threshold: '110%'
+          comment-on-alert: true
+          # Enable Job Summary for PRs
+          summary-always: true
       - name: Install nbmake
         run: uv pip install nbmake
       - name: Test notebooks
         run: uv run pytest --nbmake docs/source/examples/*ipynb
-
-      - name: Update version
-        run: |
-          uv run semantic-release version \
-            --prerelease \
-            --no-commit \
-            --no-tag \
-            --skip-build \
-            --build-metadata ${{ steps.branch-name.outputs.branch }}
-
-      - uses: stefanzweifel/git-auto-commit-action@v5
-        with:
-          commit_message: "chore: update version [skip ci]"
-          file_pattern: "pyprobe/_version.py"
