+++ .github/workflows/publish-to-testpypi.yml
@@ -31,10 +31,6 @@ jobs:
       - name: Install the project
         run: uv sync --all-extras --group format --group test --group docs --group external-integration
 
-      - name: Get branch name
-        id: branch-name
-        run: echo "branch=${GITHUB_HEAD_REF:-${GITHUB_REF#refs/heads/}}" >> $GITHUB_OUTPUT
-
       - name: Run semantic-release
         run: |
           uv run semantic-release version \
@@ -42,7 +38,6 @@ jobs:
             --no-commit \
             --no-tag \
             --skip-build \
-            --build-metadata ${{ steps.branch-name.outputs.branch }}
     
       - uses: stefanzweifel/git-auto-commit-action@v5
         with:
