### 2025-01-01 0284dbf
Checkout all git history in workflow
