+++ .github/workflows/publish-to-testpypi.yml
@@ -20,6 +20,10 @@ jobs:
 
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0  # Fetch all history
+          fetch-tags: true  # Fetch all tags
+          ref: ${{ github.head_ref || github.ref_name }}  # Correct branch reference
 
       - name: Install uv
         uses: astral-sh/setup-uv@v5
