+++ pyprobe/_version.py
@@ -1,1 +1,1 @@
-__version__ = "1.1.4"
+__version__ = "1.1.4-rc.4"
