+++ CHANGELOG.md
@@ -1,0 +1,75 @@
+# CHANGELOG
+
+
+## v1.1.4-rc.4 (2025-01-01)
+
+### Chores
+
+- Add python-semantic-release as dependency
+  ([`d917238`](https://github.com/ImperialCollegeLondon/PyProBE/commit/d9172386681c0501d1198caad402110eb7e7a1be))
+
+- Test semantic release in github action
+  ([`507e396`](https://github.com/ImperialCollegeLondon/PyProBE/commit/507e3962fa9bcec3d273253531803a384506b3d6))
+
+
+## v1.1.4-rc.1 (2025-01-01)
+
+### Chores
+
+- Add setup for semantic release
+  ([`0523185`](https://github.com/ImperialCollegeLondon/PyProBE/commit/0523185db1bdcaf3d6027d2de8e1b1c990e8b9a7))
+
+- Use _version.py for version numbering
+  ([`b621de0`](https://github.com/ImperialCollegeLondon/PyProBE/commit/b621de0f77ca0684456671b170a05d5bc94ae9aa))
+
+
+## v1.1.4 (2024-12-31)
+
+
+## v1.1.3 (2024-12-07)
+
+
+## v1.1.2 (2024-12-07)
+
+
+## v1.1.1 (2024-12-03)
+
+
+## v1.1.0 (2024-11-29)
+
+### Documentation
+
+- **contributor**: Contrib-readme-action has updated readme
+  ([`70da06d`](https://github.com/ImperialCollegeLondon/PyProBE/commit/70da06d55a8254b6b2a4a5deb67d950b68782f68))
+
+
+## v1.0.3 (2024-10-04)
+
+
+## v1.0.2 (2024-09-16)
+
+
+## v1.0.1 (2024-09-16)
+
+
+## v1.0.0 (2024-09-16)
+
+
+## v0.1.4 (2024-06-18)
+
+
+## v0.1.3 (2024-06-17)
+
+
+## v0.1.2 (2024-06-11)
+
+
+## v0.1.1 (2024-06-05)
+
+
+## v0.1.0 (2024-06-02)
+
+### Documentation
+
+- **contributor**: Contrib-readme-action has updated readme
+  ([`ba1a272`](https://github.com/ImperialCollegeLondon/PyProBE/commit/ba1a272a78893fd73aa18ee44630fd7356fefcef))
