### 2025-01-01 9458f82
chore: add python-semantic-release as dependency
