### 2025-01-01 be4aa88
chore: set up workflow for pre-release
