### 2025-01-01 c5cd674
tests: fix failing cell archive test
