+++ .github/workflows/publish-to-pypi.yml
@@ -1,44 +1,0 @@
-name: "Publish to PyPI"
-
-on:
-  release:
-    types: [published]
-  workflow_dispatch:
-
-jobs:
-  Build:
-    runs-on: ubuntu-latest
-
-    permissions:
-      id-token: write  # IMPORTANT: mandatory for trusted publishing
-
-    environment:
-      name: pypi
-      url: https://pypi.org/p/PyProBE-Data
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v5
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version-file: "pyproject.toml"
-          
-      - name: Install the project
-        run: uv sync --all-extras --group format --group test --group docs --group external-integration
-
-      - name: Build
-        run: uv build
-      
-      - name: Upload artifact
-        uses: actions/upload-artifact@v4
-        with:
-          name: dist
-          path: dist/
-          retention-days: 5  # Keep artifacts for 5 days
-
-      - name: Publish
-        run: uv publish --trusted-publishing always
