+++ .github/workflows/publish-to-testpypi.yml
@@ -1,73 +1,0 @@
-name: "Publish to TestPyPI"
-
-on:
-  push:
-    branches:
-      - main
-  workflow_dispatch:
-
-jobs:
-  Build:
-    runs-on: ubuntu-latest
-
-    permissions:
-      id-token: write  # IMPORTANT: mandatory for trusted publishing
-      contents: write    # for git operations
-
-    environment:
-      name: testpypi
-      url: https://test.pypi.org/p/PyProBE-Data
-
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0  # Fetch all history
-          fetch-tags: true  # Fetch all tags
-          ref: ${{ github.head_ref || github.ref_name }}  # Correct branch reference
-      
-      - name: Install uv
-        uses: astral-sh/setup-uv@v5
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version-file: "pyproject.toml"
-          
-      - name: Install the project
-        run: uv sync --all-extras --group format --group test --group docs --group external-integration
-
-      - name: Run semantic-release
-        id: semrel
-        run: |
-          uv run semantic-release version \
-            --prerelease \
-            --no-commit \
-            --no-tag \
-            --skip-build
-
-      - name: Update uv lockfile
-        run:  uv lock --upgrade-package pyprobe-data
-
-      - name: Get version
-        id: version
-        run: |
-          VERSION=$(grep -o '".*"' pyprobe/_version.py | sed 's/"//g')
-          echo "version=$VERSION" >> $GITHUB_OUTPUT
-
-      - uses: stefanzweifel/git-auto-commit-action@v5
-        with:
-          commit_message: Update version [skip ci]
-          tagging_message: "v${{ steps.version.outputs.version }}"
-
-      - name: Build
-        run: uv build
-      
-      - name: Upload artifact
-        uses: actions/upload-artifact@v4
-        with:
-          name: dist
-          path: dist/
-          retention-days: 5  # Keep artifacts for 5 days
-
-      - name: Publish
-        run: uv publish --publish-url https://test.pypi.org/legacy/ --trusted-publishing always
