### 2025-01-01 12c9900
chore: use open PR action only
