### 2025-01-01 e1b8e3b
Update permissions for workflow
