### 2025-01-01 f1b8122
chore: test auto publishing
