### 2025-01-01 d574f33
chore: release candidate 1.2.0 [skip ci]
