+++ CHANGELOG.md
@@ -1,0 +1,84 @@
+# CHANGELOG
+
+
+## v1.2.0 (2025-01-01)
+
+### Chores
+
+- Add python-semantic-release as dependency
+  ([`9458f82`](https://github.com/ImperialCollegeLondon/PyProBE/commit/9458f82274ddb03481fa8325fc84a753163b3574))
+
+- Add setup for semantic release
+  ([`0523185`](https://github.com/ImperialCollegeLondon/PyProBE/commit/0523185db1bdcaf3d6027d2de8e1b1c990e8b9a7))
+
+- Set up workflow for pre-release
+  ([`be4aa88`](https://github.com/ImperialCollegeLondon/PyProBE/commit/be4aa88b2f4002b950a5b65684a3065bdb7e28c2))
+
+- Test auto publishing
+  ([`f1b8122`](https://github.com/ImperialCollegeLondon/PyProBE/commit/f1b8122e12f75d54577efc47b7899ce7a056d8cf))
+
+- Update create-pr version
+  ([`e360490`](https://github.com/ImperialCollegeLondon/PyProBE/commit/e360490412e8fdef81f6958727ee8b7c15f919b8))
+
+- Use _version.py for version numbering
+  ([`b621de0`](https://github.com/ImperialCollegeLondon/PyProBE/commit/b621de0f77ca0684456671b170a05d5bc94ae9aa))
+
+- Use open PR action only
+  ([`12c9900`](https://github.com/ImperialCollegeLondon/PyProBE/commit/12c9900615179edf8cf9c9bba82222c48a1626e7))
+
+### Features
+
+- Add another feature
+  ([`fccd34a`](https://github.com/ImperialCollegeLondon/PyProBE/commit/fccd34a97f40e6b4198842fb3fb707028a9a62c6))
+
+
+## v1.1.4 (2024-12-31)
+
+
+## v1.1.3 (2024-12-07)
+
+
+## v1.1.2 (2024-12-07)
+
+
+## v1.1.1 (2024-12-03)
+
+
+## v1.1.0 (2024-11-29)
+
+### Documentation
+
+- **contributor**: Contrib-readme-action has updated readme
+  ([`70da06d`](https://github.com/ImperialCollegeLondon/PyProBE/commit/70da06d55a8254b6b2a4a5deb67d950b68782f68))
+
+
+## v1.0.3 (2024-10-04)
+
+
+## v1.0.2 (2024-09-16)
+
+
+## v1.0.1 (2024-09-16)
+
+
+## v1.0.0 (2024-09-16)
+
+
+## v0.1.4 (2024-06-18)
+
+
+## v0.1.3 (2024-06-17)
+
+
+## v0.1.2 (2024-06-11)
+
+
+## v0.1.1 (2024-06-05)
+
+
+## v0.1.0 (2024-06-02)
+
+### Documentation
+
+- **contributor**: Contrib-readme-action has updated readme
+  ([`ba1a272`](https://github.com/ImperialCollegeLondon/PyProBE/commit/ba1a272a78893fd73aa18ee44630fd7356fefcef))
