### 2025-01-01 da9afe7
chore: release candidate 1.2.0
