+++ .github/workflows/publish-to-testpypi.yml
@@ -3,8 +3,7 @@ name: "Publish to TestPyPI"
 on:
   push:
     branches:
-      - release-candidates/*
-  workflow_dispatch:
+      - release-candidates*
 
 jobs:
   Build:
