### 2025-01-01 ad1e81d
chore: correct asterisk
