### 2025-01-01 4730f97
chore: release candidate 1.2.0
