### 2025-01-01 57b9de6
chore: replace auto publishing with workflow to create an rc PR
