### 2025-01-02 f1c40a0
fix(result): bug in combine_results method
