### 2025-01-02 ae27260
chore: add ipykernel dependency
