+++ pyproject.toml
@@ -10,6 +10,7 @@ dependencies = [
     "deprecated>=1.2.15",
     "distinctipy>=1.3.4",
     "fastexcel>=0.12.0",
+    "ipykernel>=6.29.5",
     "ipython>=8.31.0",
     "kaleido==0.2.1",
     "matplotlib>=3.10.0",
