### 2025-01-02 217598c
chore: Add hvplot and seaborn as optional dependencies
