+++ pyproject.toml
@@ -26,6 +26,14 @@ dependencies = [
     "sympy>=1.13.3",
 ]
 
+[project.optional-dependencies]
+hvplot = [
+    "hvplot>=0.11.2",
+]
+seaborn = [
+    "seaborn>=0.13.2",
+]
+
 [dependency-groups]
 format = [
     "black>=24.10.0",
