### 2025-01-02 c8fdcc5
fix: pass label argument to plot through as string
Prevents collection of columns if label argument happens to match a column name
