### 2025-01-02 cf5f627
refactor: remove unit information from column definitions
