### 2025-01-02 dd2cbf9
docs: update examples to use .plot() method
