### 2025-01-02 b3011d5
chore: remove show_image method and associated dependencies
kaleido and ipython removed as required dependencies. This is not a breaking change as the .show() method remains.
