+++ pyproject.toml
@@ -11,8 +11,6 @@ dependencies = [
     "distinctipy>=1.3.4",
     "fastexcel>=0.12.0",
     "ipykernel>=6.29.5",
-    "ipython>=8.31.0",
-    "kaleido==0.2.1",
     "matplotlib>=3.10.0",
     "numpy>=2.2.1",
     "ordered-set>=4.1.0",
