+++ requirements.txt
@@ -1,250 +1,0 @@
-#
-# This file is autogenerated by pip-compile with Python 3.12
-# by the following command:
-#
-#    pip-compile
-#
-aiosignal==1.3.1
-    # via ray
-altair==5.4.0
-    # via streamlit
-annotated-types==0.7.0
-    # via pydantic
-anytree==2.12.1
-    # via pybamm
-asttokens==2.4.1
-    # via stack-data
-attrs==24.2.0
-    # via
-    #   jsonschema
-    #   referencing
-blinker==1.8.2
-    # via streamlit
-cachetools==5.5.0
-    # via streamlit
-casadi==3.6.6
-    # via pybamm
-certifi==2024.7.4
-    # via requests
-charset-normalizer==3.3.2
-    # via requests
-click==8.1.7
-    # via
-    #   ray
-    #   streamlit
-contourpy==1.3.0
-    # via matplotlib
-cycler==0.12.1
-    # via matplotlib
-decorator==5.1.1
-    # via ipython
-deprecated==1.2.15
-    # via PyProBE (pyproject.toml)
-distinctipy==1.3.4
-    # via PyProBE (pyproject.toml)
-executing==2.1.0
-    # via stack-data
-fastexcel==0.11.5
-    # via PyProBE (pyproject.toml)
-filelock==3.16.1
-    # via ray
-fonttools==4.53.1
-    # via matplotlib
-frozenlist==1.5.0
-    # via
-    #   aiosignal
-    #   ray
-gitdb==4.0.11
-    # via gitpython
-gitpython==3.1.43
-    # via streamlit
-idna==3.7
-    # via requests
-ipython==8.26.0
-    # via PyProBE (pyproject.toml)
-jedi==0.19.1
-    # via ipython
-jinja2==3.1.4
-    # via
-    #   altair
-    #   pydeck
-joblib==1.4.2
-    # via scikit-learn
-jsonschema==4.23.0
-    # via
-    #   altair
-    #   ray
-jsonschema-specifications==2023.12.1
-    # via jsonschema
-kaleido==0.2.1
-    # via PyProBE (pyproject.toml)
-kiwisolver==1.4.7
-    # via matplotlib
-markdown-it-py==3.0.0
-    # via rich
-markupsafe==2.1.5
-    # via jinja2
-matplotlib==3.9.2
-    # via PyProBE (pyproject.toml)
-matplotlib-inline==0.1.7
-    # via ipython
-mdurl==0.1.2
-    # via markdown-it-py
-mpmath==1.3.0
-    # via sympy
-msgpack==1.1.0
-    # via ray
-narwhals==1.8.0
-    # via altair
-numpy==1.26.4
-    # via
-    #   PyProBE (pyproject.toml)
-    #   casadi
-    #   contourpy
-    #   distinctipy
-    #   matplotlib
-    #   pandas
-    #   pyarrow
-    #   pybamm
-    #   pydeck
-    #   scikit-learn
-    #   scipy
-    #   streamlit
-    #   xarray
-ordered-set==4.1.0
-    # via PyProBE (pyproject.toml)
-packaging==24.1
-    # via
-    #   altair
-    #   matplotlib
-    #   plotly
-    #   pooch
-    #   ray
-    #   streamlit
-    #   xarray
-pandas==2.2.3
-    # via
-    #   pybamm
-    #   streamlit
-    #   xarray
-parso==0.8.4
-    # via jedi
-pexpect==4.9.0
-    # via ipython
-pillow==10.4.0
-    # via
-    #   matplotlib
-    #   streamlit
-platformdirs==4.2.2
-    # via pooch
-plotly==5.23.0
[…]
