### 2025-01-02 860b663
docs: simplify developer installation instructions
Remove requirement to specify all groups independently
