### 2025-01-02 fe47fc0
docs: Update contributing guidelines
