### 2025-01-02 82d6867
chore: add nbmake as developer requirement
