### 2025-01-02 0617e37
docs: add explanations for quick_add_procedure
