### 2025-01-02 28b7a48
feat: allow specification of header rows in experiment log file
