### 2025-01-02 3eb5919
chore: release candidate 1.2.0-rc.1 [skip ci]
