+++ pyprobe/_version.py
@@ -1,1 +1,1 @@
-__version__ = "1.1.4"
+__version__ = "1.2.0-rc.1"
