+++ CHANGELOG.md
@@ -1,0 +1,168 @@
+# CHANGELOG
+
+
+## v1.2.0-rc.1 (2025-01-02)
+
+### Bug Fixes
+
+- Error from rebase
+  ([`62a0e5f`](https://github.com/ImperialCollegeLondon/PyProBE/commit/62a0e5fe06f49576223c0b088d51d372c47ea21b))
+
+- Issue with missing experiment tuple on dashboard
+  ([`21d22c8`](https://github.com/ImperialCollegeLondon/PyProBE/commit/21d22c84af595df621501cba177a590bf33414ee))
+
+- Pass label argument to plot through as string
+  ([`c8fdcc5`](https://github.com/ImperialCollegeLondon/PyProBE/commit/c8fdcc5d1d0dea6f3b3321a7952b5ee518d6683e))
+
+Prevents collection of columns if label argument happens to match a column name
+
+- **result**: Bug in combine_results method
+  ([`f1c40a0`](https://github.com/ImperialCollegeLondon/PyProBE/commit/f1c40a0780245c34b1abadf6f9c993309d85271a))
+
+### Chores
+
+- Add hvplot and seaborn as optional dependencies
+  ([`217598c`](https://github.com/ImperialCollegeLondon/PyProBE/commit/217598c7a809948adaf8beab4ea321479afaea4a))
+
+- Add ipykernel dependency
+  ([`ae27260`](https://github.com/ImperialCollegeLondon/PyProBE/commit/ae272601db1bcfd5f6209f5a3ac384eeff48a364))
+
+- Add nbmake as developer requirement
+  ([`82d6867`](https://github.com/ImperialCollegeLondon/PyProBE/commit/82d68678bdeb1077e035b897077a1ff6cccd23f6))
+
+- Add python-semantic-release as dependency
+  ([`9458f82`](https://github.com/ImperialCollegeLondon/PyProBE/commit/9458f82274ddb03481fa8325fc84a753163b3574))
+
+- Add setup for semantic release
+  ([`0523185`](https://github.com/ImperialCollegeLondon/PyProBE/commit/0523185db1bdcaf3d6027d2de8e1b1c990e8b9a7))
+
+- Remove pip-tools requirements files
+  ([`5924f31`](https://github.com/ImperialCollegeLondon/PyProBE/commit/5924f311faa41540e13de85aa17df835f3331bc0))
+
+- Remove show_image method and associated dependencies
+  ([`b3011d5`](https://github.com/ImperialCollegeLondon/PyProBE/commit/b3011d570b3f54e34c1b591d5fecff4304492868))
+
+kaleido and ipython removed as required dependencies. This is not a breaking change as the .show()
+  method remains.
+
+- Replace auto publishing with workflow to create an rc PR
+  ([`57b9de6`](https://github.com/ImperialCollegeLondon/PyProBE/commit/57b9de60004474c7cb046ef83e97d5e4fa2ca564))
+
+- Set up workflow for pre-release
+  ([`be4aa88`](https://github.com/ImperialCollegeLondon/PyProBE/commit/be4aa88b2f4002b950a5b65684a3065bdb7e28c2))
+
+- Use _version.py for version numbering
+  ([`b621de0`](https://github.com/ImperialCollegeLondon/PyProBE/commit/b621de0f77ca0684456671b170a05d5bc94ae9aa))
+
+### Documentation
+
+- Add explanations for quick_add_procedure
+  ([`0617e37`](https://github.com/ImperialCollegeLondon/PyProBE/commit/0617e37791f730f93aaa0e246cfc47170b0a81ca))
+
[…]
