+++ _sources/examples/plotting.ipynb.txt
@@ -1,0 +1,230 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Plotting\n",
+    "PyProBE includes compatibility with multiple open plotting libraries in Python. This example will go through interacting with each of them."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Start by importing PyProBE and the dataset:"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import pyprobe\n",
+    "import matplotlib.pyplot as plt\n",
+    "%matplotlib inline\n",
+    "info_dictionary = {'Name': 'Sample cell',\n",
+    "                   'Chemistry': 'NMC622',\n",
+    "                   'Nominal Capacity [Ah]': 0.04,\n",
+    "                   'Cycler number': 1,\n",
+    "                   'Channel number': 1,}\n",
+    "data_directory = '../../../tests/sample_data/neware'\n",
+    "\n",
+    "# Create a cell object\n",
+    "cell = pyprobe.Cell(info=info_dictionary)\n",
+    "cell.add_procedure(procedure_name='Sample',\n",
+    "                   folder_path = data_directory,\n",
+    "                   filename = 'sample_data_neware.parquet')"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Pandas/Matplotlib"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "[Matplotlib](https://matplotlib.org/) is one of the most popular plotting libraries in Python. [Pandas](https://pandas.pydata.org/) dataframes include a built-in `plot()` method to construct plots using Matplotlib, which is also integrated into PyProBE `Result` objects. \n",
+    "\n",
+    "Call `plot()` on any result object and control the method exactly as you would expect to with Pandas:"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "plt.figure()\n",
+    "cell.procedure['Sample'].experiment('Break-in Cycles').plot(x='Time [s]', y='Voltage [V]', grid=True)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## HvPlot\n",
+    "Plots created using Matplotlib sometimes need a lot of work to look good, and they are not particularly interactive. [HvPlot](https://hvplot.holoviz.org/index.html) provides a high-level interface similar to plotting in pandas with the additional flexibility of specifying a more interactive backend. Note hvplot is an optional dependency, it can be installed seperately to enable this functionality."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "cell.procedure['Sample'].experiment('Break-in Cycles').hvplot(x='Time [s]', y='Voltage [V]')"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Seaborn\n",
+    "Seaborn is a high-level plotting library built on top of matplotlib. It provides a clean interface for dataframes, while maintaining much of matplotlib's flexibility.\n",
+    "\n",
[…]
