### 2025-01-02 7860e08
chore: release candidate 1.2.0 [skip ci]
