### 2025-01-02 35b33bd
docs: add detail for creating a jupyter kernel from uv .venv
