### 2025-01-02 80cf916
chore: add condition to release candidate workflow
Prevents running the workflow when a merge is made from a branch created by the release candidate workflow itself
