+++ .github/workflows/create-release-candidate.yml
@@ -7,6 +7,7 @@ on:
 
 jobs:
   create-rc:
+    if: ${{ !contains(github.event.head_commit.message, 'release-candidates/') }}
     runs-on: ubuntu-latest
 
     permissions:
