+++ _sources/examples/analysing-GITT-data.ipynb.txt
@@ -20,6 +20,7 @@
    "source": [
     "import pyprobe\n",
     "import matplotlib.pyplot as plt\n",
+    "%matplotlib inline\n",
     "\n",
     "info_dictionary = {'Name': 'Sample cell',\n",
     "                   'Chemistry': 'NMC622',\n",
