+++ _sources/examples/LEAN-differentiation.ipynb.txt
@@ -24,6 +24,8 @@
    "source": [
     "import pyprobe\n",
     "import matplotlib.pyplot as plt\n",
+    "%matplotlib inline\n",
+    "\n",
     "info_dictionary = {'Name': 'Sample cell',\n",
     "                   'Chemistry': 'NMC622',\n",
     "                   'Nominal Capacity [Ah]': 0.04,\n",
