### 2025-01-02 cd4b853
docs: add some missing citations and describe plotting integrations
