### 2025-01-02 2202c22
chore: move dev dependencies into optional dependency group
Dependency groups are currently not supported for backwards compatibility with pip. Wait for resolution of PEP 735
