### 2025-01-03 a1a76d2
chore: update workflows for dependency reorganisation
