### 2025-01-03 0045fd4
docs: add optional dependency detail for hvplot and seaborn
