### 2025-01-03 fc720d1
test: move pybamm out of dev dependencies
Use pytest.importorskip for skipping pybamm tests if it is not installed
