+++ pyproject.toml
@@ -48,7 +48,6 @@ dev = [
     "types-deprecated>=1.2.15.20241117",
     "types-pyyaml>=6.0.12.20241230",
     "types-toml>=0.10.8.20240310",
-    "pybamm>=24.1"
 ]
 docs = [
     "autodoc-pydantic>=2.2.0",
@@ -59,6 +58,9 @@ docs = [
     "sphinx-tabs>=3.4.7",
     "sphinxcontrib-bibtex>=2.6.3",
     "xlsxwriter>=3.2.0",
+]
+pybamm = [
+    "pybamm>=24.1",
 ]
 
 [build-system]
