### 2025-01-03 699a850
chore: recategorise xlsxwriter as dev dependency
