### 2025-01-03 5e0ce2d
test: don't check column order in add_procedure test
