+++ tests/test_cell.py
@@ -145,13 +145,19 @@ def test_add_procedure(cell_instance, procedure_fixture, benchmark):
         return cell_instance.add_procedure(title, input_path, file_name)
 
     benchmark(add_procedure)
-    assert_frame_equal(cell_instance.procedure[title].data, procedure_fixture.data)
+    assert_frame_equal(
+        cell_instance.procedure[title].data,
+        procedure_fixture.data,
+        check_column_order=False,
+    )
 
     cell_instance.add_procedure(
         "Test_custom", input_path, file_name, readme_name="README_total_steps.yaml"
     )
     assert_frame_equal(
-        cell_instance.procedure["Test_custom"].data, procedure_fixture.data
+        cell_instance.procedure["Test_custom"].data,
+        procedure_fixture.data,
+        check_column_order=False,
     )
 
 
