### 2025-01-03 #7474 @tomjholland
@jakobhaervig following your feedback on the installation process, I have made a few improvements:
- for the user installation, the package can now be simply pip-installed with `pip install PyProBE-Data` so the requirements.txt files are not required
- `pybamm` is now an optional dependency, so the warnings that you were seeing now no longer appear when importing `pyprobe` into a notebook
- I have switched from pip-tools to uv for managing the dependencies and developer install
  - This automates the specification of dependency versions in the `pyproject.toml` file, replacing the `requirements-dev.txt` with a platform-agnostic `uv.lock`
  - An editable developer install is still possible with `pip install -e '.[dev, docs]'`
