### 2025-01-04 9219eaf
refactor: remove setting of a default name and colour assignment in cell
