### 2025-01-04 5657285
feat: add a selectbox for a cell identifier to replace the "Name" field
