### 2025-01-04 6a5e1c4
refactor: move cell identifier to below plot
