### 2025-01-04 9894e58
refactor: remove search for color entry in info dict
This changes the default behaviour to cycle through the colours built-in to plotly
