+++ pyprobe/plot.py
@@ -226,11 +226,6 @@ class Plot:
             dash (str): The dash style of the line. E.g. "dash", "dot", "dashdot".
             showlegend (bool): Whether to show the legend.
         """
-        if color is None:
-            if "color" in result.info:
-                color = str(result.info["color"])
-            else:
-                color = "blue"
         if label is None:
             if "Name" in result.info:
                 label = str(result.info["Name"])
