### 2025-01-04 c610126
feat: allow any dict values in info dictionary
