### 2025-01-04 f9df697
test: fix Plot class tests
