### 2025-01-04 d7704cc
refactor: remove colour generation in make_cell_list method
