+++ pyproject.toml
@@ -8,12 +8,10 @@ requires-python = ">=3.11,<3.13"
 license = { file = "LICENSE" }
 dependencies = [
     "deprecated>=1.2.15",
-    "distinctipy>=1.3.4",
     "fastexcel>=0.12.0",
     "ipykernel>=6.29.5",
     "matplotlib>=3.10.0",
     "numpy>=2.2.1",
-    "ordered-set>=4.1.0",
     "plotly>=5.24.1",
     "polars>=1.18.0",
     "pydantic>=2.10.4",
