### 2025-01-04 cc1313f
docs: remove mention of the requirement to specify a "Name" for a cell
