### 2025-01-05 8b7c508
refactor: replace black, flake8 and isort config with ruff
