### 2025-01-05 f2a79c2
chore: ignore D103 (docstring in public function) in examples
