### 2025-01-06 e32a067
chore: fix ruff linting errors
