### 2025-01-06 e8a267c
style: run ruff format on pyprobe/ and tests/
