+++ pyprobe/__init__.py
@@ -1,4 +1,5 @@
 """The PyProBE package."""
+
 from ._version import __version__  # noqa: F401
 from .cell import Cell, load_archive, make_cell_list  # noqa: F401
 from .dashboard import launch_dashboard  # noqa: F401
