+++ tests/test_units.py
@@ -1,4 +1,5 @@
 """Test the Units class."""
+
 import numpy as np
 import polars as pl
 
