### 2025-01-07 #7474 @phibeck
Thank you very much @TomTranter, @jakobhaervig and @ritesh001 for the updates and your thorough reviews! 🎉

@tomjholland the reviewers have recommended the submission for publication. There are a few more steps before we finalize the publication.

> @phibeck I caught a typo on Figure 5 in the manuscript, so took the opportunity to correct a very minor UI change that I have made in the package since submission in the same figure. The changes are made in this PR: [ImperialCollegeLondon/PyProBE#179](https://github.com/ImperialCollegeLondon/PyProBE/pull/179), if there are no issues I will merge it with your permission. I can then create a final repo release, which I assume the version on the JOSS submission can be updated to reflect?

That's okay. You can go ahead and merge it, and ask editorialbot to regenerate the pdf once you're done using `@editorialbot generate pdf`. Indeed, the next steps for you are making a final release and archive that version, see my next comment.
