+++ _sources/examples/differentiating-voltage-data.ipynb.txt
@@ -26,20 +26,25 @@
    "source": [
     "import pyprobe\n",
     "import matplotlib.pyplot as plt\n",
+    "\n",
     "%matplotlib inline\n",
     "\n",
-    "info_dictionary = {'Name': 'Sample cell',\n",
-    "                   'Chemistry': 'NMC622',\n",
-    "                   'Nominal Capacity [Ah]': 0.04,\n",
-    "                   'Cycler number': 1,\n",
-    "                   'Channel number': 1,}\n",
-    "data_directory = '../../../tests/sample_data/neware'\n",
+    "info_dictionary = {\n",
+    "    \"Name\": \"Sample cell\",\n",
+    "    \"Chemistry\": \"NMC622\",\n",
+    "    \"Nominal Capacity [Ah]\": 0.04,\n",
+    "    \"Cycler number\": 1,\n",
+    "    \"Channel number\": 1,\n",
+    "}\n",
+    "data_directory = \"../../../tests/sample_data/neware\"\n",
     "\n",
     "# Create a cell object\n",
     "cell = pyprobe.Cell(info=info_dictionary)\n",
-    "cell.add_procedure(procedure_name='Sample',\n",
-    "                   folder_path = data_directory,\n",
-    "                   filename = 'sample_data_neware.parquet')"
+    "cell.add_procedure(\n",
+    "    procedure_name=\"Sample\",\n",
+    "    folder_path=data_directory,\n",
+    "    filename=\"sample_data_neware.parquet\",\n",
+    ")"
    ]
   },
   {
@@ -55,9 +60,9 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "final_cycle= cell.procedure['Sample'].experiment('Break-in Cycles').cycle(-1)\n",
+    "final_cycle = cell.procedure[\"Sample\"].experiment(\"Break-in Cycles\").cycle(-1)\n",
     "\n",
-    "final_cycle.discharge(0).plot(x='Time [hr]', y='Voltage [V]')"
+    "final_cycle.discharge(0).plot(x=\"Time [hr]\", y=\"Voltage [V]\")"
    ]
   },
   {
@@ -76,9 +81,11 @@
    "source": [
     "from pyprobe.analysis import differentiation\n",
     "\n",
-    "raw_data_dVdQ = differentiation.gradient(final_cycle.discharge(0),\"Capacity [Ah]\", \"Voltage [V]\")\n",
+    "raw_data_dVdQ = differentiation.gradient(\n",
+    "    final_cycle.discharge(0), \"Capacity [Ah]\", \"Voltage [V]\"\n",
+    ")\n",
     "print(raw_data_dVdQ.column_list)\n",
-    "raw_data_dVdQ.plot(x='Capacity [Ah]', y='Voltage [V]')"
+    "raw_data_dVdQ.plot(x=\"Capacity [Ah]\", y=\"Voltage [V]\")"
    ]
   },
   {
@@ -99,13 +106,18 @@
     "from pyprobe.analysis import smoothing\n",
     "\n",
     "\n",
-    "downsampled_data = smoothing.downsample(input_data = final_cycle.discharge(0),\n",
-    "                                            target_column='Voltage [V]',\n",
-    "                                                  sampling_interval=0.002,\n",
-    "                                                )\n",
+    "downsampled_data = smoothing.downsample(\n",
[…]
