### 2025-01-08 8b3e2d3
chore: update workflow to v7 of create-pull-request gh action [skip ci]
