+++ .github/workflows/create-release-candidate.yml
@@ -44,7 +44,7 @@ jobs:
         run: uv lock --upgrade-package pyprobe-data
 
       - name: Open Pull Request
-        uses: peter-evans/create-pull-request@v4
+        uses: peter-evans/create-pull-request@v7
         with:
           branch: release-candidates/rc-${{ steps.bump.outputs.version }}
           title: "Release Candidate ${{ steps.bump.outputs.version }}"
