### 2025-01-08 458e4c6
docs: fix errors and typos in documentation
