+++ docs/source/api/pyprobe.cyclers.rst
@@ -12,8 +12,11 @@ Submodules
 .. toctree::
    :maxdepth: 4
 
+   pyprobe.cyclers.arbin
    pyprobe.cyclers.basecycler
+   pyprobe.cyclers.basytec
    pyprobe.cyclers.biologic
+   pyprobe.cyclers.maccor
    pyprobe.cyclers.neware
 
 
