### 2025-01-08 0e4a25f
chore: add readme info to pyproject.toml
