+++ pyproject.toml
@@ -6,6 +6,7 @@ authors = [
 dynamic = ["version"]
 requires-python = ">=3.11,<3.13"
 license = { file = "LICENSE" }
+readme = {file = "README.md", content-type = "text/markdown"}
 dependencies = [
     "deprecated>=1.2.15",
     "fastexcel>=0.12.0",
