### 2025-01-11 7ad73e3
docs: create pybamm integration example
