### 2025-01-11 e8549d9
docs: add mention of pybamm integration to paper
