### 2025-01-12 ea8e10b
chore: update nbstripout precommit hook to remove kernelspec
