+++ .pre-commit-config.yaml
@@ -29,6 +29,7 @@ repos:
     rev: 0.7.1
     hooks:
       - id: nbstripout
+        args: ['--extra-keys="metadata.kernelspec"']
   - repo: https://github.com/astral-sh/uv-pre-commit
     # uv version.
     rev: 0.5.8
