### 2025-01-12 62154c3
chore: use % magic to install pybamm and matplotlib in example notebooks
