+++ pyproject.toml
@@ -10,7 +10,6 @@ dependencies = [
     "deprecated>=1.2.15",
     "fastexcel>=0.12.0",
     "ipykernel>=6.29.5",
-    "matplotlib>=3.10.0",
     "numpy>=2.2.1",
     "plotly>=5.24.1",
     "polars>=1.18.0",
@@ -53,9 +52,6 @@ docs = [
     "sphinx-design>=0.6.1",
     "sphinx-tabs>=3.4.7",
     "sphinxcontrib-bibtex>=2.6.3",
-]
-pybamm = [
-    "pybamm>=24.1",
 ]
 
 [build-system]
