### 2025-01-12 2ffc0a2
chore: make ruff a dev dependency
