+++ pyproject.toml
@@ -16,7 +16,6 @@ dependencies = [
     "pydantic>=2.10.4",
     "pyyaml>=6.0.2",
     "ray>=2.40.0",
-    "ruff>=0.8.5",
     "scikit-learn>=1.6.0",
     "streamlit>=1.41.1",
     "sympy>=1.13.3",
@@ -43,6 +42,7 @@ dev = [
     "types-pyyaml>=6.0.12.20241230",
     "types-toml>=0.10.8.20240310",
     "xlsxwriter>=3.2.0",
+    "ruff>=0.9.1",
 ]
 docs = [
     "autodoc-pydantic>=2.2.0",
