+++ _sources/examples/differentiating-voltage-data.ipynb.txt
@@ -24,6 +24,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
+    "%pip install matplotlib\n",
     "import pyprobe\n",
     "import matplotlib.pyplot as plt\n",
     "\n",
@@ -231,11 +232,6 @@
   }
  ],
  "metadata": {
-  "kernelspec": {
-   "display_name": "pyprobe-dev",
-   "language": "python",
-   "name": "python3"
-  },
   "language_info": {
    "codemirror_mode": {
     "name": "ipython",
