+++ _sources/examples/ocv-fitting.ipynb.txt
@@ -15,6 +15,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
+    "%pip install matplotlib\n",
     "import pyprobe\n",
     "import polars as pl\n",
     "from pyprobe.analysis import degradation_mode_analysis as dma\n",
@@ -234,11 +235,6 @@
   }
  ],
  "metadata": {
-  "kernelspec": {
-   "display_name": "pyprobe-dev",
-   "language": "python",
-   "name": "python3"
-  },
   "language_info": {
    "codemirror_mode": {
     "name": "ipython",
