+++ _sources/examples/analysing-GITT-data.ipynb.txt
@@ -18,6 +18,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
+    "%pip install matplotlib\n",
     "import pyprobe\n",
     "import matplotlib.pyplot as plt\n",
     "\n",
@@ -221,11 +222,6 @@
   }
  ],
  "metadata": {
-  "kernelspec": {
-   "display_name": ".venv",
-   "language": "python",
-   "name": "python3"
-  },
   "language_info": {
    "codemirror_mode": {
     "name": "ipython",
