### 2025-01-12 62adee0
chore: ignore performance example from pytest on notebook examples
