+++ .github/workflows/ci.yml
@@ -36,6 +36,12 @@ jobs:
       - name: Run tests
         run: uv run pytest tests --benchmark-json output.json --cov-report=xml
 
+      - name: Upload coverage artifact
+        uses: actions/upload-artifact@v3
+        with:
+          name: coverage
+          path: coverage.xml
+
       - name: Store benchmark result
         uses: benchmark-action/github-action-benchmark@v1
         with:
