+++ _sources/examples/filtering-data.ipynb.txt
@@ -21,12 +21,20 @@
    "metadata": {},
    "outputs": [],
    "source": [
+    "%%capture\n",
     "%pip install matplotlib\n",
     "import pyprobe\n",
     "import matplotlib.pyplot as plt\n",
     "\n",
-    "%matplotlib inline\n",
-    "\n",
+    "%matplotlib inline"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
     "info_dictionary = {\n",
     "    \"Name\": \"Sample cell\",\n",
     "    \"Chemistry\": \"NMC622\",\n",
