### 2025-01-12 5017aec
docs: minor typo fix to paper
