### 2025-01-12 ef08ea5
chore: add urls and classifiers to pyproject.toml
