+++ pyproject.toml
@@ -22,6 +22,19 @@ dependencies = [
     "sympy>=1.13.3",
 ]
 
+classifiers = [
+  "License :: OSI Approved :: BSD License",
+  "Intended Audience :: Science/Research",
+  "Topic :: Scientific/Engineering",
+]
+
+[project.urls]
+Homepage = "https://github.com/ImperialCollegeLondon/PyProBE"
+Documentation = "https://imperialcollegelondon.github.io/PyProBE/"
+Repository = "https://github.com/ImperialCollegeLondon/PyProBE"
+Releases = "https://github.com/ImperialCollegeLondon/PyProBE/releases"
+Changelog = "https://github.com/ImperialCollegeLondon/PyProBE/blob/main/CHANGELOG.md"
+
 [project.optional-dependencies]
 hvplot = [
     "hvplot>=0.11.2",
