### 2025-01-12 1a07ae9
chore: remove lean differentiation example
