### 2025-01-12 #7474 @tomjholland
Thanks @phibeck, I think I have done all of the required tasks listed above:
- Final version: `v1.3.0`
- DOI: https://doi.org/10.5281/zenodo.14635071
