### 2025-01-12 87de879
chore: add status badges to readme
