### 2025-01-21 b158397
chore: add --frozen flag to uv sync commands in workflows
