### 2025-01-22 dbd16c9
chore: replace warnings with logger for missing columns in data
