### 2025-01-22 246d7b1
fix: fix uv version in pre-commit and workflows
