+++ .github/workflows/ci.yml
@@ -24,6 +24,9 @@ jobs:
       - uses: actions/setup-python@v4
       - name: Install uv
         uses: astral-sh/setup-uv@v5
+        with:
+          # Install a specific version of uv.
+          version: "0.5.22"
 
       - name: Set up Python
         uses: actions/setup-python@v5
