+++ .github/workflows/sphinx.yml
@@ -17,6 +17,9 @@ jobs:
       - uses: actions/setup-python@v4
       - name: Install uv
         uses: astral-sh/setup-uv@v5
+        with:
+          # Install a specific version of uv.
+          version: "0.5.22"
 
       - name: Set up Python
         uses: actions/setup-python@v5
