### 2025-01-22 #7474 @tomjholland
@phibeck apologies, I noticed another typo in one of my figures. I have released a new version, which has a new DOI on zenodo:
- Version: `v1.3.2`
- DOI: https://doi.org/10.5281/zenodo.14720989
I will re-generate the paper draft below
