### 2025-01-28 d5b94a6
feat: add capability to export any Result object to a .mat file
