### 2025-01-29 24f7119
chore: specify paths for ci workflow
