### 2025-01-29 480dcca
chore: add path filters for sphinx workflow
