+++ .github/workflows/sphinx.yml
@@ -7,6 +7,13 @@ on:
   pull_request:
     branches:
       - main
+    paths:
+      - pyprobe/*
+      - tests/sample_data/*
+      - docs/*
+      - pyproject.toml
+      - .github/workflows/sphinx.yml
+      - uv.lock
   workflow_dispatch:
 
 jobs:
