+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-name: "Build and test"
+name: "Test and QA"
 on:
   pull_request:
     branches:
