### 2025-01-29 d360ec4
chore: replace pre-commit action with pre-commit.ci
