### 2025-01-29 e62a02d
chore: release candidate 1.4.0
