+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # CHANGELOG
 
 
-## v1.4.0 (2025-01-30)
+## v1.4.0 (2025-01-31)
 
 ### Chores
 
 - Add a .zenodo.json file
   ([`29be1c1`](https://github.com/ImperialCollegeLondon/PyProBE/commit/29be1c1e275561eebebbfcddb4f144ee9fa45cf6))
+
+- Add isort setting to ruff
+  ([`24408a1`](https://github.com/ImperialCollegeLondon/PyProBE/commit/24408a133e4d2265f703d7d335ace9aed3ddba38))
 
 - Add JOSS status badge
   ([`43a94a9`](https://github.com/ImperialCollegeLondon/PyProBE/commit/43a94a9b5db654e90424105ab0e11e623f0a17bd))
@@ -19,6 +22,9 @@
 
 - Add status badges to readme
   ([`87de879`](https://github.com/ImperialCollegeLondon/PyProBE/commit/87de879e1b330b3c55dbaaa35e8e1a9c58bb5b54))
+
+- Release candidate 1.4.0
+  ([`0d3181d`](https://github.com/ImperialCollegeLondon/PyProBE/commit/0d3181d2ecda0c41cc6568748e0682026d557ad4))
 
 - Release candidate 1.4.0
   ([`e62a02d`](https://github.com/ImperialCollegeLondon/PyProBE/commit/e62a02d2554a7b6c3fb6412ffedd6f504bb87bc1))
