### 2025-01-31 ff88c3d
fix: move critical plot functionality for dashboard into dashboard script
