+++ pyproject.toml
@@ -9,6 +9,7 @@ license = { file = "LICENSE" }
 readme = {file = "README.md", content-type = "text/markdown"}
 dependencies = [
     "deprecated>=1.2.15",
+    "distinctipy>=1.3.4",
     "fastexcel>=0.12.0",
     "ipykernel>=6.29.5",
     "numpy>=2.2.1",
