### 2025-02-01 bf38f5f
refactor: split dashboard into functions and add tests
