+++ pyprobe/dashboard.py
@@ -5,7 +5,7 @@ import os
 import pickle
 import platform
 import subprocess
-from typing import List
+from typing import Any, List
 
 import distinctipy
 import plotly.graph_objects as go
@@ -54,50 +54,39 @@ def launch_dashboard(cell_list: List[Cell]) -> None:
         )
 
 
-if __name__ == "__main__":
-    with open("dashboard_data.pkl", "rb") as f:
-        cell_list = pickle.load(f)
+def _dataframe_with_selections(df: pl.DataFrame) -> List[int]:
+    """Create a dataframe with a selection column for user input.
 
-    st.title("PyProBE Dashboard")
-    st.sidebar.title("Select data to plot")
+    Args:
+        df (pd.DataFrame): The dataframe to display.
 
-    info_list = []
-    for i in range(len(cell_list)):
-        info_list.append(cell_list[i].info)
-    info = pl.DataFrame(info_list)
+    Returns:
+        list: The list of selected row indices.
+    """
+    df = df.to_pandas()
+    df_with_selections = copy.deepcopy(df)
+    df_with_selections.insert(0, "Select", False)
 
-    def dataframe_with_selections(df: pl.DataFrame) -> List[int]:
-        """Create a dataframe with a selection column for user input.
+    # Get dataframe row-selections from user with st.data_editor
+    edited_df = st.sidebar.data_editor(
+        df_with_selections,
+        hide_index=True,  # Keep the index visible
+        column_config={"Select": st.column_config.CheckboxColumn(required=True)},
+        disabled=df.columns.tolist(),
+    )
 
-        Args:
-            df (pd.DataFrame): The dataframe to display.
+    # Filter the dataframe using the temporary column, then drop the column
+    selected_rows = edited_df[edited_df.Select]
+    selected_indices = (
+        selected_rows.index.tolist()
+    )  # Get the indices of the selected rows
+    return selected_indices
 
-        Returns:
-            list: The list of selected row indices.
-        """
-        df = df.to_pandas()
-        df_with_selections = copy.deepcopy(df)
-        df_with_selections.insert(0, "Select", False)
 
-        # Get dataframe row-selections from user with st.data_editor
-        edited_df = st.sidebar.data_editor(
-            df_with_selections,
-            hide_index=True,  # Keep the index visible
-            column_config={"Select": st.column_config.CheckboxColumn(required=True)},
-            disabled=df.columns.tolist(),
-        )
-
-        # Filter the dataframe using the temporary column, then drop the column
-        selected_rows = edited_df[edited_df.Select]
-        selected_indices = (
-            selected_rows.index.tolist()
-        )  # Get the indices of the selected rows
-        return selected_indices
-
-    # Display the DataFrame in the sidebar
-    selected_indices = dataframe_with_selections(info)
-
-    # Get the procedure names from the selected cells
+def _get_procedure_names(
+    cell_list: List[Cell], selected_indices: List[int]
+) -> List[str]:
+    """Get the common procedure names from the selected cells."""
     procedure_names_sets = [
         list(cell_list[i].procedure.keys()) for i in selected_indices
     ]
@@ -109,13 +98,22 @@ if __name__ == "__main__":
         procedure_names = list(procedure_names_sets[0])
         for s in procedure_names_sets[1:]:
             procedure_names = [x for x in procedure_names if x in s]
-    procedure_names = list(procedure_names)
-    selected_raw_data = st.sidebar.selectbox("Select a procedure", procedure_names)
+    return list(procedure_names)
+
+
+def _get_data_filters(
[…]
