+++ tests/test_dashboard.py
@@ -1,0 +1,367 @@
+"""Tests for the dashboard module."""
+
+import subprocess
+from unittest.mock import MagicMock, patch
+
+import pandas as pd
+import polars as pl
+import pytest
+import streamlit as st
+from streamlit.testing.v1 import AppTest
+
+from pyprobe.dashboard import (
+    _dataframe_with_selections,
+    _get_procedure_names,
+    launch_dashboard,
+)
+
+
+@pytest.fixture
+def mock_cell():
+    """Create a mock Cell object."""
+    cell = MagicMock()
+    cell.info = {"id": "test_cell"}
+    cell.procedure = {"test_proc": MagicMock()}
+    return cell
+
+
+@pytest.fixture
+def cell_list(mock_cell):
+    """Create a list of mock Cell objects."""
+    return [mock_cell, MagicMock()]
+
+
+def test_pickle_dump(cell_list, mocker):
+    """Test that cell list is properly pickled with correct context handling."""
+    mock_open = mocker.mock_open()
+
+    with (
+        patch("builtins.open", mock_open),
+        patch("pickle.dump") as mock_dump,
+        patch("subprocess.Popen"),
+    ):
+        launch_dashboard(cell_list)
+
+        mock_open.assert_called_once_with("dashboard_data.pkl", "wb")
+        mock_dump.assert_called_once_with(cell_list, mock_open())
+
+
+def test_windows_launch(cell_list):
+    """Test Windows subprocess launch."""
+    with patch("platform.system", return_value="Windows"):
+        with patch("subprocess.Popen") as mock_popen:
+            with patch("builtins.open"):
+                with patch("pickle.dump"):
+                    launch_dashboard(cell_list)
+
+                    mock_popen.assert_called_once()
+                    args = mock_popen.call_args[0][0]
+
+                    assert args[0:5] == ["cmd", "/c", "start", "/B", "streamlit"]
+                    assert "dashboard.py" in args[6]
+                    assert args[-3:] == [">", "nul", "2>&1"]
+                    assert mock_popen.call_args[1]["shell"] is True
+
+
+def test_darwin_launch(cell_list):
+    """Test Darwin/MacOS subprocess launch."""
+    with patch("platform.system", return_value="Darwin"):
+        with patch("subprocess.Popen") as mock_popen:
+            with patch("builtins.open"):
+                with patch("pickle.dump"):
+                    launch_dashboard(cell_list)
+
+                    mock_popen.assert_called_once()
+                    args = mock_popen.call_args[0][0]
+
+                    assert args[0:3] == ["nohup", "streamlit", "run"]
+                    assert "dashboard.py" in args[3]
+                    assert mock_popen.call_args[1]["stdout"] is subprocess.DEVNULL
+                    assert mock_popen.call_args[1]["stderr"] is subprocess.STDOUT
+
+
+def test_other_platform(cell_list):
+    """Test that no subprocess is launched on unsupported platforms."""
+    with patch("platform.system", return_value="Linux"):
+        with patch("subprocess.Popen") as mock_popen:
+            with patch("builtins.open"):
+                with patch("pickle.dump"):
+                    launch_dashboard(cell_list)
+                    mock_popen.assert_not_called()
+
+
+def test_empty_cell_list():
+    """Test handling of empty cell list."""
+    with patch("builtins.open"):
+        with patch("pickle.dump") as mock_dump:
+            with patch("subprocess.Popen"):
[…]
