### 2025-02-01 #7474 @phibeck
Hi @tomjholland thanks for your patience, I will send comments shortly.

Thanks for the update. Just to let you know, changes to the paper only don't necessarily require updating the version / archives, as long as the software didn't change.
