### 2025-02-01 #7474 @phibeck
@tomjholland I see there's a new release on the repo. Is `v1.3.2` still the version that should be used for the JOSS paper?
