### 2025-02-01 63a5bd7
refactor: reorganise data processing methods
