+++ pyprobe/cell.py
@@ -34,31 +34,22 @@ class Cell(BaseModel):
 
     def _convert_to_parquet(
         self,
-        cycler: Literal[
-            "neware", "biologic", "biologic_MB", "arbin", "basytec", "maccor", "generic"
-        ],
+        importer: basecycler.BaseCycler,
         folder_path: str,
-        input_filename: str | Callable[[str], str],
         output_filename: str | Callable[[str], str],
         filename_inputs: Optional[List[str]] = None,
         compression_priority: Literal[
             "performance", "file size", "uncompressed"
         ] = "performance",
         overwrite_existing: bool = False,
-        column_dict: Optional[Dict[str, str]] = None,
-        header_row_index: int = 0,
-        date_column_format: Optional[str] = None,
     ) -> None:
         """Convert a file into PyProBE format.
 
         Args:
-            cycler:
-                The cycler used to produce the data.
+            importer:
+                The cycler object to import the data.
             folder_path:
                 The path to the folder containing the data file.
-            input_filename:
-                A filename string or a function to generate the file name for cycler
-                data.
             output_filename:
                 A filename string or a function to generate the file name for PyProBE
                 data.
@@ -75,22 +66,7 @@ class Cell(BaseModel):
                 If True, any existing parquet file with the output_filename will be
                 overwritten. If False, the function will skip the conversion if the
                 parquet file already exists.
-            column_dict:
-                A dictionary mapping the column names in the cycler file to the PyProBE
-                column names. The keys of the dictionary are the cycler column names and
-                the values are the PyProBE column names. You must use asterisks to
-                indicate the units of the columns. Only used for the 'generic' cycler.
-                E.g. :code:`{"V (*)": "Voltage [*]"}`.
-            header_row_index:
-                The index of the header row in the file. Only used for the 'generic'
-                cycler.
-            date_column_format:
-                The format of the date column in the generic file. Only used for the
-                'generic' cycler.
         """
-        input_data_path = self._get_data_paths(
-            folder_path, input_filename, filename_inputs
-        )
         output_data_path = self._get_data_paths(
             folder_path, output_filename, filename_inputs
         )
@@ -101,28 +77,7 @@ class Cell(BaseModel):
             raise ValueError(error_msg)
 
         if not os.path.exists(output_data_path) or overwrite_existing:
-            cycler_dict = {
-                "neware": neware.Neware,
-                "biologic": biologic.Biologic,
-                "biologic_MB": biologic.BiologicMB,
-                "arbin": arbin.Arbin,
-                "basytec": basytec.Basytec,
-                "maccor": maccor.Maccor,
-            }
             t1 = time.time()
-            if cycler == "generic":
-                if column_dict is None:
-                    error_msg = "column_dict must be provided for the generic cycler."
-                    logger.error(error_msg)
-                    raise ValueError(error_msg)
-                importer = basecycler.BaseCycler(
-                    input_data_path=input_data_path,
-                    column_dict=column_dict,
-                    header_row_index=header_row_index,
-                    datetime_format=date_column_format,
-                )
-            else:
-                importer = cycler_dict[cycler](input_data_path=input_data_path)
             compression_dict = {
                 "uncompressed": "uncompressed",
                 "performance": "lz4",
@@ -179,10 +134,21 @@ class Cell(BaseModel):
                 overwritten. If False, the function will skip the conversion if the
                 parquet file already exists.
         """
+        cycler_dict = {
+            "neware": neware.Neware,
+            "biologic": biologic.Biologic,
+            "biologic_MB": biologic.BiologicMB,
+            "arbin": arbin.Arbin,
+            "basytec": basytec.Basytec,
+            "maccor": maccor.Maccor,
+        }
[…]
