### 2025-02-01 11cfa02
tests: add tests for errors when processing input files
