### 2025-02-01 b03ab52
refactor: remove deprecated Plot class
BREAKING CHANGE: removal of deprecated class due to maintenance overhead. Use Result.plot(), Result.hvplot() or seaborn wrapper instead
