+++ pyprobe/plot.py
@@ -148,308 +148,3 @@ Other modifications include:
 
 See the seaborn documentation for more information: https://seaborn.pydata.org/
 """
-
-
-@deprecated(
-    reason="The Plot class is deprecated as a user facing class. It is kept for "
-    "powering the backend of the dashboard. For user facing plotting, use the "
-    "plot, hvplot or seaborn plotting integrations.",
-    version="1.1.4",
-)
-class Plot:
-    """A class for plotting result objects with plotly.
-
-    The Plot class is deprecated as a user facing class. It is kept for
-    powering the backend of the dashboard. For user facing plotting, use the
-    plot, hvplot or seaborn plotting integrations."
-
-    Args:
-        layout (go.Layout): The plotly layout to use.
-    """
-
-    title_font_size = 18
-    axis_font_size = 14
-    default_layout = go.Layout(
-        template="simple_white",
-        title_font=dict(size=title_font_size),
-        xaxis_title_font=dict(size=title_font_size),
-        yaxis_title_font=dict(size=title_font_size),
-        xaxis_tickfont=dict(size=axis_font_size),
-        yaxis_tickfont=dict(size=axis_font_size),
-        legend_font=dict(size=axis_font_size),
-        width=800,
-        height=600,
-    )
-
-    def __init__(
-        self,
-        layout: go.Layout = default_layout,
-    ):
-        """Initialize the Plot object."""
-        self.layout = layout
-        self._fig = make_subplots(specs=[[{"secondary_y": True}]])
-        self._fig.update_layout(layout)
-        self.x_min: Optional[float] = None
-        self.x_max: Optional[float] = None
-        self.y_min: Optional[float] = None
-        self.y_max: Optional[float] = None
-        self.y2_min: Optional[float] = None
-        self.y2_max: Optional[float] = None
-
-        self.xaxis_title: Optional[str] = None
-        self.yaxis_title: Optional[str] = None
-        self.secondary_y: Optional[str] = None
-
-    def show(self) -> None:
-        """Show the plot."""
-        self.fig.show()
-
-    def add_line(
-        self,
-        result: "Result",
-        x: str,
-        y: str,
-        secondary_y: Optional[str] = None,
-        color: Optional[str] = None,
-        label: Optional[str] = None,
-        dash: str = "solid",
-        showlegend: bool = True,
-    ) -> "Plot":
-        """Add a line to the plot.
-
-        Args:
-            result (Result): The result object.
-            x (str): The x-axis dataframe column.
-            y (str): The y-axis dataframe column.
-            secondary_y (str): The secondary y-axis dataframe column.
-            color (str): The color of the line.
-            label (str): The label of the line.
-            dash (str): The dash style of the line. E.g. "dash", "dot", "dashdot".
-            showlegend (bool): Whether to show the legend.
-        """
-        if label is None:
-            if "Name" in result.info:
-                label = str(result.info["Name"])
-            else:
-                label = "Data"
-        x_data, y_data = result.get(x, y)
-        self._check_limits(x_data, y_data)
-        self.xaxis_title = x
-        self.yaxis_title = y
-        x_data, y_data = result.get(x, y)
-        self._fig.add_trace(
-            go.Scatter(
-                x=x_data,
-                y=y_data,
-                mode="lines",
-                line=dict(color=color, dash=dash),
-                name=label,
-                showlegend=showlegend,
-            )
-        )
-
-        if secondary_y is not None:
-            self.secondary_y = secondary_y
-            self._add_secondary_y_line(result, x, secondary_y)
-        return self
-
-    def _add_secondary_y_line(
-        self,
-        result: "Result",
-        x: str,
-        y: str,
-        color: Optional[str] = None,
[…]
