+++ pyprobe/__init__.py
@@ -4,7 +4,6 @@ from ._version import __version__  # noqa: F401
 from .cell import Cell, load_archive, make_cell_list  # noqa: F401
 from .dashboard import launch_dashboard  # noqa: F401
 from .logger import configure_logging  # noqa: F401
-from .plot import Plot  # noqa: F401
 from .result import Result  # noqa: F401
 
 configure_logging()
