+++ tests/test_plot.py
@@ -9,7 +9,6 @@ from plotly.express.colors import sample_colorscale
 from sklearn.preprocessing import minmax_scale
 
 from pyprobe import plot
-from pyprobe.plot import Plot
 from pyprobe.result import Result
 
 
@@ -176,305 +175,3 @@ def test_seaborn_wrapper_complete_coverage():
     sns_attrs = {attr for attr in dir(sns) if not attr.startswith("_")}
     wrapper_attrs = {attr for attr in dir(wrapper) if not attr.startswith("_")}
     assert sns_attrs == wrapper_attrs
-
-
-@pytest.fixture
-def Plot_fixture():
-    """Return a Plot instance."""
-    return Plot()
-
-
-def test_init(Plot_fixture):
-    """Test the __init__ method."""
-    assert isinstance(Plot_fixture, Plot)
-    assert Plot_fixture.layout == Plot.default_layout
-    assert isinstance(Plot_fixture._fig, go.Figure)
-
-
-def test_check_limits(Plot_fixture):
-    """Test the _check_limits method."""
-    data = pl.DataFrame({"x": [5, 6, 7, 8], "y": [1, 2, 3, 4]})
-    x = data["x"]
-    y = data["y"]
-    Plot_fixture._check_limits(x, y)
-    assert Plot_fixture.x_max == 8
-    assert Plot_fixture.x_min == 5
-    assert Plot_fixture.y_max == 4
-    assert Plot_fixture.y_min == 1
-
-    data = pl.DataFrame({"x": [-4, 6, 14, 12], "y": [2, 2, 3, 3]})
-    x = data["x"]
-    y = data["y"]
-
-    Plot_fixture._check_limits(x, y)
-    assert Plot_fixture.x_max == 14
-    assert Plot_fixture.x_min == -4
-    assert Plot_fixture.y_max == 4
-    assert Plot_fixture.y_min == 1
-
-    Plot_fixture._check_limits(x, y, secondary_y=True)
-    assert Plot_fixture.y2_max == 3
-    assert Plot_fixture.y2_min == 2
-
-
-def test_ranges(Plot_fixture):
-    """Test the ranges method."""
-    Plot_fixture.x_max = 14
-    Plot_fixture.x_min = -4
-    Plot_fixture.y_max = 4
-    Plot_fixture.y_min = 1
-    Plot_fixture.y2_max = 3
-    Plot_fixture.y2_min = 2
-
-    assert Plot_fixture.x_range == [-4.9, 14.9]
-    assert Plot_fixture.y_range == [0.85, 4.15]
-    assert Plot_fixture.y2_range == [1.95, 3.05]
-
-
-@pytest.fixture
-def plot_result_fixture():
-    """Return a Plot instance."""
-    info = {"color": "blue", "Name": "Line 1"}
-    data = pl.DataFrame(
-        {"x": [1, 2, 3, 4], "y": [5, 6, 7, 8], "secondary_y": [2, 3, 4, 5]}
-    )
-
-    return Result(base_dataframe=data, info=info)
-
-
-def test_make_colorscale(Plot_fixture):
-    """Test the make_colorscale method."""
-    colormap = "viridis"
-    points = np.array([0, 3, 1, 2])
-    colors = Plot_fixture.make_colorscale(points, colormap)
-
-    assert isinstance(colors, list)
-    assert colors == [
-        "rgb(68, 1, 84)",
-        "rgb(253, 231, 37)",
-        "rgb(49, 104, 142)",
-        "rgb(53, 183, 121)",
-    ]
-
-
-def test_add_line(Plot_fixture, plot_result_fixture):
-    """Test the add_line method."""
-    result = plot_result_fixture
-    x = "x"
-    y = "y"
-    secondary_y = None
[…]
