### 2025-02-01 29241f3
tests(cell): add missing tests to cell module
