### 2025-02-01 600c04a
refactor: remove deprecated analysis classes
BREAKING CHANGE: class-based structure for the analysis module has been replaced with single functions within the same module
