+++ pyprobe/analysis/differentiation.py
@@ -147,154 +147,3 @@ def differentiate_LEAN(
     gradient_result.column_definitions = input_data.column_definitions
     gradient_result.column_definitions[gradient_title] = "The calculated gradient."
     return gradient_result
-
-
-# 1. Define the class as a Pydantic BaseModel.
-class Differentiation(BaseModel):
-    """A class for differentiating experimental data."""
-
-    # 2. Define the input_data attribute, giving it a type
-    input_data: PyProBEDataType
-    """The input data for the differentiation."""
-
-    @deprecated("Use differentiation.gradient instead.", version="1.1.0")
-    def differentiate_FD(  # 1. Define the method
-        self,
-        x: str,
-        y: str,
-        gradient: str = "dydx",
-    ) -> Result:
-        """Differentiate smooth data with the finite difference method.
-
-        A light wrapper of the numpy.gradient function.
-
-        Args:
-            x (str): The name of the x variable.
-            y (str): The name of the y variable.
-            gradient (str, optional):
-                The gradient to calculate, either 'dydx' or 'dxdy'.
-                Defaults to "dydx".
-
-        Returns:
-            Result:
-                A result object containing the columns, `x`, `y` and the
-                calculated gradient.
-        """
-        # 2. Validate the inputs to the method
-        validator = AnalysisValidator(
-            input_data=self.input_data,
-            required_columns=[x, y],
-            # required_type not neccessary here as type specified when declaring
-            # input_data attribute is strict enough
-        )
-        # 3. Retrieve the validated columns as numpy arrays
-        x_data, y_data = validator.variables
-
-        # 4. Perform the computation
-        if gradient == "dydx":
-            gradient_title = f"d({y})/d({x})"
-            gradient_data = np.gradient(y_data, x_data)
-        elif gradient == "dxdy":
-            gradient_title = f"d({x})/d({y})"
-            gradient_data = np.gradient(x_data, y_data)
-        else:
-            raise ValueError("Gradient must be either 'dydx' or 'dxdy'.")
-
-        # 5. Create a Result object to store the results
-        gradient_result = self.input_data.clean_copy(
-            pl.DataFrame({x: x_data, y: y_data, gradient_title: gradient_data})
-        )
-        # 6. Define the column definitions for the Result object
-        gradient_result.column_definitions = self.input_data.column_definitions
-        gradient_result.column_definitions[gradient_title] = "The calculated gradient."
-        # 7. Return the Result object
-        return gradient_result
-
-    @deprecated("Use differentiation.differentiate_LEAN instead.", version="1.1.0")
-    def differentiate_LEAN(
-        self,
-        x: str,
-        y: str,
-        k: int = 1,
-        gradient: str = "dydx",
-        smoothing_filter: List[float] = [0.0668, 0.2417, 0.3830, 0.2417, 0.0668],
-        section: str = "longest",
-    ) -> Result:
-        r"""A method for differentiating noisy data.
-
-        Uses 'Level Evaluation ANalysis' (LEAN) method described in the paper of
-        :footcite:t:`Feng2020`.
-
-        This method assumes :math:`x` datapoints to be evenly spaced, it can return
-        either :math:`\frac{dy}{dx}` or :math:`\frac{dx}{dy}` depending on the argument
-        provided to the `gradient` parameter.
-
-        Args:
-            x (str):
-                The name of the x variable.
-            y (str):
-                The name of the y variable.
-            k (int, optional):
-                The integer multiple to apply to the sampling interval for the bin size
-                (:math:`\delta R` in paper). Default is 1.
-            gradient (str, optional):
-                The gradient to calculate, either 'dydx' or 'dxdy'. Default is 'dydx'.
[…]
