### 2025-02-01 eaf42e1
chore(ci): update ci workflow to install pybamm before running tests
