+++ .github/workflows/ci.yml
@@ -41,9 +41,11 @@ jobs:
         run: |
           uv venv --seed
           uv sync --all-extras --frozen
+          source .venv/bin/activate
+          uv pip install pybamm
 
       - name: Run tests
-        run: uv run pytest tests --benchmark-json output.json --cov-report=xml
+        run: pytest tests --benchmark-json output.json --cov-report=xml
 
       - name: Store benchmark result
         uses: benchmark-action/github-action-benchmark@v1
