### 2025-02-01 f8f21cc
chore: add pybamm installation to ci workflow
