### 2025-02-01 db65402
chore: run pytest with uv
