### 2025-02-03 a06ec46
refactor: reorganise dashboard into class and add tests for full coverage
