+++ pyprobe/dashboard.py
@@ -5,7 +5,7 @@ import os
 import pickle
 import platform
 import subprocess
-from typing import Any, List
+from typing import TYPE_CHECKING, Any, List
 
 import distinctipy
 import plotly.graph_objects as go
@@ -13,6 +13,9 @@ import polars as pl
 import streamlit as st
 
 from pyprobe.cell import Cell
+
+if TYPE_CHECKING:
+    from pyprobe.result import Result
 
 
 def launch_dashboard(cell_list: List[Cell]) -> None:
@@ -53,85 +56,14 @@ def launch_dashboard(cell_list: List[Cell]) -> None:
         )
 
 
-def _dataframe_with_selections(df: pl.DataFrame) -> List[int]:
-    """Create a dataframe with a selection column for user input.
+class _Dashboard:
+    """Class to create a Streamlit dashboard for PyProBE."""
 
-    Args:
-        df (pd.DataFrame): The dataframe to display.
+    def __init__(self, cell_list: List[Cell]) -> None:
+        """Initialize the dashboard with the cell list."""
+        self.cell_list = cell_list
+        self.info = self.get_info(self.cell_list)
 
-    Returns:
-        list: The list of selected row indices.
-    """
-    df = df.to_pandas()
-    df_with_selections = copy.deepcopy(df)
-    df_with_selections.insert(0, "Select", False)
-
-    # Get dataframe row-selections from user with st.data_editor
-    edited_df = st.sidebar.data_editor(
-        df_with_selections,
-        hide_index=True,  # Keep the index visible
-        column_config={"Select": st.column_config.CheckboxColumn(required=True)},
-        disabled=df.columns.tolist(),
-    )
-
-    # Filter the dataframe using the temporary column, then drop the column
-    selected_rows = edited_df[edited_df.Select]
-    selected_indices = (
-        selected_rows.index.tolist()
-    )  # Get the indices of the selected rows
-    return selected_indices
-
-
-def _get_procedure_names(
-    cell_list: List[Cell], selected_indices: List[int]
-) -> List[str]:
-    """Get the common procedure names from the selected cells."""
-    procedure_names_sets = [
-        list(cell_list[i].procedure.keys()) for i in selected_indices
-    ]
-
-    # Find the common procedure names
-    if len(procedure_names_sets) == 0:
-        procedure_names: List[str] = []
-    else:
-        procedure_names = list(procedure_names_sets[0])
-        for s in procedure_names_sets[1:]:
-            procedure_names = [x for x in procedure_names if x in s]
-    return list(procedure_names)
-
-
-def _get_data_filters(
-    cell_list: List[Cell], selected_indices: List[int], procedure_names: List[str]
-) -> tuple[Any, tuple[Any, ...], Any]:
-    import streamlit as st
-
-    selected_procedure = st.sidebar.selectbox("Select a procedure", procedure_names)
-
-    # Select an experiment
-    if selected_procedure is not None:
-        experiment_names = (
-            cell_list[selected_indices[0]]
-            .procedure[selected_procedure]
-            .experiment_names
-        )
-        selected_experiment = st.sidebar.multiselect(
-            "Select an experiment", experiment_names
-        )
-        selected_experiment_tuple = tuple(selected_experiment)
-    else:
-        selected_experiment_tuple = ()
-
-    # Get the cycle and step numbers from the user
-    cycle_step_input = st.sidebar.text_input(
-        'Enter the cycle and step numbers (e.g., "cycle(1).step(2)")'
-    )
-    return selected_procedure, selected_experiment_tuple, cycle_step_input
-
-
-def _get_graph_inputs(
[…]
