+++ tests/test_dashboard.py
@@ -3,15 +3,16 @@
 import subprocess
 from unittest.mock import MagicMock, patch
 
+import numpy as np
 import pandas as pd
 import polars as pl
+import polars.testing as pl_testing
 import pytest
 import streamlit as st
 from streamlit.testing.v1 import AppTest
 
 from pyprobe.dashboard import (
-    _dataframe_with_selections,
-    _get_procedure_names,
+    _Dashboard,
     launch_dashboard,
 )
 
@@ -99,269 +100,374 @@ def test_empty_cell_list():
                 mock_dump.assert_called_once_with([], mock_dump.call_args[0][1])
 
 
-@pytest.fixture
-def sample_df():
-    """Create a sample Polars DataFrame for testing."""
-    return pl.DataFrame({"col1": [1, 2, 3], "col2": ["a", "b", "c"]})
+def test_dashboard_init(mock_cell):
+    """Test Dashboard initialization."""
+    dashboard = _Dashboard([mock_cell])
+    assert dashboard.cell_list == [mock_cell]
+    assert isinstance(dashboard.info, pl.DataFrame)
 
 
-def test_dataframe_with_selections(sample_df):
-    """Test basic functionality of _dataframe_with_selections."""
-    with patch("streamlit.sidebar.data_editor") as mock_editor:
-        # Mock the returned DataFrame from st.data_editor
-        mock_edited_df = pd.DataFrame(
-            {"Select": [True, False, True], "col1": [1, 2, 3], "col2": ["a", "b", "c"]}
+def test_dashboard_get_info(mock_cell):
+    """Test info DataFrame creation."""
+    mock_cell.info = {"id": "test1", "value": 10}
+    mock_cell2 = MagicMock()
+    mock_cell2.info = {"id": "test2", "value": 20}
+
+    dashboard = _Dashboard([mock_cell, mock_cell2])
+    expected_df = pl.DataFrame(
+        [{"id": "test1", "value": 10}, {"id": "test2", "value": 20}]
+    )
+    pl_testing.assert_frame_equal(
+        dashboard.get_info([mock_cell, mock_cell2]), expected_df
+    )
+
+
+def testdataframe_with_selections():
+    """Test DataFrame filtering based on selections."""
+    data = pl.DataFrame({"id": ["test1", "test2"], "value": [10, 20]})
+    df_with_selections = _Dashboard.dataframe_with_selections(data)
+    assert "Select" in df_with_selections.columns
+    assert (df_with_selections["Select"].to_numpy() == False).all()
+
+
+def test_select_cell_indices(mock_cell):
+    """Test cell selection functionality."""
+    mock_cell.info = {"id": "test1", "value": 10}
+    mock_cell2 = MagicMock()
+    mock_cell2.info = {"id": "test2", "value": 20}
+    dashboard = _Dashboard([mock_cell, mock_cell2])
+
+    with patch("streamlit.sidebar.data_editor") as mock_select:
+        mock_select.return_value = pd.DataFrame(
+            {"id": ["test1", "test2"], "value": [10, 20], "Select": [True, False]}
         )
-        mock_editor.return_value = mock_edited_df
-
-        result = _dataframe_with_selections(sample_df)
-
-        assert result == [0, 2]
-        mock_editor.assert_called_once()
-
-        empty_df = pl.DataFrame()
-        mock_editor.return_value = pd.DataFrame({"Select": []})
-        result = _dataframe_with_selections(empty_df)
-        assert result == []
-
-        mock_edited_df = pd.DataFrame(
-            {
-                "Select": [False, False, False],
-                "col1": [1, 2, 3],
-                "col2": ["a", "b", "c"],
-            }
-        )
[…]
