### 2025-02-03 3510ddb
tests: increase timeout on dashboard apptest
