### 2025-02-03 df3ff12
fix: remove blank title from plot
