### 2025-02-03 42d0f98
chore: release candidate 2.0.0
