### 2025-02-03 #7474 @tomjholland
> [@tomjholland](https://github.com/tomjholland) I see there's a new release on the repo. Is `v1.3.2` still the version that should be used for the JOSS paper?

Hi @phibeck,
I caught a bug over the weekend and removed some deprecated functions. These changes might as well be in the release for the JOSS paper, so the version should now be `v2.0.0`.

The zenodo entry has been updated so the latest DOI is: https://doi.org/10.5281/zenodo.14796110
