### 2025-02-04 #7474 @phibeck
@tomjholland thanks for your patience. Here are my very few comments for the manuscript. Please have a look when you have a moment.

- lines 3f: in your affiliations, please spell out UK as United Kingdom
- line 79f: missing word: "...are described in Figure 5."
- line 106 and 111: just a note, for me the DOIs are stuck redirecting currently, but it's the same for the links on the respective Elsevier side so I would hope they fix this soon
