### 2025-02-04 3c04e46
docs: correct missing word typo in paper
