### 2025-02-04 #7474 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
